% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_edge_list.R
\name{add_edge_ids}
\alias{add_edge_ids}
\title{Add edge ID numbers to an edge list}
\usage{
add_edge_ids(edge_list, edge_id = "edge_id")
}
\arguments{
\item{edge_list}{the edge list to which edge ID numbers are to be added}

\item{edge_id}{the name of the edge ID column in the outgoing edge list. (Default is "\code{edge_id}".)}
}
\value{
\code{edge_list} with an added column containing the edge ID numbers.
}
\description{
The edges in an edge list can have ID numbers.
This functions adds a column of edge ID numbers.
}
\examples{
library(tidyr)
library(matsbyname)
sutmats <- UKEnergy2000mats \%>\% spread(key = matrix.name, value = matrix)
# Suppress adding edge IDs
elDF <- edge_list(sutmats, edge_id = NULL)$`Edge list`[[1]]
add_node_ids(elDF)
}
