% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep_utilities.R
\name{add_matnames}
\alias{add_matnames}
\title{Add a column of matrix names to tidy data frame}
\usage{
add_matnames(.DF, ledger_side = "Ledger.side", energy = "EX.ktoe",
  flow_aggregation_point = "Flow.aggregation.point", supply_side = "Supply",
  consumption_side = "Consumption", eiou = "Energy industry own use",
  matname = "matname", U = "U", U_EIOU = "U_EIOU", V = "V", Y = "Y")
}
\arguments{
\item{.DF}{a data frame with \code{ledger_side} and \code{energy} columns.}

\item{ledger_side}{the name of the column in \code{.DF} that contains ledger side
(a string). Default is "\code{Ledger.side}".}

\item{energy}{the name of the column in \code{.DF} that contains energy and exergy values
(a string). Default is "\code{EX.ktoe}".}

\item{supply_side}{the identifier for items on the supply side of the ledger (a string).
Default is "\code{Supply}".}

\item{consumption_side}{the identifier for items on the consumption side
of the ledger (a string). Default is "\code{Consumption}".}

\item{eiou}{the identifier for items that are energy industry own use.
Default is "\code{"Energy industry own use"}.}

\item{matname}{the name of the output column containing the name of the matrix
in which this row belongs (a string). Default is "\code{matname}".}

\item{U}{the name for the use matrix (a string). Default is "\code{U}".}

\item{U_EIOU}{the name for the energy industry own use matrix. Default is "\code{U_EIOU}".}

\item{V}{the name for the make matrix (a string). Default is "\code{V}".}

\item{Y}{the name for the final demand matrix (a string). Default is "\code{Y}".}
}
\value{
\code{.DF} with an added column, \code{matname}.
}
\description{
Add a column of matrix names to tidy data frame
}
\examples{
UKEnergy2000tidy \%>\%
  add_matnames()
}
