% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aggregate_cols}
\alias{aggregate_cols}
\title{Aggregate data frame column names}
\format{
A string list with 9 entries.
\describe{
\item{aggregate_primary}{The name of a column in a wide-by-matrices data frame containing aggregates of primary energy.}
\item{net_aggregate_primary}{The name of a column in a wide-by-matrices data frame containing aggregates of net primary energy.}
\item{gross_aggregate_primary}{The name of a column in a wide-by-matrices data frame containing aggregates of gross primary energy.}
\item{net_aggregate_demand}{The name of a column in a wide-by-matrices data frame containing aggregates of net final demand energy (excluding energy industry own use), regardless of whether the last stage is final energy or useful energy.}
\item{gross_aggregate_demand}{The name of a column in a wide-by-matrices data frame containing aggregates of gross final demand energy (including energy industry own use), regardless of whether the last stage is final energy or useful energy.}
\item{region}{The name of a column in a wide-by-matrices data frame containing regions.}
\item{aggregated_suffix}{The suffix for column names containing aggregated matrices.}
\item{product_sector}{The name of a column containing names of products, industries, or sectors. Default is "Product.Industry.Sector".}
\item{aggregates_df}{The name of a column containing a nested data frame of aggregates. Default is "Aggregates".}
}
}
\usage{
aggregate_cols
}
\description{
A string list containing named names of columns in aggregate data frames.
Items in the list provide default values for column name function arguments
to aggregation functions throughout the \code{Recca} package.
}
\examples{
aggregate_cols
}
\keyword{datasets}
