% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alloc_cols}
\alias{alloc_cols}
\title{Columns in a data frame that contain final-to-useful allocations}
\format{
A string list with 2 entries.
\describe{
\item{C_Y}{The name of a column in a wide-by-matrices data frame containing allocations to final-to-useful machines used in final demand. "C_Y"}
\item{C_eiou}{The name of a column in a wide-by-matrices data frame containing allocations to final-to-useful machines used in final demand. "C_EIOU"}
}
}
\usage{
alloc_cols
}
\description{
A string list containing named names of columns in PSUT data frames.
Items in the list provide default values for column name function arguments
throughout the \code{Recca} package.
}
\examples{
alloc_cols
}
\keyword{datasets}
