% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{any_start_with}
\alias{any_start_with}
\title{Tell if any of a list of strings starts with another string}
\usage{
any_start_with(x, target)
}
\arguments{
\item{x}{a vector of strings}

\item{target}{}
}
\value{
\code{TRUE} if any of \code{x} starts with \code{target}, \code{FALSE} otherwise.

importFrom Hmisc escapeRegex
}
\description{
Tell if any of a list of strings starts with another string
}
\examples{
any_start_with(x = c("ad", "bd", "cd"), target = "b") # TRUE
any_start_with(x = c("Production - Crude", "Production - NG", "Bogus"), target = "Production") # TRUE
any_start_with(x = c("Production - Crude", "Production - NG", "Bogus"), target = "Offshore") # FALSE
}
