% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{balance_cols}
\alias{balance_cols}
\title{Balance column names}
\format{
A string list with 8 entries.
\describe{
\item{inter_industry_balance_colname}{The name of a column that contains inter-industry balance vectors.}
\item{inter_industry_balanced_colname}{The name of a column that tells whether inter-industry balance vectors are all \strong{0}.}
\item{between_industry_balance_colname}{The name of a column that contains inter-industry balance vectors.}
\item{between_industry_balanced_colname}{The name of a column that contains inter-industry balance vectors are all \strong{0}.}
\item{intra_industry_balance_colname}{The name of a column that contains intra-industry balance vectors.}
\item{intra_industry_balanced_colname}{The name of a column that contains intra-industry balance vectors are all \strong{0}.}
\item{across_industry_balance_colname}{The name of a column that contains intra-industry balance vectors.}
\item{across_industry_balanced_colname}{The name of a column that contains intra-industry balance vectors are all \strong{0}.}
}
}
\usage{
balance_cols
}
\description{
A string list containing named names of columns in balance data frames.
Items in the list provide default values for column name function arguments
to balance functions throughout the \code{Recca} package.
}
\examples{
balance_cols
}
\keyword{datasets}
