% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregates.R
\name{calc_aggregates_from_ecc_prime}
\alias{calc_aggregates_from_ecc_prime}
\title{Calculate aggregates from list of reconstructed ECCs}
\usage{
calc_aggregates_from_ecc_prime(
  ecc_prime,
  p_industries,
  fd_sectors,
  pattern_type,
  by,
  product_sector,
  aggregates_df,
  aggregate_primary,
  gross_aggregate_demand,
  net_aggregate_demand,
  R_prime_colname,
  U_prime_colname,
  U_feed_prime_colname,
  U_eiou_prime_colname,
  r_eiou_prime_colname,
  V_prime_colname,
  Y_prime_colname
)
}
\arguments{
\item{ecc_prime}{A list of reconstructed energy conversion chains.}
}
\value{
A data frame containing reconstructed (prime) matrices and
primary and final demand aggregates in a list suitable for use in \code{matsindf::matsindf_apply()}.
}
\description{
This is a helper function for \code{footprint_aggregates()} and \code{effects_aggregates()}.
It calculates the primary and final demand aggregates for a list of
reconstructed energy conversion chains (ECCs) in \code{ecc_prime}.
}
