% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embodied.R
\name{calc_embodied_EIOU}
\alias{calc_embodied_EIOU}
\title{Calculate various embodied EIOU matrices}
\usage{
calc_embodied_EIOU(
  .iomats = NULL,
  e_EIOU = "e_EIOU",
  Y = "Y",
  y = "y",
  L_ixp = "L_ixp",
  L_ixp_feed = "L_ixp_feed",
  Q_EIOU_p = "Q_EIOU_p",
  Q_EIOU_s = "Q_EIOU_s",
  Q_EIOU_feed_p = "Q_EIOU_feed_p",
  Q_EIOU_feed_s = "Q_EIOU_feed_s"
)
}
\arguments{
\item{.iomats}{a data frame containing matrices that describe the Input-Output structure
(using the supply-use table format) of an Energy Conversion Chain.
\code{.iomats} will likely have been obtained from the \code{\link{calc_io_mats}} function.}

\item{Y}{final demand (\code{Y}) matrix or name of the column in \code{.iodata} containing same. Default is "\code{Y}".}

\item{L_ixp}{industry-by-product Leontief (\code{L_ixp}) matrix or name of the column in \code{.iodata} containing same. Default is "\code{L_ixp}"}

\item{q}{final demand (\code{q}) vector or name of the column in \code{.iodata} containing same. Default is "\code{q}".}

\item{g}{name of the \code{g} vector on output. Default is "\code{g}".}

\item{W}{name of the \code{W} matrix on output. Default is "\code{W}".}

\item{U_EIOU}{name of the \code{U_EIOU} matrices on output. Default is "\code{U_EIOU}".}

\item{G}{name of the \code{G} matrix on output.
\code{G} is calculated by \code{L_ixp * y_hat}. Default is "\code{G}".}

\item{H}{name of the \code{H} matrix on output.
\code{H} is calculated by \code{L_ixp * Y}. Default is "\code{H}".}

\item{E}{name of \code{E} matrix on output.
\code{E} is calculated by \code{W * g_hat_inv}. Default is "\code{E}".}
}
\value{
a list or data frame containing EIOU embodied matrices
}
\description{
Calculate various embodied EIOU matrices
}
