% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embodied.R
\name{calc_embodied_etas}
\alias{calc_embodied_etas}
\title{Embodied energy efficiencies}
\usage{
calc_embodied_etas(
  .embodiedmats = NULL,
  primary_machine_names,
  Y = "Y",
  G = "G",
  H = "H",
  eta_p = "eta_p",
  eta_s = "eta_s"
)
}
\arguments{
\item{.embodiedmats}{A data frame containing columns of \strong{Y}, \strong{G}, and \strong{H} matrices.}

\item{primary_machine_names}{A list of strings representing names of Industries whose output is counted in Total Energy Supply (TPES).}

\item{Y}{A final demand (\strong{Y}) matrix or name of a column in \code{.embodiedmats} containing same. Default is "Y".}

\item{G}{A \strong{G}  matrix or name of a column in \code{.embodiedmats} containing same. Default is "G".}

\item{H}{An \strong{H} matrix or name of a column in \code{.embodiedmats} containing same. Default is "H".}

\item{eta_p}{The name for product-based efficiencies on output. Default is "eta_p".}

\item{eta_s}{The name for final-demand-sector-based efficiencies on output. Default is "eta_s".}
}
\value{
A list or data frame containing embodied energy efficiencies.
}
\description{
Embodied energy efficiencies are based on the total upstream primary energy demand
for a product produced by the ECC or
for the energy consumed by a final demand sector of the ECC.
This function calculates both.
\strong{eta_s} gives sector-based embodied energy efficiency, and
\strong{eta_p} gives product-based embodied energy efficiency.
}
\details{
Note that these efficiencies (\strong{eta_s} and \strong{eta_p}) are different from
energy conversion industry efficiencies.
To calculate energy conversion industry efficiencies, use the
\code{calc_eta_i()} function.
}
