% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embodied.R
\name{calc_embodied_mats}
\alias{calc_embodied_mats}
\title{Calculate various embodied energy matrices}
\usage{
calc_embodied_mats(.iodata = NULL, Y_colname = "Y", q_colname = "q",
  L_ixp_colname = "L_ixp", g_colname = "g", W_colname = "W",
  U_EIOU_colname = "U_EIOU", G_colname = "G", H_colname = "H",
  E_colname = "E", M_p_colname = "M_p", M_s_colname = "M_s",
  F_footprint_p_colname = "F_footprint_p",
  F_effects_p_colname = "F_effects_p",
  F_footprint_s_colname = "F_footprint_s",
  F_effects_s_colname = "F_effects_s")
}
\arguments{
\item{.iodata}{a data frame containing matrices that describe the Input-Output structure
(using the supply-use table format) of an Energy Conversion Chain.
\code{.iodata} will likely have been obtained from the \code{calc_io_mats} function.}

\item{Y_colname}{the name of the column in \code{.iodata} containing final demand (\code{Y}) matrices.}

\item{q_colname}{the name of the column in \code{.iodata} containing final demand (\code{q}) vectors.}

\item{L_ixp_colname}{the name of the column in \code{.iodata} containing Industry-by-Product
Leontief (\code{L_ixp}) matrices.}

\item{g_colname}{the name of the output column containing \code{g} vectors.}

\item{W_colname}{the name of the output column containing \code{W} matrices.}

\item{U_EIOU_colname}{the name of the output column containing \code{U_EIOU} matrices.}

\item{G_colname}{the name of the output column containing \code{G} matrices.
\code{G} is calculated by \code{L_ixp * y_hat}.}

\item{H_colname}{the name of the output column containing \code{H} matrices.
\code{H} is calculated by \code{L_ixp * Y}.}

\item{E_colname}{the name of the output column containing \code{E} matrices.
\code{E} is calculated by \code{W * g_hat_inv}.}

\item{M_p_colname}{the name of the output column containing \code{M_p} matrices.
\code{M_p} is formed from column sums of positive entries in the various Qx matrices}

\item{M_s_colname}{the name of the output column containing \code{M_s} matrices.
\code{M_s} is constructed by \code{M_p * q_hat_inv * Y}.}

\item{F_footprint_p_colname}{the name of the output column containing \code{F_footprint_p} matrices.
\code{F}\code{_footprint_p} is calculated by \code{M_p} (\code{M_p}^T\code{i})_hat_inv.}

\item{F_effects_p_colname}{the name of the output column containing \code{F_effects_p} matrices.
\code{F}\code{_effects_p} is calculated by \code{M_p i}_hat_inv \code{M_p}.}

\item{F_footprint_s_colname}{the name of the output column containing \code{F_footprint_s} matrices.
\code{F}\code{_footprint_s} is calculated by \code{M_s} (\code{M_s}^T\code{i})_hat_inv.}

\item{F_effects_s_colname}{the name of the output column containing \code{F_effects_s} matrices.
\code{F}\code{_effects_p} is calculated by \code{M_s i}_hat_inv \code{M_s}.}
}
\value{
\code{.iodata} with columns
\code{G_colname}, \code{H_colname}, \code{E_colname}, and \code{Q_colname} added
}
\description{
Calculate various embodied energy matrices
}
