% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency.R
\name{calc_eta_fu_Y_eiou_ls_useful}
\alias{calc_eta_fu_Y_eiou_ls_useful}
\title{Calculates final-to-useful efficiencies when last stage is useful}
\usage{
calc_eta_fu_Y_eiou_ls_useful()
}
\description{
There are times when it is helpful to know the
final-to-useful efficiency of end uses at the useful stage of an energy conversion chain.
This function calculates those efficiencies.
}
\details{
The output of this function is matrices
with same structure as \strong{Y} and \strong{U_EIOU} when
last stage is useful.
}
