% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency.R
\name{calc_eta_i}
\alias{calc_eta_i}
\title{Calculate industry efficiencies}
\usage{
calc_eta_i(.sutdata, g_colname = "g", U_colname = "U",
  S_units_colname = "S_units", eta_colname = "eta")
}
\arguments{
\item{g_colname}{a string for the name of a column of \code{g} matrices in \code{.sutmats}. (Default is "\code{g}".)}

\item{U_colname}{a string for the name of a column of \code{U} matrices in \code{.sutmats}. (Default is "\code{U}".)}

\item{S_units_colname}{a string for the name of a column of \code{S_units} matrices in \code{.sutmats}. (Default is "\code{S_units}".)}

\item{eta_colname}{the name of the industry efficiency column in output. Default is "\code{eta}".}

\item{.sutmats}{a data frame containing columns for \code{g}, \code{U}, and \code{S_units} matrices}
}
\value{
\code{.sutmats} with an additional column "\code{eta}"
}
\description{
Calculates industry efficiencies for all energy conversion industries in the ECC.
}
\details{
Note that these efficiencies (\code{eta}) are different from
final demand sector and product efficiencies (\code{eta_s} and \code{eta_p}, respectively).
Both final demand sector and product efficiencies
(\code{eta_s} and \code{eta_p}) are based on embodied energy, whereas
industry efficiencies (\code{eta}) is based on direct energy consumed and produced
by the energy conversion industry.

To calculate energy conversion industry efficiencies, use the
\code{\link{calc_embodied_etas}} function.

The "\code{g}" column can be calculated by \code{\link{calc_io_mats}}.
}
