% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency.R
\name{calc_eta_pfu}
\alias{calc_eta_pfu}
\title{Calculate aggregate efficiencies}
\usage{
calc_eta_pfu(
  p_aggregates,
  finaldemand_aggregates,
  aggregate_primary_colname = Recca::aggregate_cols$aggregate_primary,
  net_aggregate_demand_colname = Recca::aggregate_cols$net_aggregate_demand,
  gross_aggregate_demand_colname = Recca::aggregate_cols$gross_aggregate_demand
)
}
\arguments{
\item{p_aggregates}{A data frame of primary aggregates, probably calculated by
\code{primary_aggregates()}.}

\item{finaldemand_aggregates}{A data frame of final demand aggregates, probably calculated by
\code{finaldemand_aggregates()}.}

\item{aggregate_primary_colname}{The name of the column in \code{p_aggregates} that contains primary energy or exergy aggregates.
Default is \code{Recca::aggregate_cols$aggregate_primary}.}

\item{net_aggregate_demand_colname}{The name of the column in \code{finaldemand_aggregates}
that contains net final demand aggregates.
Default is \code{Recca::aggregate_cols$net_aggregate_demand}.}

\item{gross_aggregate_demand_colname}{The name of the column in \code{finaldemand_aggregates}
that contains gross final demand aggregates.
Default is \code{Recca::aggregate_cols$gross_aggregate_demand}.}
}
\value{
A data frame of aggregate efficiencies.
}
\description{
Calculates aggregate primary-to-final demand (gross and net) efficiencies.
If final demand is at the final stage, calculates primary-to-final (gross and net) efficiencies.
If final demand is at the useful stage, calculates primary-to-useful (gross and net) efficiencies.
If final demand is at the services stage, calculates primary-to-services (gross and net) efficiencies.
}
