% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embodied.R
\name{calc_F_footprint_effects}
\alias{calc_F_footprint_effects}
\title{Upstream footprint and downstream effects matrices}
\usage{
calc_F_footprint_effects(.Mmats = NULL, M_p = "M_p", M_s = "M_s",
  F_footprint_p = "F_footprint_p", F_effects_p = "F_effects_p",
  F_footprint_s = "F_footprint_s", F_effects_s = "F_effects_s")
}
\arguments{
\item{.Mmats}{a data frame containing a column of embodied matrices}

\item{M_p}{embodied product matrix or name of the column in \code{.Mmats} containing same. Default is "\code{M_p}".}

\item{M_s}{embodied sector matrix or name of the column in \code{.Mmats} containing same. Default is "\code{M_s}".}

\item{F_footprint_p}{the name for \code{F_footprint_p} matrices on output. Default is "\code{F_footprint_p}".}

\item{F_effects_p}{the name for \code{F_effects_p} matrices on output. Default is "\code{F_effects_p}".}

\item{F_footprint_s}{the name for \code{F_footprint_s} matrices on output. Default is "\code{F_footprint_s}".}

\item{F_effects_s}{the name for \code{F_effects_s} matrices on output. Default is "\code{F_effects_s}".}
}
\value{
a list or data frame containing \code{F_footprint_p}, \code{F_effects_p},
\code{F_footprint_s}, and \code{F_effects_s} matrices
}
\description{
Calculates upstream footprint matrices (\code{F_footprint_p}, \code{F_footprint_s})
and downstream effects matrices (\code{F_effects_p}, \code{F_effects_s})
given an embodied matries \code{M_p} and \code{M_s}.
Column sums of \code{F_footprint} are 1.
Row sums of \code{F_effects} are 1.
}
