% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embodied.R
\name{calc_GH}
\alias{calc_GH}
\title{Calculate the \code{G} and \code{H} matrices for embodied energy calculations}
\usage{
calc_GH(
  .iomats = NULL,
  Y = "Y",
  L_ixp = "L_ixp",
  R = "R",
  A = "A",
  q = "q",
  G_V = "G_V",
  G_R = "G_R",
  G = "G",
  H_V = "H_V",
  H_R = "H_R",
  H = "H"
)
}
\arguments{
\item{.iomats}{a data frame containing matrices that describe the Input-Output structure of an Energy Conversion Chain.
\code{.iomats} will likely have been obtained from the \code{\link{calc_io_mats}} function.}

\item{Y}{final demand (\code{Y}) matrix or name of the column in \code{.iodata} containing same. Default is "\code{Y}".}

\item{L_ixp}{industry-by-product Leontief (\code{L_ixp}) matrix or name of the column in \code{.iodata} containing same. Default is "\code{L_ixp}".}

\item{R}{Resources (\code{R}) matrix or name of the column in \code{.iodata} containing same. Default is "\code{R}".}

\item{A}{The name of the \code{A} matrix column in the \code{.iomats} data frame.
Default is "A".}

\item{q}{The name of the \code{q} vector in the \code{.iomats} data frame.
Default is "q".}

\item{G_V}{name for the \code{G_V} matrix on output. Default is "\code{G_V}".
\code{G_V} is calculated by \code{L_ixp * y_hat}.}

\item{G_R}{name for the \code{G_R} matrix on output. Default is "\code{G_R}".
\code{G_R} is calculated by \code{R * q_hat_inv * L_pxp * y_hat}.}

\item{G}{name for the \code{G} matrix on output. Default is "\code{G}".
\code{G} is calculated by \code{G_R + G_V}.}

\item{H_V}{name for the \code{H_V} matrix on output. Default is "\code{H_V}".
\code{H_V} is calculated by \code{L_ixp * Y}.}

\item{H_R}{name for the \code{H_R} matrix on output. Default is "\code{H_R}".
\code{H_R} is calculated by \code{R * q_hat_inv * L_pxp * Y}.}

\item{H}{name for the \code{H} matrix on output. Default is "\code{H}".
\code{H} is calculated by \code{H_V + H_R}.}
}
\value{
a list or data frame containing \code{G} and \code{H} matrices.
}
\description{
Calculate the \code{G} and \code{H} matrices for embodied energy calculations
}
