% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy_balance.R
\name{calc_inter_industry_balance}
\alias{calc_inter_industry_balance}
\title{Calculate inter-industry energy balances for PSUT (\strong{RUVY}) matrices.}
\usage{
calc_inter_industry_balance(
  .sutmats = NULL,
  R = "R",
  U = "U",
  V = "V",
  Y = "Y",
  balance_colname = Recca::balance_cols$inter_industry_balance_colname
)
}
\arguments{
\item{.sutmats}{A named list of matrices or
an SUT-style data frame with columns of matrices,
including \code{R}, \code{U}, \code{V}, and \code{Y} columns.}

\item{R}{resources (\strong{R}) matrix or name of the column in \code{.sutmats}
that contains same. Default is "R".}

\item{U}{use (\strong{U}) matrix or name of the column in \code{.sutmats}
that contains same. Default is "U".}

\item{V}{make (\strong{V}) matrix or name of the column in \code{.sutmats}
that contains same. Default is "V".}

\item{Y}{final demand (\strong{Y}) matrix or name of the column in \code{.sutmats}
that contains same. Default is "Y".}

\item{balance_colname}{The name of the column containing energy balance vectors.
Default is \code{Recca::balance_cols$inter_industry_balance_colname}.}
}
\value{
A list or data frame containing inter-industry balances.
}
\description{
In a PSUT description of an economy, all of every flow
leaving one industry must arrive at another industry.
This function calculates those balances via
(\strong{R} + \strong{V}) - (\strong{U} + \strong{Y}).
The output of this function is named \code{balance_colname}.
}
\details{
Often, balances are calculated on energy or exergy flows.
Balances can be calculated for mass flows or monetary flows, too.

Inter-industry balances are calculated for products (not industries),
and the result is a column vector of product balances.
}
\examples{
result <- UKEnergy2000mats |>
  dplyr::filter(LastStage \%in\% c("Final", "Useful")) |>
  tidyr::pivot_wider(names_from = matrix.name, values_from = matrix) |>
  calc_inter_industry_balance()
result[[Recca::balance_cols$inter_industry_balance_colname]][[1]]
result[[Recca::balance_cols$inter_industry_balance_colname]][[2]]
}
\seealso{
\code{\link[=verify_inter_industry_balance]{verify_inter_industry_balance()}}
}
