% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{calc_io_mats}
\alias{calc_io_mats}
\title{Calculate various input-output matrices}
\usage{
calc_io_mats(.sutdata, U_colname = "U", V_colname = "V", Y_colname = "Y",
  y_colname = "y", q_colname = "q", g_colname = "g", W_colname = "W",
  Z_colname = "Z", D_colname = "D", C_colname = "C", A_colname = "A",
  L_ixp_colname = "L_ixp", L_pxp_colname = "L_pxp")
}
\arguments{
\item{.sutdata}{a data frame of supply-use table matrices with matrices arranged in columns.}

\item{U_colname}{the name of the column in \code{.sutdata} containing Use (\code{U}) matrices.}

\item{V_colname}{the name of the column in \code{.sutdata} containing Make (\code{V}) matrices.}

\item{Y_colname}{the name of the column in \code{.sutdata} containing final demand (\code{Y}) matrices.}

\item{y_colname}{the name of the output column containing \code{y} vectors.
\code{y} is calculated by \code{rowsums(Y)}.}

\item{q_colname}{the name of the output column containing \code{q} vectors.
\code{q} is calculated by \code{rowsums(U) + y}.}

\item{g_colname}{the name of the output column containing \code{g} vectors.
\code{g} is calculated by \code{rowsums(V)}.}

\item{W_colname}{the name of the output column containing \code{W} matrices.
\code{W} is calculated by \code{transpose(V) - U}.}

\item{Z_colname}{the name of the output column containing \code{Z} matrices.
\code{Z} is calculated by \code{U * g_hat_inv}.}

\item{D_colname}{the name of the output column containing \code{D} matrices.
\code{D} is calculated by \code{V * q_hat_inv}.}

\item{C_colname}{the name of the output column containing \code{C} matrices.
\code{C} is calculated by \code{transpose(V) * g_hat_inv}.}

\item{A_colname}{the name of the output column containing \code{A} matrices.
\code{A} is calculated by \code{Z * D}.}

\item{L_ixp_colname}{the name of the output column containing industry-by-product L matrices.
\code{L_ixp} is calculated by \code{D * L_pxp}.}

\item{L_pxp_colname}{the name of the output column containing product-by-product L matrices.
\code{L_pxp} is calculated by \code{(I - Z*D)^-1}.}
}
\value{
a data frame containing columns specified in \code{keep_cols},
\code{y_colname}, \code{q_colname}, \code{g_colname}, \code{W_colname},
\code{Z_colname}, \code{D_colname}, \code{C_colname}, \code{A_colname},
\code{L_ixp_colname}, and \code{L_pxp_colname}.
}
\description{
Calculate various input-output matrices
}
