% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embodied.R
\name{calc_M}
\alias{calc_M}
\title{Calculate embodied energy matrices}
\usage{
calc_M(
  .YqGHEdata = NULL,
  Y = "Y",
  q = "q",
  G = "G",
  E = "E",
  tol = 1e-04,
  M_p = "M_p",
  M_s = "M_s"
)
}
\arguments{
\item{.YqGHEdata}{A data frame containing columns of \strong{q} vectors
and \strong{Y}, \strong{G}, \strong{H}, and \strong{E} matrices.
\code{.YqGEdata} will likely have been obtained from the \code{calc_GH()} and \code{calc_E()} functions.}

\item{Y}{A final demand (\strong{Y}) matrix or name of the column in \code{.YqHGEdata} containing same. Default is "Y".}

\item{q}{A \strong{q} column vector or name of the column in \code{.YqHGEdata} containing same. Default is "q".}

\item{G}{A \strong{G} matrix or name of the column in \code{.YqHGEdata} containing same. Default is "G".}

\item{E}{An \strong{E} matrix or name of the column in \code{.YqHGEdata} containing same. Default is "E".}

\item{tol}{The allowable energy balance error.}

\item{M_p}{The name for matrices of embodied energy in products on output. Default is "M_p".
These matrices contain embodied products in rows and embodying products in columns.}

\item{M_s}{The name for matrices of embodied energy consumed by final demand sectors. Default is "M_s".
These matrices contain embodied products in rows and consuming final demand sectors in columns.}
}
\value{
A list or data frame of embodied energy matrices.
}
\description{
\strong{M} is calculated by \code{e_hat * G},
but \code{e_hat} contains lists of matrices,
so the \strong{M} is also contain lists of matrices.
In each list, there is one \strong{M} matrix for each Product in the Energy conversion chain.
}
