% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{calc_mats_start_positions_excel}
\alias{calc_mats_start_positions_excel}
\title{Calculate the origin and extent for each matrix}
\usage{
calc_mats_start_positions_excel(
  R,
  U,
  V,
  Y,
  r_eiou,
  U_eiou,
  U_feed,
  S_units,
  pad = 2
)
}
\arguments{
\item{R, U, V, Y, r_eiou, U_eiou, U_feed, S_units}{Matrices to be arranged on an Excel worksheet.}

\item{pad}{The number of blank rows or columns between matrices.}
}
\value{
A nested list of origins and extents.
}
\description{
The origin is defined as the upper-left corner of the matrix on the worksheet.
The extent is defined as the lower-right corner of the matrix on the worksheet.
}
\details{
The outer structure of the return value is matrices,
in the order provided in the argument list.
The inner structure of the return value is a list of "origin" and "extent,"
in that order.

This is a helper function, so it is not public.
}
