% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{calc_yqgW}
\alias{calc_yqgW}
\title{Calculate y, g, and q vectors and W matrices}
\usage{
calc_yqgW(.sutdata = NULL, U_colname = "U", V_colname = "V",
  Y_colname = "Y", S_units = "S_units", y_colname = "y",
  q_colname = "q", g_colname = "g", W_colname = "W")
}
\arguments{
\item{.sutdata}{a data frame of supply-use table matrices with matrices arranged in columns.}

\item{U_colname}{the name of the column in \code{.sutdata} containing Use (\code{U}) matrices.}

\item{V_colname}{the name of the column in \code{.sutdata} containing Make (\code{V}) matrices.}

\item{Y_colname}{the name of the column in \code{.sutdata} containing final demand (\code{Y}) matrices.}

\item{S_units}{the name of the column in \code{.sutdata} containing \code{S_units} matrices.}

\item{y_colname}{the name of the output column containing \code{y} vectors.
\code{y} is calculated by \code{rowsums_byname(Y)}.}

\item{q_colname}{the name of the output column containing \code{q} vectors.
\code{q} is calculated by \code{rowsums_byname(U) + y}.}

\item{g_colname}{the name of the output column containing \code{g} vectors.
\code{g} is calculated by \code{rowsums_byname(V)}.}

\item{W_colname}{the name of the output column containing \code{W} matrices.
\code{W} is calculated by \code{transpose_byname(V) - U}.}
}
\value{
\code{.sutdata} with columns \code{y_colname}, \code{q_colname}, \code{g_colname}, and \code{W_colname} added
}
\description{
Calculate y, g, and q vectors and W matrices
}
