% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{combine_RV}
\alias{combine_RV}
\title{Combine resource (\code{R}) and make (\code{V}) matrices into a make plus resource (\code{R_plus_V}) matrix}
\usage{
combine_RV(.sutdata = NULL, R_colname = "R", V_colname = "V",
  R_plus_V_colname = "R_plus_V")
}
\arguments{
\item{.sutdata}{a list or data frame containing use matrix(ces) and make matrix(ces)}

\item{R_colname}{identifier for the resource matrix (a string). Default is "\code{R}".}

\item{V_colname}{identifier for the make matrix (a string). Default is "\code{V}".}

\item{R_plus_V_colname}{identifier for the make matrix in \code{.sutdata} (a string). Default is "\code{R_plus_V}".}
}
\value{
\code{sum_by_name(R, V)}
}
\description{
Combine resource (\code{R}) and make (\code{V}) matrices into a make plus resource (\code{R_plus_V}) matrix
}
\examples{
library(dplyr)
library(tidyr)
UKEnergy2000mats \%>\%
  spread(key = "matrix.name", value = "matrix") \%>\%
  # Delete next line when switch to using R everywhere
  rename(R_plus_V = V) \%>\% separate_RV() \%>\% select(-R_plus_V) \%>\%
  combine_RV()
}
