% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cooling_eu.products}
\alias{cooling_eu.products}
\title{A list of Cooling useful work products}
\format{
A list with 2 entries.
\describe{
\item{LTC.-10.C}{The useful work product (Eu.product) "LTC.-10.C"}
\item{LTC.20.C}{The useful work product (Eu.product) "LTC.20.C"}
}
}
\usage{
cooling_eu.products
}
\description{
A list of Cooling useful work products
}
\keyword{datasets}
