% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecc_layout.R
\name{ecc_layout}
\alias{ecc_layout}
\title{Layout for graph representation of an energy conversion chain}
\usage{
ecc_layout(industry_stages, product_stages, fd_stage, scbsd_stage,
  industry_stage_names_colname = "Industries",
  industry_stage_colname = "Stage", industry_group_colname = "Groups",
  product_stage_names_colname = "Products", product_stage_colname = "Stage",
  product_group_colname = "Groups")
}
\arguments{
\item{industry_stage_name}{the column in \code{industry_stages} containing names of industries
(a string)}

\item{industry_stage_number}{the column in \code{industry_stages} containing the left-to-right
position of this stage (a string)}

\item{industry_group}{the name of an optional column in \code{industry_stages} containing
names of groups for industries (a string).
The order of groups in the graph is determined by the top-to-bottom
order in which group names appear in the \code{industry_stages} data frame.}

\item{product_group}{}
}
\description{
Information for the layout is provided by the following arguments:
\code{industry_stages},
\code{product_stages},
\code{fd_stage}, and
\code{scbsd_stage},
all of which are data frames.
}
