% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{edge_list}
\alias{edge_list}
\title{Create an edge list}
\usage{
edge_list(.sutdata = NULL, U = "U", V = "V", Y = "Y",
  edge_list = "Edge list", from = "From", to = "To", value = "Value",
  product = "Product", waste = "Waste", node_id = "node_id",
  edge_id = "edge_id", simplify_edges = TRUE)
}
\arguments{
\item{.sutdata}{Optionally, a data frame containing columns named with the values of the \code{U}, \code{V}, and \code{Y} arguments.}

\item{U}{a use matrix of the name of a column in \code{.sutdata} containing use matrices. (Default is "\code{U}".)}

\item{V}{a make matrix of the name of a column in \code{.sutdata} containing make matrices. (Default is "\code{V}".)}

\item{Y}{a final demand matrix of the name of a column in \code{.sutdata} containing final demant matrices. (Default is "\code{Y}".)}

\item{edge_list}{the name of the column in the output data frame containing edge lists. (Default is "\code{Edge list}".)}

\item{from}{the name of the edge list column containing source nodes. (Default is "\code{From}".)}

\item{to}{the name of the edge list column containing destination nodes. (Default is "\code{To}".)}

\item{value}{the name of the edge list column containing magnitudes of the flows. (Default is "\code{Value}".)}

\item{product}{the name of the edge list column containing the product of the edge list flow. (Default is "\code{Product}".)}

\item{waste}{the name of the waste product and the destination node for wastes.
Set \code{NULL} to suppress addition of waste edges. (Default is "\code{Waste}".)}

\item{simplify_edges}{if \code{TRUE}, products with only one source will not have a node.
The source of the product will be connected directly to its consumers. If \code{FALSE}, no simplifications are made.
(Default is \code{TRUE}.)}
}
\value{
an edge list or a column of edge lists
}
\description{
An edge list is a data frame in which each row describes a flow from one entity to another.
Columns in the edge list data frame are \code{From}, \code{To}, \code{Value}, and \code{Product}.
The edge list can be created from the \code{U}, \code{V}, and \code{Y} matrices of an energy conversion chain.
Edge lists is a typical data format for visulaization software.
}
\details{
Optionally, waste streams can be calculated from the \code{U} and \code{V} matrices and
added to the edge list.
Optionally, edges can be simplified when a product has a single source.
In that event, the node named after the product is removed,
and destinations are connected to the sources.
}
\examples{
sutmats <- UKEnergy2000mats \%>\% spread(key = matrix.name, value = matrix)
# Don't simplify edges and don't include waste edges
el_basic <- edge_list(sutmats, simplify_edges = FALSE, include_waste = FALSE)
head(el_basic$`Edge list`[[1]])
tail(el_basic$`Edge list`[[1]])
# Simplify edges and include waste
el <- edge_list(sutmats)
head(el$`Edge list`[[1]])
# Now includes waste edges
tail(el$`Edge list`[[1]])
# Works with single matrices, too.
elmats <- edge_list(U = sutmats$U[[1]], V = sutmats$V[[1]], Y = sutmats$Y[[1]])
head(elmats[["Edge list"]])
tail(elmats[["Edge list"]])
}
