% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregates.R
\name{effects_aggregates}
\alias{effects_aggregates}
\title{Title}
\usage{
effects_aggregates(
  .sut_data = NULL,
  p_industries,
  fd_sectors,
  pattern_type = c("exact", "leading", "trailing", "anywhere"),
  unnest = FALSE,
  method = c("solve", "QR", "SVD"),
  tol_invert = .Machine$double.eps,
  tol_chop_sum = 1e-04,
  R = Recca::psut_cols$R,
  U = Recca::psut_cols$U,
  U_feed = Recca::psut_cols$U_feed,
  V = Recca::psut_cols$V,
  Y = Recca::psut_cols$Y,
  S_units = Recca::psut_cols$S_units,
  aggregates_df = Recca::aggregate_cols$aggregates_df,
  product_sector = Recca::aggregate_cols$product_sector,
  aggregate_primary = Recca::aggregate_cols$aggregate_primary,
  net_aggregate_demand = Recca::aggregate_cols$net_aggregate_demand,
  gross_aggregate_demand = Recca::aggregate_cols$gross_aggregate_demand,
  .prime = "_prime",
  R_colname = Recca::psut_cols$R,
  U_colname = Recca::psut_cols$U,
  U_feed_colname = Recca::psut_cols$U_feed,
  U_eiou_colname = Recca::psut_cols$U_eiou,
  r_eiou_colname = Recca::psut_cols$r_eiou,
  V_colname = Recca::psut_cols$V,
  Y_colname = Recca::psut_cols$Y,
  R_prime_colname = paste0(R_colname, .prime),
  U_prime_colname = paste0(U_colname, .prime),
  U_feed_prime_colname = paste0(U_feed_colname, .prime),
  U_eiou_prime_colname = paste0(U_eiou_colname, .prime),
  r_eiou_prime_colname = paste0(r_eiou_colname, .prime),
  V_prime_colname = paste0(V_colname, .prime),
  Y_prime_colname = paste0(Y_colname, .prime)
)
}
\arguments{
\item{p_industries}{A vector of names of industries to be aggregated as "primary."
If \code{.sut_data} is a data frame, \code{p_industries} should be the name of a column in the data frame.
If \code{.sut_data} is \code{NULL}, \code{p_industries} can be a single vector of industry names.
These industries in \code{p_industries} will appear in rows of the resource (\strong{R}) and make (\strong{V}) matrices and
columns of the final demand matrix (\strong{Y}).
Entries in \strong{Y_p} will be subtracted from entries in \strong{R_p} \code{+} \strong{V_p} to obtain
the total primary energy aggregate,
where \verb{*_p} is the primary part of those matrices.
The function \code{find_p_industry_names()} might be helpful to find
primary industry names if they can be identified by prefixes.
This argument is passed to \code{primary_aggregates()}.}

\item{fd_sectors}{A vector of names of sectors in final demand.
Names should include columns in the \strong{Y} and \strong{U_EIOU} matrices
to cover both net (in \strong{Y}) and gross (in \strong{Y} and \strong{U_EIOU}) final demand.
This argument is passed to \code{finaldemand_aggregates()}.}

\item{pattern_type}{One of "exact", "leading", "trailing", or "anywhere" which specifies
how matches are made for \code{p_industries}.
If "exact", exact matches specify the sectors to be aggregated.
If "leading", sectors are aggregated if any entry in \code{p_industries} matches the leading part of a final demand sector's name.
If "trailing", sectors are aggregated if any entry in \code{p_industries} matches the trailing part of a final demand sector's name.
If "anywhere", sectors are aggregated if any entry in \code{p_industries} matches any part of a final demand sector's name.
Default is "exact".
This argument is passed to both \code{primary_aggregates()} and \code{finaldemand_aggregates()}.}

\item{unnest}{A boolean that tells whether to unnest the outgoing data.
When \code{TRUE}, creates a new column called \code{product_sector} and columns of primary and final demand aggregates.
Default is \code{FALSE}.}

\item{Y_prime_colname}{}
}
\description{
Title
}
