% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{efficiency_cols}
\alias{efficiency_cols}
\title{Efficiency data frame column names}
\format{
A string list with 4 entries.
\describe{
\item{eta_i}{The name of a column in a wide-by-matrices data frame containing industry efficiencies.}
\item{eta_pfd_gross}{The name of a column in a wide-by-matrices data frame containing  efficiencies between the primary and gross final demand stages.}
\item{eta_pfd_net}{The name of a column in a wide-by-matrices data frame containing efficiencies between the primary and net final demand stages.}
\item{efficiency_name_suffix}{The suffix for names of columns containing efficiency names.}
}
}
\usage{
efficiency_cols
}
\description{
A string list containing named names of columns in efficiency data frames.
Items in the list provide default values for column name function arguments
to efficiency functions throughout the \code{Recca} package.
}
\examples{
efficiency_cols
}
\keyword{datasets}
