% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endogenize_losses.R
\name{endogenize_losses}
\alias{endogenize_losses}
\title{Endogenize losses into PSUT matrices}
\usage{
endogenize_losses(
  .sutmats = NULL,
  R = Recca::psut_cols$R,
  U = Recca::psut_cols$U,
  V = Recca::psut_cols$V,
  Y = Recca::psut_cols$Y,
  intra_industry_balance = Recca::balance_cols$intra_industry_balance_colname,
  losses_alloc = Recca::balance_cols$losses_alloc_colname,
  losses_sector = Recca::balance_cols$losses_sector,
  replace_cols = FALSE,
  clean = FALSE,
  tol = 1e-06,
  V_prime = "V_prime",
  Y_prime = "Y_prime"
)
}
\arguments{
\item{.sutmats}{A \code{matsindf} data frame, wide by matrices, or
a list of lists of matrices.
Default is \code{NULL}.}

\item{R}{Resources (\strong{R}) matrix or name of the column in \code{.sutmats}
that contains same.
Default is \link{psut_cols}\verb{$R} or
"R".}

\item{U}{Use (\strong{U}) matrix or name of the column in \code{.sutmats}
that contains same.
Necessary for verifying calculating losses.
Default is \link{psut_cols}\verb{$U} or
"U".}

\item{V}{Make (\strong{V}) matrix or name of the column in \code{.sutmats}
that contains same.
Default is \link{psut_cols}\verb{$V} or
"V".}

\item{Y}{Final demand (\strong{Y}) matrix or name
of the column in \code{.sutmats} that contains same.
Default is \link{psut_cols}\verb{$Y} or
"Y".}

\item{intra_industry_balance}{A vector or the name of the column containing
intra-industry balance vectors.
If missing, losses are calculated internally
with \code{\link[=calc_intra_industry_balance]{calc_intra_industry_balance()}}
before endogenizing.
Default is
\link{balance_cols}\verb{$intra_industry_balance_colname} or
"SUTIntraIndustryBalance".}

\item{losses_alloc}{A matrix or the name of the column containing
loss allocation matrices.
See details for structure of this matrix.
Default is \link{balance_cols}\verb{$losses_alloc_colname} or
"LossesAlloc".}

\item{losses_sector}{The string name of the sector
that will absorb losses in the \strong{Y} matrix.
Default is \link{balance_cols}\verb{$losses_sector}
or "Transformation losses".}

\item{replace_cols}{A boolean that tells whether to
(a) replace
the \code{V} and \code{Y} columns with
\code{V_prime} and \code{Y_prime} columns, respectively and
(b) delete the \code{V_prime}, \code{Y_prime}, \code{balance_colname}, and
\code{losses_alloc_colname} columns
after endogenizing the losses
when \code{.sutmats} is a data frame or a list.
Default is \code{FALSE}.}

\item{clean}{A boolean that tells whether the outgoing
\code{V_prime} and \code{Y_prime} matrices should have
\code{0} rows and columns removed.
Default is \code{FALSE}.}

\item{tol}{The maximum allowable difference from \code{1} for the rowsums of
loss allocation matrices.
Default is \code{1e-6}.}

\item{V_prime}{The name of the \strong{V} matrix with endogenized losses.
Default is "V_prime".}

\item{Y_prime}{The name of the \strong{Y} matrix with endogenized losses.
Default is "Y_prime".}
}
\value{
A version of the conversion chain with losses endogenized.
}
\description{
When a conversion chain does \emph{not} include losses in the
\strong{RUVY} matrices of the PSUT framework,
it may be helpful to endogenize the losses.
This function performs the endogenization.
}
\details{
This function endogenizes losses into the \strong{V} and \strong{Y} matrices,
because losses are made (\strong{V}) by industries
and gathered by final demand (\strong{Y}).
By default, this function creates new
\code{V_prime} and \code{Y_prime} matrices.
Setting \code{replace_cols = TRUE}
(default is \code{FALSE})
replaces the existing
\code{V} and \code{Y} matrices with \code{V_prime} and \code{Y_prime},
respectively.

All losses are allocated to the
\code{losses_sector} column in the \strong{Y} matrix,
by default named
\link{balance_cols}\verb{$losses_sector or "}r Recca::balance_cols$losses_sector`".
\subsection{Endogenizing algorithm}{

The endogenizing algorithm is this:
\itemize{
\item If not present,
calculate the balance vector with
\code{\link[=calc_intra_industry_balance]{calc_intra_industry_balance()}}.
\item Hatize the balance vector
with \code{\link[matsbyname:hatize_byname]{matsbyname::hatize_byname()}}.
\item Matrix multiply the hatized balance vector and
the losses_allocation matrix (\code{losses_alloc})
to obtain a matrix to be added to \strong{V}.
\item Transpose the matrix to be added to \strong{V},
calculate rowsums, and
set the column name to \code{losses_sector}
to obtain a matrix to be added to \strong{Y}.
\item Add the matrices to \strong{V} and \strong{Y}, respectively.
}
}

\subsection{Losses allocation matrix}{

The losses allocation matrix (\code{losses_alloc}) tells how losses from
each industry should be allocated to products.
The losses allocation matrix should have
industries in rows and
products (losses) in rows.
The values in the losses allocation matrix are fractions
of each industry's losses (in rows) that are allocated
to loss products (in columns).
Rows of losses allocation matrix must sum to \code{1}.
The value of \code{losses_alloc} must resolve to a matrix
of this form.
Options include:
\itemize{
\item The string name of a column in \code{.sutmats}
that contains loss allocation matrices for every row
in \code{.sutmats}.
The default value is
\link{balance_cols}\verb{$losses_alloc_colname},
the string "LossesAlloc".
All industries in the \strong{V} matrix must be present
in the rows of the matrices in the \code{losses_alloc} column.
If any of the matrices in the column names \code{losses_alloc}
has a single row,
the row is assumed to apply to all industries.
\item A reasonable value, such as
\link{balance_cols}\verb{$default_losses_alloc},
a 1x1 matrix with
a row named "All industries",
a column named "Waste heat", and
a value of \code{1}.
This default matrix ascribes losses from all industries
in the \strong{V} matrix
to a product called "Waste heat".
\item A matrix of the same sense
(industries in rows,
waste products in columns,
rows sum to 1.0)
to be applied to all
rows in the wide-by-matrices data frame supplied
in \code{.sutmats}.
All industries in the \strong{V} matrix must be present
in the rows of \code{losses_alloc}.
If the matrix has a single row
(as \link{balance_cols}\verb{$default_losses_alloc}),
it is assumed to apply to all industries.
}
}

\subsection{Intra-industry balances}{

The intra-industry losses to be endogenized
are found in \code{intra_industry_balance}
and can be calculated with \code{\link[=calc_intra_industry_balance]{calc_intra_industry_balance()}}.
If \code{intra_industry_balance} is not present,
it is calculated internally via
\code{\link[=calc_intra_industry_balance]{calc_intra_industry_balance()}}.
}

\subsection{Cleaning}{

When this function operates
on a conversion chain with already-balanced industries,
the matrices to be added to \strong{V} and \strong{Y}
will be the \strong{0} matrix to within \code{tol}.
Setting \code{clean = TRUE} removes those rows or columns
from the output \code{V_prime} and \code{Y_prime} matrices.
}

\subsection{Checks}{

Prior to performing any calculations,
\code{\link[=verify_inter_industry_balance]{verify_inter_industry_balance()}} is checked.
There is no point endogenizing losses for
a conversion chain that is not internally consistent.

After endogenizing the losses,
all industries in the conversion chain
should pass \code{\link[=verify_intra_industry_balance]{verify_intra_industry_balance()}},
a condition that is checked before returning.
}
}
\examples{
mats <- UKEnergy2000mats |>
  tidyr::pivot_wider(names_from = matrix.name,
                     values_from = matrix) |>
  dplyr::filter(.data[[IEATools::iea_cols$last_stage]] \%in\%
                  c(IEATools::last_stages$final, IEATools::last_stages$useful)) |>
  dplyr::mutate(
    # Add a matrix column of loss allocations.
    # This bit of code adds a default loss allocation matrix
    # to every row of the data frame.
    "{Recca::balance_cols$losses_alloc_colname}" :=
      RCLabels::make_list(Recca::balance_cols$default_losses_alloc,
                          n = dplyr::n(),
                          lenx = 1)
    )
dplyr::glimpse(mats)
mats |>
  calc_intra_industry_balance() |>
  endogenize_losses() |>
  dplyr::glimpse()
# Replace original matrices with endogenized matrices
mats |>
  calc_intra_industry_balance() |>
  endogenize_losses(replace_cols = TRUE) |>
  # Check the intra-industry balance.
  # Everything should be balanced now.
  calc_intra_industry_balance() |>
  verify_intra_industry_balance() |>
  dplyr::glimpse()
}
