% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{energy_types}
\alias{energy_types}
\title{A list of energy types}
\format{
A list with 3 entries.
\describe{
\item{energy_type}{The name of the energy type column, "EnergyType".}
\item{e}{energy}
\item{x}{exergy}
}
}
\usage{
energy_types
}
\description{
A list of energy type options in the "EnergyType" column.
}
\keyword{datasets}
