% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{euproduct_aggregation_map}
\alias{euproduct_aggregation_map}
\title{A list of Name - Eu.product list combinations}
\format{
A list with 10 entries.
}
\usage{
euproduct_aggregation_map
}
\description{
A list of Name - Eu.product list combinations for use in aggregating individual
useful work products in to groups
}
\keyword{datasets}
