% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy_to_exergy.R
\name{extend_fu_detailed_to_exergy}
\alias{extend_fu_detailed_to_exergy}
\title{Extend the final-to-useful details matrices from energy to exergy}
\usage{
extend_fu_detailed_to_exergy(
  .fu_detailed_mats = NULL,
  Y_fu_detailed = Recca::psut_cols$Y_fu_detailed,
  U_eiou_fu_detailed = Recca::psut_cols$U_eiou_fu_detailed,
  clean_up_df = TRUE,
  phi = Recca::psut_cols$phi,
  .exergy_suffix = "_exergy",
  mat_piece = "noun",
  phi_piece = "all",
  energy_type = Recca::psut_cols$energy_type,
  mat_col_notation = RCLabels::from_notation,
  mat_colname_preposition =
    RCLabels::prepositions_list[[which(RCLabels::prepositions_list == "from")]],
  Y_fu_detailed_colname = Recca::psut_cols$Y_fu_detailed,
  U_eiou_fu_detailed_colname = Recca::psut_cols$U_eiou_fu_detailed,
  phi_colname = Recca::psut_cols$phi,
  energy = Recca::energy_types$e,
  exergy = Recca::energy_types$x,
  industry_type = IEATools::row_col_types$industry,
  product_type = IEATools::row_col_types$product
)
}
\arguments{
\item{phi}{}
}
\description{
The details matrices contain (in row and column names)
details about the move from the final energy stage to
the useful energy stage.
Four pieces of information are provided:
\itemize{
\item final energy product,
\item final demand sector,
\item useful energy product, and
\item final-to-useful machine.
}
}
\details{
Two details matrices are available:
\itemize{
\item \code{Y_fu_details} and
\item \code{U_EIOU_fu_details}.
}

The format for the row and column names for both details matrices is
\itemize{
\item row names
\itemize{
\item \code{RCLabels::arrow_notation}
\item prefix: final energy product
\item suffix: final demand sector
\item example: "Aviation gasoline -> Domestic aviation"
}
\item column names
\itemize{
\item \code{RCLabels::from_notation}
\item noun: useful energy product
\item object of from: final-to-useful machine
\item example: "HPL [from Electric pumps]"
}
}

The energy stage of the entries in the details matrices are indicated
by the entry in the \code{Energy.type} column,
typically "Useful".
}
