% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy_to_exergy.R
\name{extend_one_matrix_to_exergy}
\alias{extend_one_matrix_to_exergy}
\title{Convert one matrix from energy to exergy}
\usage{
extend_one_matrix_to_exergy(
  .m,
  phi_vec,
  product_margin,
  mat_piece = "all",
  phi_piece = "all",
  notation = RCLabels::bracket_notation,
  prepositions = RCLabels::prepositions_list
)
}
\arguments{
\item{.m}{An energy matrix to be converted to exergy.}

\item{phi_vec}{A vector of exergy-to-energy ratios (phi).}

\item{product_margin}{The margin of \code{.m} containing products,
\code{1} means rows, \code{2} means columns.}

\item{mat_piece}{The piece of row and column names for \code{.m}
against which row names of the \code{phi} vector is to be matched.
Default is "all", meaning that entire names are to be matched.}

\item{phi_piece}{The piece of row names in the \code{phi} vector against which
row and column names for \code{.m}
is to be matched.
Default is "all", meaning that entire names are to be matched.}

\item{notation}{The nomenclature for the row and column labels.
Default is \code{RCLabels::bracket_notation}.}

\item{prepositions}{The prepositions to be used row and column notation.
Default is \code{RCLabels::prepositions_list}.}
}
\value{
A matrix with same size as \code{.m} but converted from energy to exergy.
}
\description{
A helper function for
\code{\link[=extend_to_exergy]{extend_to_exergy()}}
and
\code{\link[=extend_fu_details_to_exergy]{extend_fu_details_to_exergy()}}
that is helpful in other circumstances, too.
}
