% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy_to_exergy.R
\name{extend_to_exergy}
\alias{extend_to_exergy}
\title{Extend an ECC in PSUT format from energy to exergy}
\usage{
extend_to_exergy(
  .sutmats = NULL,
  clean_up_df = TRUE,
  R = Recca::psut_cols$R,
  U = Recca::psut_cols$U,
  U_feed = Recca::psut_cols$U_feed,
  U_eiou = Recca::psut_cols$U_eiou,
  r_eiou = Recca::psut_cols$r_eiou,
  V = Recca::psut_cols$V,
  Y = Recca::psut_cols$Y,
  phi = Recca::psut_cols$phi,
  .exergy_suffix = "_exergy",
  R_name = Recca::psut_cols$R,
  U_name = Recca::psut_cols$U,
  U_feed_name = Recca::psut_cols$U_feed,
  U_eiou_name = Recca::psut_cols$U_eiou,
  r_eiou_name = Recca::psut_cols$r_eiou,
  V_name = Recca::psut_cols$V,
  Y_name = Recca::psut_cols$Y,
  phi_name = Recca::psut_cols$phi,
  energy_type = IEATools::iea_cols$energy_type,
  energy = Recca::energy_types$e,
  exergy = Recca::energy_types$x,
  s_units = Recca::psut_cols$s_units
)
}
\arguments{
\item{.sutmats}{An optional data frame of energy conversino chain matrices.}

\item{clean_up_df}{When \code{.sutmats} is a data frame, tells whether to \code{tidyr::pivot_longer()} the result,
remove no-longer-needed input column \code{phi}, and
fill the \code{energy_type} column with "X" for the exergy versions of the ECC matrices.
Default is \code{TRUE}.}

\item{R, U, U_feed, U_eiou, r_eiou, V, Y, phi}{Names of columns in \code{.sutmats} or single matriecs. See \code{Recca::psut_cols}.}

\item{.exergy_suffix}{The string suffix to be appended to exergy versions of columns.}

\item{R_name, U_name, U_feed_name, U_eiou_name, r_eiou_name, V_name, Y_name}{Names of output matrices}
}
\value{
A data frame or list of matrices that represent the exergy version of the ECC.
}
\description{
An energy conversion chain can be represented in energy or exergy quantifications of energy.
This function moves from an energy quantification to an exergy quantification,
given the matrices for the energy quantification and
phi (exergy-to-energy ratio) vectors.
}
\details{
Internally, this function uses \code{matsindf::apply}, so
the ECC matrices can be provided
as individual matrices
or in a data frame
(in which case the arguments should
given the string names of columns in the \code{.sutmats} data frame).
}
