% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/losses_irrev.R
\name{extend_to_exergy_with_losses_irrev}
\alias{extend_to_exergy_with_losses_irrev}
\title{Calculate exergy losses and irreversibility}
\usage{
extend_to_exergy_with_losses_irrev(
  .sutmats = NULL,
  clean_up_df = TRUE,
  R = Recca::psut_cols$R,
  U = Recca::psut_cols$U,
  V = Recca::psut_cols$V,
  Y = Recca::psut_cols$Y,
  U_feed = Recca::psut_cols$U_feed,
  U_eiou = Recca::psut_cols$U_eiou,
  r_eiou = Recca::psut_cols$r_eiou,
  intra_industry_balance = Recca::balance_cols$intra_industry_balance_colname,
  losses_alloc = Recca::balance_cols$losses_alloc_colname,
  losses_sector = Recca::balance_cols$losses_sector,
  phi = Recca::psut_cols$phi,
  irrev_alloc = Recca::balance_cols$irrev_alloc_colname,
  irrev_sector = Recca::balance_cols$irrev_sector,
  replace_cols = FALSE,
  clean = FALSE,
  tol = 1e-06,
  .exergy_suffix = "_exergy",
  energy_type = Recca::psut_cols$energy_type,
  exergy = Recca::energy_types$x,
  R_exergy = paste0(Recca::psut_cols$R, .exergy_suffix),
  U_exergy = paste0(Recca::psut_cols$U, .exergy_suffix),
  V_exergy = paste0(Recca::psut_cols$V, .exergy_suffix),
  Y_exergy = paste0(Recca::psut_cols$Y, .exergy_suffix),
  U_feed_exergy = paste0(Recca::psut_cols$U_feed, .exergy_suffix),
  U_eiou_exergy = paste0(Recca::psut_cols$U_eiou, .exergy_suffix),
  r_eiou_exergy = paste0(Recca::psut_cols$r_eiou, .exergy_suffix)
)
}
\arguments{
\item{.sutmats}{An optional wide-by-matrices data frame
containing conversion chains quantified
in conserved quantities (such as mass or energy).
Losses may be included in the matrices,
but that is not required.}

\item{R}{Resources (\strong{R}) matrix or name of the column in \code{.sutmats}
that contains same.
Default is \link{psut_cols}\verb{$R} or
"R".}

\item{U}{Use (\strong{U}) matrix or name of the column in \code{.sutmats}
that contains same.
Necessary for verifying calculating losses.
Default is \link{psut_cols}\verb{$U} or
"U".}

\item{V}{Make (\strong{V}) matrix or name of the column in \code{.sutmats}
that contains same.
Default is \link{psut_cols}\verb{$V} or
"V".}

\item{Y}{Final demand (\strong{Y}) matrix or name
of the column in \code{.sutmats} that contains same.
Default is \link{psut_cols}\verb{$Y} or
"Y".}

\item{U_feed}{The feedstock portion of the use (\strong{U}) matrix
or name of the column in \code{.sutmats} that contains same.
Default is \link{psut_cols}\verb{$U_feed} or
"U_feed".}

\item{U_eiou}{The energy industry own use portion of the use (\strong{U}) matrix
or name of the column in \code{.sutmats} that contains same.
Default is \link{psut_cols}\verb{$U_eiou} or
"U_EIOU".}

\item{r_eiou}{A matrix of the ratio of energy industry own use
to total use for the use (\strong{U}) matrix
or name of the column in \code{.sutmats} that contains same.
Default is \link{psut_cols}\verb{$r_eiou} or
"r_EIOU".}

\item{intra_industry_balance}{A vector or the name of the column containing
intra-industry balance vectors for the conserved quantity.
If missing, losses are calculated internally
with \code{\link[=calc_intra_industry_balance]{calc_intra_industry_balance()}}
before endogenizing.
Default is
\link{balance_cols}\verb{$intra_industry_balance_colname} or
"SUTIntraIndustryBalance".}

\item{losses_alloc}{A matrix or the name of the column containing
loss allocation matrices.
See details for structure of this matrix.
Default is \link{balance_cols}\verb{$losses_alloc_colname} or
"LossesAlloc".}

\item{losses_sector}{The string name of the sector
that will absorb losses in the \strong{Y} matrix for the conserved quantity.
Default is \link{balance_cols}\verb{$losses_sector}
or "Transformation losses".}

\item{irrev_alloc}{An irreversibility allocation matrix or the name of a column in
\code{.sutmats} containing same.
Default is \link{balance_cols}\verb{$irrev_alloc_colname} or
"IrrevAlloc".}

\item{irrev_sector}{The string name of the sector that will absorb irreversibilities
(exergy destruction)
in the \strong{Y} matrix.
Default is \link{balance_cols}\verb{$irrev_sector}
or "Irreversibilities".}

\item{replace_cols}{A boolean that tells whether to
(a) replace
the \code{V} and \code{Y} columns with
\code{V_prime} and \code{Y_prime} columns, respectively and
(b) delete the \code{V_prime}, \code{Y_prime}, \code{balance_colname}, and
\code{losses_alloc_colname} columns
after endogenizing the losses
when \code{.sutmats} is a data frame or a list.
Default is \code{FALSE}.}

\item{clean}{A boolean that tells whether the outgoing
\code{V_prime} and \code{Y_prime} matrices should have
\code{0} rows and columns removed.
Default is \code{FALSE}.}

\item{tol}{The maximum allowable difference from \code{1} for the rowsums of
loss allocation matrices.
Default is \code{1e-6}.}

\item{phi_vec}{A vector of exergy-to-energy ratios (phi)
or the string name of a column in \code{.sutmats} containing same.
Default is \link{psut_cols}\verb{$phi} or
"phi".}

\item{irreversibility}{}
}
\description{
Given a conversion chain quantified in conserved quantities,
such as mass or energy,
calculate an exergy version of the conversion chain with
endogenized losses and irreversibility
(exergy destruction)
}
\details{
The input should be a conversion chain
that contains conserved quantities (probably mass or energy).
Optionally, the conversion chain may already
have endogenized transformation losses (wastes).
The output contains a new version of the input conversion chain
in exergy terms with all transformation losses endogenized and
exergy destruction (irreversibility) calculated.

Between input and output, the following steps are taken:
\enumerate{
\item Transformation losses of the conserved quantity
are calculated and endogenized into the \strong{V} and \strong{Y} matrices
via a call to \code{\link[=endogenize_losses]{endogenize_losses()}}.
If \code{intra_industry_balance} is \code{NULL},
losses of the conserved quantity are calculated
within \code{\link[=endogenize_losses]{endogenize_losses()}}
via a call to \code{\link[=calc_intra_industry_balance]{calc_intra_industry_balance()}}.
If the input already has endogenized losses,
there will be no change to the incoming conversion chain.
The \code{losses_alloc} and \code{losses_sector} arguments are passed
to \code{\link[=endogenize_losses]{endogenize_losses()}}.
\item All matrices of the conversion chain are extended to exergy
via a call to \code{\link[=extend_to_exergy]{extend_to_exergy()}}.
The \code{phi} argument is passed to \code{\link[=extend_to_exergy]{extend_to_exergy()}}.
Transformation losses calculated previously (in Step 1.)
are now exergy losses,
i.e., the exergy of transformation losses.
\item Transformation losses of the (un-conserved) exergy flows are
endogenized into the \strong{V} and \strong{Y} matrices
via another call to \code{\link[=endogenize_losses]{endogenize_losses()}}.
These transformation losses are interpreted as
exergy destruction (irreversibility).
The \code{irrev_alloc} and \code{irrev_sector} arguments are apassed
to \code{\link[=endogenize_losses]{endogenize_losses()}}.
}

Throughout the function,
inter-industry balances are verified via
\code{\link[=verify_inter_industry_balance]{verify_inter_industry_balance()}}.
After Steps 1. and 3.,
across-industry balances are verfied via
\code{\link[=verify_intra_industry_balance]{verify_intra_industry_balance()}}.

The new version of the conversion chain is contained in named arguments
(or columns if \code{.sutmats} is a data frame) named with the suffix
\code{.exergy_suffix}, by default "_exergy".
If \code{.sutmats} is a data frame,
a tidy data frame can be output with an \code{energy_type} column
by setting \code{clean_up_df = TRUE} (the default).
The \code{energy_type} column is filled with
\code{energy_type_value}, by default
\link{energy_types}\verb{$x} or "X".

To extend to exergy without
calculation of losses and irreversibility,
use \code{\link[=extend_to_exergy]{extend_to_exergy()}}.
}
\seealso{
\code{\link[=extend_to_exergy]{extend_to_exergy()}} for a version of this function that
does not endogenize losses of the conserved quantity and
does not calculate exergy destruction (irreversibility).
}
