% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{extract_R}
\alias{extract_R}
\title{Extract resource matrices (\code{R}) from make plus resource (\code{V_plus_R}) and use (\code{U}) matrices}
\usage{
extract_R(.sutdata = NULL, U_colname = "U",
  V_plus_R_colname = "V_plus_R", V_colname = "V", R_colname = "R")
}
\arguments{
\item{.sutdata}{a list or data frame containing use matrix(ces) and make matrix(ces)}

\item{U_colname}{identifier for the use matrix (a string) \code{.sutdata}. Default is "\code{U}".}

\item{V_plus_R_colname}{identifier for the make matrix in \code{.sutdata} (a string). Default is "\code{V_plus_R}".}

\item{V_colname}{identifier for the make matrix (a string). Default is "\code{V}".}

\item{R_colname}{identifier for the resource matrix (a string). Default is "\code{R}".}
}
\value{
a version of the \code{V} matrix without resource industries and an \code{R} matrix
}
\description{
Resource industries are industries that make a product without using any products.
Resource industries are identified by interrogating
the use (\code{U}) and make (\code{V_plus_R}) matrices.
Resource industries have all zeroes in their column of the use matrix (\code{U})
and at least one non-zero value in their row of the make (\code{V_plus_R}) matrix.
}
\details{
A resource matrix (\code{R}) has industries in rows and products in columns.
The elements of of \code{R} indicate extraction of resources from the biosphere.
The industries of \code{R} are the reserves of the extracted products.

This function uses the \code{\link{resource_industries}} function to
identify the resource industries in the \code{V_plus_R} matrix.
Thereafter, the function extracts the resource industries from the \code{V_plus_R} matrix
to form the \code{R} matrix.
Finally, the \code{R} matrix is subtracted from the \code{V_plus_R} matrix
and saved as the \code{V} matrix.
If there are no resource industries in the \code{V_plus_R} matrix,
a warning is emitted,
no \code{R} matrix is created, and
no changes are made to the \code{V_plus_R} matrix.
}
