% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregates.R
\name{finaldemand_aggregates_IEA}
\alias{finaldemand_aggregates_IEA}
\title{Final demand aggregate energy from IEA tables}
\usage{
finaldemand_aggregates_IEA(.ieadata, UVY_colname = "UVY",
  energy_colname = "EX.ktoe", Y_fd_varname = "Y_fd",
  Y_ne_varname = "Y_ne", U_EIOU_varname = "U_EIOU",
  aggregate_net_finaldemand_colname, aggregate_gross_finaldemand_colname)
}
\arguments{
\item{.ieadata}{a data frame with columns of IEA data.}

\item{UVY_colname}{the name of the column in \code{.ieadata}
containing names of submatrices of U, V, and G to which this
row of data belongs.}

\item{energy_colname}{the name of the column in \code{.ieadata}
that contains energy data.}

\item{Y_fd_varname}{the name of the variable in \code{.ieadata} containing final demand (\code{Y_fd}) matrices.
\code{Y_fd} will typically contain energy consumption by industries, transport, and other.}

\item{aggregate_net_finaldemand_colname}{the name of the output column containing aggregates of net energy demand.
Each entry in this column is \code{sumall(Y_fd)}.}

\item{aggregate_gross_finaldemand_colname}{the name of the output column containing aggregates of gross energy demand.
Each entry in this column is calculated by \code{sumall(Y_fd)} + \code{sumall(U_EIOU)}.}

\item{U_EIOU_colname}{the name of the column in \code{.ieadata} containing Energy industry own use
(\code{U_EIOU}) matrices.}

\item{keep_cols}{the names of colums in \code{.sutdata} to retain in the output}
}
\value{
a two-column data fram containing the gross aggregate energy demand
and net aggregate energy demand for each row of \code{.sutdata}.
}
\description{
Calculates aggregate final demand energy from a data frame of IEA data.
}
