% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregates.R
\name{finaldemand_aggregates_with_units}
\alias{finaldemand_aggregates_with_units}
\title{Final demand aggregate energy with units}
\usage{
finaldemand_aggregates_with_units(
  .sutdata,
  fd_sectors,
  U = Recca::psut_cols$U,
  Y = Recca::psut_cols$Y,
  r_EIOU = Recca::psut_cols$r_eiou,
  S_units = Recca::psut_cols$s_units,
  by = c("Total", "Product", "Sector"),
  net_aggregate_demand = Recca::aggregate_cols$net_aggregate_demand,
  gross_aggregate_demand = Recca::aggregate_cols$gross_aggregate_demand
)
}
\arguments{
\item{.sutdata}{A data frame with columns of matrices from a supply-use analysis.}

\item{fd_sectors}{A vector of names of sectors in final demand.
Names should include columns in the \code{Y} and \code{U_EIOU} matrices
to cover both net (in \code{Y}) and gross (in \code{Y} and \code{U_EIOU}) final demand.}

\item{U}{Use (\code{U}) matrix or name of the column in \code{.sutdata} containing same}

\item{Y}{Final demand (\code{Y}) matrix or name of the column in \code{.sutdata} containing same}

\item{r_EIOU}{Matrix of ratios of EIOU for the make (\code{U}) matrix or name of the column in \code{.sutdata} containing same}

\item{S_units}{The name of the column in \code{.sutdata} containing \code{S_units} matrices.}

\item{by}{One of "Product", "Sector", or "Total" to indicate the desired aggregation:
"Product" for aggregation by energy carrier (Crude oil, Primary solid biofuels, etc.),
"Sector" for aggregation by final demand sector (Agriculture/forestry, Domestic navigation, etc.), or
"Total" for aggregation over both Product and Sector (the default).}

\item{net_aggregate_demand}{The name of net energy demand (which excludes energy industry own use) on output.}

\item{gross_aggregate_demand}{The name of gross energy demand (which includes energy industry own use) on output.}
}
\value{
A list or data frame containing net aggregate energy demand
and gross aggregate energy demand
}
\description{
Calculates aggregate final demand energy and services from a data frame of Supply-Use matrices.
The calculation includes non-energy uses if they are present in the final demand matrix.
The calculation does not include balancing items (Losses and Statistical differences).
If \code{.sutdata} is a data frame, \code{fd_sectors} should be the name of a column in the data frame.
If \code{.sutdata} is \code{NULL}, \code{fd_sectors} can be a single vector of industry names.
}
