% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{flows_unit_homogeneous}
\alias{flows_unit_homogeneous}
\title{Tell whether industry flows (inputs and outputs) are unit-homogeneous}
\usage{
flows_unit_homogeneous(.sutdata = NULL, U_colname = "U",
  V_colname = "V", S_units_colname = "S_units", keep_details = FALSE,
  flows_unit_homogeneous_colname = "flows_unit_homogeneous")
}
\arguments{
\item{.sutdata}{a data frame of supply-use table matrices with matrices arranged in columns.}

\item{U_colname}{the name of the column in \code{.sutdata} that contains
\code{U} matrices. Default is "\code{U}".}

\item{V_colname}{the name of the column in \code{.sutdata} that contains
\code{V} matrices. Default is "\code{V}".}

\item{S_units_colname}{the name of the column in \code{.sutdata} that contains
\code{S_units} matrices. Default is "\code{S_units}".}

\item{keep_details}{if \code{TRUE}, per-industry results are returned;
if \code{FALSE}, per-ECC results are returned.}

\item{flows_unit_homogeneous_colname}{the name of the output column
that tells whether each industry's outputs are unit-homogeneous.
Default is "\code{flows_unit_homogeneous}".}
}
\value{
\code{.sutdata} with additional column "\code{flows_unit_homogeneous}" containing
        \code{TRUE} if each industry's flows are unit-homogeneous,
        \code{FALSE} if each industry's flows are unit-heterogeneous.
}
\description{
Returns \code{TRUE} if each industry's flows (all inputs and outputs) are unit-homogeneous.
}
\details{
The \code{V_bar} matrix is queried for the number of non-zero entries in each row.
If the number of non-zero entries in each row is exactly 1,
industry outputs are unit-homogeneous.
Note that \code{V_bar = \link[matsbyname]{matrixproduct_byname}(V, S_units)}.
}
\examples{
library(magrittr)
library(tidyr)
UKEnergy2000mats \%>\%
  spread(key = "matrix.name", value = "matrix") \%>\%
  flows_unit_homogeneous()
}
