% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregates.R
\name{footprint_aggregates}
\alias{footprint_aggregates}
\title{Calculate footprint aggregates}
\usage{
footprint_aggregates(
  .sut_data = NULL,
  fd_sectors,
  pattern_type = c("exact", "leading", "trailing", "anywhere"),
  R = Recca::psut_cols$R,
  U = Recca::psut_cols$U,
  V = Recca::psut_cols$V,
  Y = Recca::psut_cols$Y,
  r_EIOU = Recca::psut_cols$r_eiou,
  by = c("Total", "Product", "Sector"),
  add_primary_net_gross_cols = FALSE,
  aggregate_primary = Recca::aggregate_cols$aggregate_primary,
  net_aggregate_primary = Recca::aggregate_cols$net_aggregate_primary,
  gross_aggregate_primary = Recca::aggregate_cols$gross_aggregate_primary,
  net_aggregate_demand = Recca::aggregate_cols$net_aggregate_demand,
  gross_aggregate_demand = Recca::aggregate_cols$gross_aggregate_demand
)
}
\arguments{
\item{fd_sectors}{A vector of names of sectors in final demand.
Names should include columns in the \code{Y} and \code{U_EIOU} matrices
to cover both net (in \code{Y}) and gross (in \code{Y} and \code{U_EIOU}) final demand.}

\item{gross_aggregate_demand}{}

\item{p_industries}{a vector of names of industries to be aggregated as "primary."
If \code{.sut_data} is a data frame, \code{p_industries} should be the name of a column in the data frame.
If \code{.sut_data} is \code{NULL}, \code{p_industries} can be a single vector of industry names.
These industries in \code{p_industries} will appear in rows of the resource (\code{R}) and make (\code{V}) matrices and
columns of the final demand matrix (\code{Y}).
Entries in \code{Y_p} will be subtracted from entries in \code{R_p + V_p} to obtain
the total primary energy aggregate,
where \verb{*_p} is the primary part of those matrices.
The function \code{find_p_industry_names()} might be helpful to find
primary industry names if they can be identified by prefixes.}
}
\description{
A footprint aggregate are isolated measures of primary and final demand energy
required to supply a specific amount of final demand energy.
This function calculates footprint aggregates for several categories of final demand.
}
\details{
By default, footprint aggregates are calculated for each individual
product and sector of final demand in the \code{Y} matrix.
This calculation is accomplished for each description of an energy conversion chain (ECC)
by the following algorithm:
\enumerate{
\item Calculate io matrices with \code{calc_io_mats()}.
\item Identify each product and sector from rows and columns of the \code{Y} matrix.
\item For each product and sector independently,
perform an upstream swim with \code{new_Y()}
to obtain the ECC requirements to supply that product or sector.
\item Calculate primary and final demand aggregates using \code{primary_aggregates()} and
\code{finaldemand_aggregates()}.
\item Add the primary and final demand aggregates as columns at the right side of \code{.sut_data}.
}
}
