% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregates.R
\name{group_aggregates}
\alias{group_aggregates}
\title{Perform grouping aggregations on PSUT matrices}
\usage{
group_aggregates(
  .sut_data = NULL,
  aggregation_map,
  margin = c(1, 2),
  pattern_type = "exact",
  R = Recca::psut_cols$R,
  U = Recca::psut_cols$U,
  V = Recca::psut_cols$V,
  Y = Recca::psut_cols$Y,
  r_EIOU = Recca::psut_cols$r_eiou,
  U_EIOU = Recca::psut_cols$U_eiou,
  U_feed = Recca::psut_cols$U_feed,
  S_units = Recca::psut_cols$S_units,
  R_aggregated_colname = paste0(Recca::psut_cols$R, aggregated_suffix),
  U_aggregated_colname = paste0(Recca::psut_cols$U, aggregated_suffix),
  V_aggregated_colname = paste0(Recca::psut_cols$V, aggregated_suffix),
  Y_aggregated_colname = paste0(Recca::psut_cols$Y, aggregated_suffix),
  r_EIOU_aggregated_colname = paste0(Recca::psut_cols$r_eiou, aggregated_suffix),
  U_EIOU_aggregated_colname = paste0(Recca::psut_cols$U_eiou, aggregated_suffix),
  U_feed_aggregated_colname = paste0(Recca::psut_cols$U_feed, aggregated_suffix),
  S_units_aggregated_colname = paste0(Recca::psut_cols$S_units, aggregated_suffix),
  aggregated_suffix = Recca::aggregate_cols$aggregated_suffix
)
}
\arguments{
\item{.sut_data}{A data frame of matrices to be despecified and aggregated.}

\item{aggregation_map}{Aggregation details. See documentation for \code{matsbyname::aggregate_byname()} for further information.}

\item{margin}{\code{1}, \code{2}, or \code{c(1, 2)} for row aggregation, column aggregation, or both.
Can be a row or column type.
Default is \code{c(1, 2)}.}

\item{pattern_type}{See \code{RCLabels::make_or_pattern()}.
Default is "exact".}

\item{R, U, V, Y, r_EIOU, U_EIOU, U_feed, S_units}{Matrices or names of columns in \code{.sut_data} to be despecified and aggregated. See \code{Recca::psut_cols}.}

\item{R_aggregated_colname, U_aggregated_colname, V_aggregated_colname, Y_aggregated_colname, r_EIOU_aggregated_colname, U_EIOU_aggregated_colname, U_feed_aggregated_colname, S_units_aggregated_colname}{Names of
aggregated matrices or columns.}

\item{aggregated_suffix}{A string suffix used to form the names for aggregated matrices.
Default is "_aggregated".}
}
\value{
PSUT matrices aggregated according to \code{aggregation_map}.
}
\description{
It is often helpful to aggregate data into industry or product categories,
such as "Anthracite" and "Brown coal" to "Coal and coal products" or
"Domestic aviation" and "Domestic navigation" to "Transport".
With the help of an \code{aggregation_map}, this function
performs such aggregations for a set of PSUT matrices.
}
\details{
Internally, this function uses \code{matsbyname::aggregate_byname()}.
See its documentation for details on the format for the \code{aggregation_map}.
}
\examples{
UKEnergy2000mats \%>\%
  tidyr::pivot_wider(names_from = matrix.name, values_from = matrix) \%>\%
  group_aggregates(aggregation_map = list(`Oil and oil products` =
                                            c("Crude", "Diesel", "Petrol")),
                   pattern_type = "leading",
                   margin = "Product")
}
