% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hpm_eu.products}
\alias{hpm_eu.products}
\title{A list of High pressure matter (HPM) useful work products}
\format{
A list with 3 entries.
\describe{
\item{HPA}{The useful work product (Eu.product) "High pressure air"}
\item{HPL}{The useful work product (Eu.product) "High pressure liquids"}
\item{HPNG}{The useful work product (Eu.product) "High pressure natural gas"}
}
}
\usage{
hpm_eu.products
}
\description{
A list of High pressure matter (HPM) useful work products.
}
\keyword{datasets}
