% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hth_eu.products}
\alias{hth_eu.products}
\title{A list of High temperature heat (HTH) useful work products}
\format{
A list with 7 entries.
\describe{
\item{HTH.400.C}{The useful work product (Eu.product) "HTH.400.C"}
\item{HTH.600.C}{The useful work product (Eu.product) "HTH.600.C"}
\item{HTH.850.C}{The useful work product (Eu.product) "HTH.850.C"}
\item{HTH.960.C}{The useful work product (Eu.product) "HTH.960.C"}
\item{HTH.1000.C}{The useful work product (Eu.product) "HTH.1000.C"}
\item{HTH.1300.C}{The useful work product (Eu.product) "HTH.1300.C"}
\item{HTH.1600.C}{The useful work product (Eu.product) "HTH.1600.C"}
}
}
\usage{
hth_eu.products
}
\description{
A list of High temperature heat (HTH) useful work products
}
\keyword{datasets}
