% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecc_layout.R
\name{identify_industry_stages}
\alias{identify_industry_stages}
\title{Identify industry stages}
\usage{
identify_industry_stages(
  GeneralIndustries,
  SpecificIndustries,
  general_industry_colname = "Industry",
  stage_colname = "Stage",
  specific_industry_colname = "Industry"
)
}
\arguments{
\item{GeneralIndustries}{a data frame containing at least two columns,
\code{general_industry_colname} and \code{stage_colname}.}

\item{SpecificIndustries}{a data frame containing at least \code{specific_industry_colname}.}

\item{general_industry_colname}{a string for the name of the column in \code{GeneralIndustries}
that contains the general names of industries (such as "\code{Production}").
Default is "\code{Industry}".}

\item{stage_colname}{a string for the name of the column in \code{GeneralIndustries}
that contains the names of ECC stages (such as "\code{Primary industry}",
"\code{Primary-to-final industry}", or "\code{Final-to-useful industry}").
Default is "\code{Stage}".}

\item{specific_industry_colname}{a string for the name
of the column in \code{SpecificIndustries} that contains the specific names of industries
(such as "\code{Production - Hydro}").
Default is "\code{Industry}".}
}
\value{
a data frame containing \code{specific_industry_colname} and \code{stage_colname},
        wherein the stages associated with general industry names in \code{GeneralIndustries}
        are assigned to the specific industries from \code{SpecificIndustries}
        according to matches between \code{GeneralIndustries$general_industry_colname}
        and \code{SpecificIndustries$specific_industry_colname}.
}
\description{
Names for industries in an ECC may be comprised of a general industry name
and the name of specific product,
such as "\code{Production - Hydro}".
Stages of an ECC are most easily identified simply by the general industry name
(e.g., all "\code{Production}" industries are \code{primary} industries),
but simple matching of specific industry names with general industry names to identify
industry stages will fail ("\code{Production - Hydro}" does not match "\code{Production}", e.g.).
This function assigns ECC stages to specific industry names
using a table of general industry names and associated ECC stages.
A general industry name matches a specific industry name if the specific industry name
starts with or is equal to the general industry name.
}
\examples{
GeneralIndustries <- data.frame(Industry = c("Production", "Plants", "Cars"),
                                Stage = c("Primary industry", "p-->f industry",
                                          "f-->u industry"), stringsAsFactors = FALSE)
SpecificIndustries <- data.frame(Industry = c("Production - Coal", "Production - Hydro",
                                              "Plants - Electric", "Plants - refinery",
                                              "Cars - Electric", "Cars - Petrol"),
                                              stringsAsFactors = FALSE)
identify_industry_stages(GeneralIndustries = GeneralIndustries,
                         SpecificIndustries = SpecificIndustries)
}
