% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{industry_cols}
\alias{industry_cols}
\title{Primary industry column names}
\format{
A string list with 2 entries.
\describe{
\item{p_industries_prefixes}{The name of a column in a wide-by-matrices data frame containing prefixes for names of primary industries.}
\item{p_industries_complete}{The name of a column in a wide-by-matrices data frame containing complete names of primary industries.}
}
}
\usage{
industry_cols
}
\description{
A string list containing named names of columns in SUT data frames.
Items in the list provide default values for column name function arguments
throughout the \code{Recca} package.
}
\examples{
industry_cols
}
\keyword{datasets}
