% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{inputs_outputs_unit_homogeneous}
\alias{inputs_outputs_unit_homogeneous}
\title{Tell whether industry inputs are unit-homogeneous and industry outputs are unit-homogeneous}
\usage{
inputs_outputs_unit_homogeneous(.sutmats = NULL, U = "U", V = "V",
  S_units = "S_units", keep_details = FALSE,
  ins_outs_unit_homogeneous = ".inputs_outputs_unit_homogeneous")
}
\arguments{
\item{.sutmats}{a data frame of supply-use table matrices with matrices arranged in columns.}

\item{U}{a use (\code{U}) matrix or name of the column in \code{.sutmats} that contains same. Default is "\code{U}".}

\item{V}{a make (\code{V}) matrix or name of the column in \code{.sutmats} that contains same. Default is "\code{V}".}

\item{S_units}{an \code{S_units} matrix or name of a column in \code{.sutmats} that contains same. Default is "\code{S_units}".}

\item{keep_details}{if \code{TRUE}, per-industry results are returned;
if \code{FALSE}, per-ECC results are returned.}

\item{ins_outs_unit_homogeneous}{the name of the output column
that tells whether each industry's inputs and outputs are unit-homogeneous
(though not necessarily in the same units).
Default is "\code{.inputs_outputs_unit_homogeneous}".}
}
\value{
a list or data frame containing
        \code{TRUE} if inputs from each energy conversion industry are unit-homogeneous
        and outputs from each energy conversion industry are unit-homogeneous,
        \code{FALSE} otherwise.
}
\description{
Returns \code{TRUE} if each industry's inputs are unit-homogeneous and each industry's outputs are unit homogeneous.
When inputs have different units from outputs, (but all inputs are unit-homogeneous and all outputs are unit-homogeneous),
\code{TRUE} is returned.
}
\details{
This function uses both \link{inputs_unit_homogeneous} and \link{outputs_unit_homogeneous} internally.
This function differs from \link{flows_unit_homogeneous}, because \link{flows_unit_homogeneous}
requires that all flows are unit-homogeneous before returning \code{TRUE}.
This function (\link{inputs_outputs_unit_homogeneous}) will return true when all
inputs are unit-homogeneous with different units from outputs (which are also unit-homogeoenous).
}
\seealso{
\link{inputs_unit_homogeneous}, \link{outputs_unit_homogeneous}, \link{flows_unit_homogeneous}
}
