% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ip_eu.products}
\alias{ip_eu.products}
\title{A list of Information processing useful work products}
\format{
A list with 1 entries.
\describe{
\item{IP}{The useful work product (Eu.product) Information processing}
}
}
\usage{
ip_eu.products
}
\description{
A list of Information processing useful work products
}
\keyword{datasets}
