% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{l_eu.products}
\alias{l_eu.products}
\title{A list of Light useful work products}
\format{
A list with 1 entries.
\describe{
\item{L}{The useful work product (Eu.product) Light}
}
}
\usage{
l_eu.products
}
\description{
A list of Light useful work products
}
\keyword{datasets}
