% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{load_expected_data}
\alias{load_expected_data}
\title{Load expected results}
\usage{
load_expected_data(expectation_name, dir = file.path("tests", "expectations"),
  prefix = "expected_")
}
\arguments{
\item{expectation_name}{the name of the object for which you want to load an expected result.
Do not include the "\code{expected_}" prefix.}

\item{dir}{the directory in which to look for expected results.
Default is \code{file.path("tests", "expectations")}}

\item{prefix}{the prefix to the file name.  Default is "\code{expected_}".}
}
\value{
the R object contained in the file named \code{<prefix><expectation_name>.rds}
        loaded from the directory at \code{path}.
}
\description{
Loads an expected result for a test.
Use \link[Recca]{use_expected_data} to save the expectation to disk.
}
\details{
The default directory for expectatations is \code{tests/expectations}.
}
\examples{
afortesting <- 5
use_expected_data(afortesting)
load_expected_data("afortesting")
}
