% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lth_eu.products}
\alias{lth_eu.products}
\title{A list of Low temperature heat (LTH) useful work products}
\format{
A list with 2 entries.
\describe{
\item{LTH.20.C}{The useful work product (Eu.product) "LTH.20.C"}
\item{LTH.60.C}{The useful work product (Eu.product) "LTH.60.C"}
}
}
\usage{
lth_eu.products
}
\description{
A list of Low temperature heat (LTH) useful work products
}
\keyword{datasets}
