% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{match_p_industry_names}
\alias{match_p_industry_names}
\title{Scrape primary industry names from R, V, and Y matrices}
\usage{
match_p_industry_names(
  .sutdata = NULL,
  p_industries_prefixes = Recca::industry_cols$p_industries_prefixes,
  R = Recca::psut_cols$R,
  V = Recca::psut_cols$V,
  Y = Recca::psut_cols$Y,
  p_industries_complete = Recca::industry_cols$p_industries_complete
)
}
\arguments{
\item{.sutdata}{An optional data frame containing columns of PSUT matrices}

\item{p_industries_prefixes}{A vector of prefixes for primary industry names.}

\item{R}{The name of the \code{R} matrix column in \code{.sutdata} or an \code{R} matrix.}

\item{V}{The name of the \code{V} matrix column in \code{.sutdata} or a \code{V} matrix.}

\item{Y}{The name of the \code{Y} matrix column in \code{.sutdata} or a \code{Y} matrix.}

\item{p_industries}{The name of the output column containing complete names of primary industries.
Default is \code{Recca::industry_cols$p_industries_complete}.}
}
\value{
A vector or vectors of full names of primary industries in the \code{R}, \code{V}, and \code{Y} matrices.
}
\description{
Primary industry names are needed for aggregation.
This function interrogates the row names of R and V and the column names Y matrices
for names that start with \code{p_industries}.
The assumption is that many of these row and column names may have
compound names of the form "Resources [of Oil and gas extraction]".
So this function looks for leading strings.
If "Resources" is in \code{p_industries},
"Resources [of Oil and gas extraction]" will be among the returned strings.
}
\details{
Note all of \code{R}, \code{V}, and \code{Y} need to be specified.
}
