% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matsindf_apply.R
\name{matsindf_apply}
\alias{matsindf_apply}
\title{Apply a function to a \code{matsindf} data frame (and more)}
\usage{
matsindf_apply(.DF = NULL, FUN, ...)
}
\arguments{
\item{.DF}{the \code{matsindf} data frame}

\item{FUN}{the function to be applied to \code{.DF}}

\item{...}{named arguments to be passed by name to \code{FUN}.}
}
\value{
a named list or a data frame. (See details.)
}
\description{
Applies \code{FUN} to \code{.DF} or
performs the calculation specified by \code{FUN}
on numbers or matrices.
\code{FUN} must return a named list.
}
\details{
If \code{...} are all named numbers or matrices
of the form \code{argname = m},
\code{.DF} is ignored, and
\code{m}s are passed to \code{FUN} by \code{argname}s.
The return value is a named list provided by \code{FUN}.

If \code{...} are all lists of numbers or matrices
of the form \code{argname = l},
\code{.DF} is ignored, and
\code{FUN} is \code{Map}ped across the various \code{l}s
to obtain a list of named lists returned from \code{FUN}.
The return value is a data frame
whose rows are the named lists returned from \code{FUN} and
whose column names are the names of the list items returned from \code{FUN}.
The series of named lists are \code{rbind}-ed to create the output data frame.
Columns of \code{.DF} are not present in the return value.

If \code{...} are all named character strings
of the form \code{argname = string},
\code{.DF} is required,
\code{string}s are expected to be column names in \code{.DF}, and
\code{argname}s are expected to be names of arguments to \code{FUN}.
The return value is \code{.DF} with additional columns (at right)
whose names are the names of list items returned from \code{FUN}.
}
