% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mechanical_eu.products}
\alias{mechanical_eu.products}
\title{A list of Mechanical useful work products}
\format{
A list with 4 entries.
\describe{
\item{MD}{The useful work product (Eu.product) Mechanical drive}
\item{KE}{The useful work product (Eu.product) Kinetic energy}
\item{MW}{The useful work product (Eu.product) Mechanical work}
\item{MF}{The useful work product (Eu.product) Material fracture}
}
}
\usage{
mechanical_eu.products
}
\description{
A list of Mechanical useful work products
}
\keyword{datasets}
