% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mth_eu.products}
\alias{mth_eu.products}
\title{A list of Medium temperature heat (MTH) useful work products}
\format{
A list with 2 entries.
\describe{
\item{MTH.100.C}{The useful work product (Eu.product) "MTH.100.C"}
\item{MTH.200.C}{The useful work product (Eu.product) "MTH.200.C"}
}
}
\usage{
mth_eu.products
}
\description{
A list of Medium temperature heat (MTH) useful work products
}
\keyword{datasets}
