% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{neu_eu.products}
\alias{neu_eu.products}
\title{A list of Non-energy use useful work products}
\format{
A list with 1 entries.
\describe{
\item{NEU}{The useful work product (Eu.product) Non-energy use}
}
}
\usage{
neu_eu.products
}
\description{
A list of Non-energy use useful work products
}
\keyword{datasets}
