% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruct.R
\name{new_k_ps}
\alias{new_k_ps}
\title{Assess the effect of changing perfectly substitutable intermediate inputs to an industry}
\usage{
new_k_ps(.sutdata = NULL, k_prime_colname = "k_prime",
  U_colname = "U", V_colname = "V", Y_colname = "Y",
  K_colname = "K", L_ixp_colname = "L_ixp", L_pxp_colname = "L_pxp",
  Z_colname = "Z", D_colname = "D", f_colname = "f",
  U_prime_colname = "U_prime", V_prime_colname = "V_prime")
}
\arguments{
\item{.sutdata}{a data frame of supply-use table matrices with matrices arranged in columns.}

\item{k_prime_colname}{the name of a column in \code{.sutdata} containing vectors representing new
columns of the \code{K} matrix in each row of \code{.sutdata}.
Each entry in the \code{k_prime} column of \code{.sutdata} must be a single-column matrix, and
the name of the single column must match the name of one of the columns of matrix \code{K}.}

\item{U_colname}{the name of a column in \code{.sutdata} containing \code{U} matrices for the base ECC.  Default is "\code{U}".}

\item{V_colname}{the name of a column in \code{.sutdata} containing \code{V} matrices for the base ECC.  Default is "\code{V}".}

\item{Y_colname}{the name of a column in \code{.sutdata} containing \code{Y} matrices for the base ECC.  Default is "\code{Y}".}

\item{K_colname}{the name of a column in \code{.sutdata} containing product-by-industry \code{K} matrices.  Default is "\code{K}".
\code{K} consists of columns that sum to 1.
Elements of \code{K} indicate the fraction of total input to industries (in columns)
provided by products (in rows).
\code{K} can be calculated by \code{\link{calc_io_mats}}.}

\item{L_ixp_colname}{the name of a column in \code{.sutdata} containing \code{L_ixp} matrices for the base ECC.  Default is "\code{L_ixp}".}

\item{L_pxp_colname}{the name of a column in \code{.sutdata} containing \code{L_pxp} matrices for the base ECC.  Default is "\code{L_pxp}".}

\item{Z_colname}{the name of a column in \code{.sutadata} containing \code{Z} matrices for the base ECC.  Default is "\code{Z}".}

\item{D_colname}{the name of a column in \code{.sutadata} containing \code{D} matrices for the base ECC.  Default is "\code{D}".}

\item{f_colname}{the name of a column in \code{.sutadata} containing \code{f} vectors for the base ECC.  Default is "\code{f}".}

\item{U_prime_colname}{the name of the output column that contains new Use (\code{U}) matrices.  Default is "\code{U_prime}".}

\item{V_prime_colname}{the name of the output column that contains new Make (\code{V}) matrices.  Default is "\code{V_prime}".}
}
\value{
\code{.sutdata} with additional columns \code{U_prime_colname} and \code{V_prime_colname}
}
\description{
This function calculates the effect of changing perfectly-substitutable (ps) inputs
to an intermediate industry.
New versions of \code{U} and \code{V} matrices are returned
as \code{U_prime} and \code{V_prime}.
Final demand (\code{Y}) is unchanged.
}
\details{
Note that inputs \code{K_colname}, \code{L_ixp_colname}, \code{L_pxp_colname},
\code{Z_colname}, \code{D_colname}, and \code{f_colname} can be
conveniently calculated by the function \code{\link{calc_io_mats}}.

Internally, this function uses \code{\link[matsindf]{matsindf_apply}},
and documentation assumes that
\code{.sutdata} is not \code{NULL} and is a data frame.
If \code{.sutdata} is present, output is a data frame with columns named by string values of output arguments, and
input arguments should be character strings that name columns in \code{.sutdata}.
If \code{.sutdata} is \code{NULL} (the default), output is a list with items named by output strings,
and input arguments should be single matrices or vectors.
}
\examples{
library(dplyr)
library(magrittr)
library(matsbyname)
library(tidyr)
# To demonstrate calculating changes to an energy conversion chain due to changes
# in perfectly-substitutable inputs to an intermediate industry,
# we use the PerfectSubmats data frame.
# But we need to calculate several important input-output matrices first.
io_mats <- PerfectSubmats \%>\%
  spread(key = "matrix.name", value = "matrix") \%>\%
  calc_io_mats()
# Next, find the K matrix that contains the fraction of each type of energy
# that enters each industry
K <- io_mats$K[[1]]
# Develop a new column vector for inputs to the Electric transport sector.
# As provided, the Electric transport sector is dominated by Renewable elec.
# What if the electricity input to the Electric transport sector
# were split 50/50 between Renewable elect and FF elec?
k_prime_vec <- K[, "Electric transport", drop = FALSE]
k_prime_vec["FF elec", "Electric transport"] <- 0.5
k_prime_vec["Ren elec", "Electric transport"] <- 0.5
# Add k_prime_vec to the io_mats data frame.
io_mats <- io_mats \%>\%
  mutate(
    # Set up a new k_prime vector for Electric transport.
    # That vector will be used for the infininte substitution calculation.
    k_prime = select_cols_byname(K, retain_pattern = make_pattern("Electric transport",
                                                                  pattern_type = "exact")),
    k_prime = make_list(k_prime_vec, n = 1)
  )
# Now do the calculation of U_prime and V_prime matrices.
new_UV <- new_k_ps(io_mats)
# There is much more FF extraction now than before.
io_mats$U[[1]]["FF", "FF extraction"]
new_UV$U_prime[[1]]["FF", "FF extraction"]
}
