% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruct.R
\name{new_R_ps}
\alias{new_R_ps}
\title{Calculates downstream effects of a new level of extracted resources}
\usage{
new_R_ps(
  .sutmats = NULL,
  R_prime = "R_prime",
  U = "U",
  V = "V",
  Y = "Y",
  q = "q",
  f = "f",
  U_prime = "U_prime",
  V_prime = "V_prime",
  W_prime = "W_prime",
  Y_prime = "Y_prime"
)
}
\arguments{
\item{.sutmats}{a data frame of supply-use table matrices with matrices arranged in columns.}

\item{R_prime}{The name of the new R matrix column in the input data frame, for which the new ECC must be assessed.
Default is "R_prime".}

\item{U}{The name of the U matrix column in the input data frame.
Default is "U".}

\item{V}{The name of the V matrix column in the input data frame.
Default is "V".}

\item{Y}{The name of the Y matrix column in the input data frame.
Default is "Y".}

\item{q}{The name of the q vector column in the input data frame.
Default is "q".}

\item{f}{The name of the f vector column in the input data frame.
Default is "f".}

\item{U_prime}{The name of the output column containing the new U matrices.
Default is "U_prime".}

\item{V_prime}{The name of the output column containing the new V matrices.
Default is "V_prime".}

\item{W_prime}{The name of the output column containing the new W matrices.
Default is "W_prime".}

\item{Y_prime}{The name of the output column containing the new Y matrices.
Default is "Y_prime".}
}
\value{
A data frame with added columns representing each of the new \code{U_prime}, \code{V_prime}, \code{W_prime}, and \code{Ý_prime} matrices.
}
\description{
This function calculates the effect of changing the resources available to an ECC,
i.e. of a new resources matrix \code{R} on the rest of the ECC matrices (\code{U}, \code{V}, \code{W}, and \code{Ý}).
New versions of the \code{U}, \code{V}, \code{W}, and \code{Ý} matrices are returned,
and respectively called \code{U_prime}, \code{V_prime}, \code{W_prime}, and \code{Ý_prime}.
This function assumes that each industry's inputs are perfectly substitutable (ps).
}
\details{
Each industry must be unit-homogeneous on its inputs.
If not, a matrix populated with \code{NA} is returned as the result for
\code{U_prime}, \code{V_prime}, and \code{Y_prime}.
}
\examples{
UKEnergy2000mats \%>\%
tidyr::spread(key = "matrix.name", value = "matrix") \%>\%
# Calculate the input-output matrices which are inputs to the new_R function.
 calc_io_mats() \%>\%
# Make an R_prime matrix that gives twice the resource inputs to the economy.
 dplyr::mutate(
   R_prime = matsbyname::hadamardproduct_byname(2, R)
 ) \%>\%
# Now call the new_R function which will calculate
# updated U, V, and Y matrices (U_prime, V_prime, and Y_prime)
# given R_prime.
# Each of the *_prime matrices should be 2x their originals,
# because R_prime is 2x relative to R.
 new_R_ps()
}
