% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruct.R
\name{new_Y}
\alias{new_Y}
\title{Reconstruct an economy given a new final demand matrix}
\usage{
new_Y(
  .sutmats = NULL,
  suffix = "_prime",
  Y_prime = "Y_prime",
  L_ixp = "L_ixp",
  L_pxp = "L_pxp",
  Z = "Z",
  D = "D",
  R = "R",
  r = "r",
  h = "h",
  O = "O",
  U_prime = "U_prime",
  V_prime = "V_prime",
  W_prime = "W_prime",
  R_prime = "R_prime"
)
}
\arguments{
\item{.sutmats}{A data frame of supply-use table matrices with matrices arranged in columns.}

\item{suffix}{The suffix for new matrices. Default is "_prime".}

\item{Y_prime}{A new final demand matrix or name of a column in \code{.sutmats} containing same. Default is "\code{Y_prime}".}

\item{L_ixp}{An \code{L_ixp} matrix or name of a column in \code{.sutmats} containing same. Default is "\code{L_ixp}".}

\item{L_pxp}{An \code{L_pxp} matrix or name of a column in \code{.sutmats} containing same. Default is "\code{L_pxp}".}

\item{Z}{A \code{Z} matrix or name of a column in \code{.sutmats} containing same. Default is "\code{Z}".}

\item{D}{A \code{D} matrix or name of a column in \code{.sutmats} containing same. Default is "\code{D}".}

\item{R}{A \code{R} matrix or name of a column in \code{.sutmats} containing same. Default is "\code{R}".}

\item{r}{The name of the \code{r} vector.
Default is "r".}

\item{h}{The name of the \code{h} vector.
Default is "h".}

\item{O}{A \code{O} matrix or name of a column in \code{.sutmats} containing same. Default is "\code{O}".}

\item{U_prime}{The name for new \code{U} matrices. Default is "U_prime".}

\item{V_prime}{The name for new \code{V} matrices. Default is "V_prime".}

\item{W_prime}{The name for new \code{W} matrices. Default is "W_prime".}

\item{R_prime}{The name for new \code{R} matrices. Default is "R_prime".}
}
\value{
A list or data frame with \code{R_prime}, \code{U_prime}, and \code{V_prime} matrices.
}
\description{
When the final demand matrix changes from \strong{Y} to \strong{Y_prime},
this function calculates new
resource (\strong{R_prime}),
use (\strong{U_prime}),
feed (\strong{U_feed}),
energy industry own use (\strong{U_eiou}),
ratio (\strong{r_eiou}),
and make (\strong{V_prime}) matrices
that would be required to meet the new final demand (\strong{Y_prime}).
}
\details{
Note that inputs \strong{L_ixp}, \strong{L_pxp},
\strong{Z}, and \strong{D} can be
conveniently calculated by the function \code{calc_io_mats()}.

Internally, this function uses \code{matsindf::matsindf_apply()},
and documentation assumes that
\code{.sutmats} is not \code{NULL} and is a data frame.
But \code{.sutmats} can also be a named list of matrices.
Or matrices can be supplied individually to the \code{Y_prime}, \code{L_ixp},
\code{L_pxp}, \code{Z}, \code{D}, \code{h}, and \code{O} arguments.
If \code{.sutmats} is present, output is a data frame with columns named by string values of output arguments, and
input arguments should be character strings that name columns in \code{.sutmats}.
If \code{.sutmats} is \code{NULL} (the default), output is a list with items named by output strings,
and input arguments should be single matrices or vectors.
}
\examples{
library(dplyr)
library(matsbyname)
library(tidyr)
UKEnergy2000mats \%>\%
  spread(key = matrix.name, value = matrix) \%>\%
  select(Country, Year, Energy.type, Last.stage, R, U, U_feed, V, Y, r_EIOU, S_units) \%>\%
  calc_io_mats() \%>\%
  mutate(
    # Give new Y matrices that are double the existing Y matrices
    Y_prime = matsbyname::hadamardproduct_byname(2, Y)
  ) \%>\%
  # Should give U_prime and V_prime matrices that are double the existing U and V matrices
  new_Y()
}
