% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_edge_list.R
\name{node_list}
\alias{node_list}
\title{Create a node list}
\usage{
node_list(edge_list, from = "From", to = "To", node = "Node",
  node_id = "node_id")
}
\arguments{
\item{edge_list}{an edge list data frame.}

\item{from}{the name of the \code{edge_list} column containing names of source nodes. (Default is "\code{From}".)}

\item{to}{the name of the \code{edge_list} column containing names of destination nodes. (Default is "\code{To}".)}

\item{node}{the name of the output column containing node names. (Default is "\code{Node}".)}

\item{node_id}{the name of the output column containing node ID numbers. (Default is "\code{node_id}".)}
}
\value{
a node list
}
\description{
A node list is a data frame containing node names and associated node ID numbers.
This function creates a node list from an edge list, as shown in the examples.
}
\details{
See \code{\link{edge_list}} for a function to create edge lists.
}
\examples{
sutmats <- UKEnergy2000mats \%>\% spread(key = matrix.name, value = matrix)
el <- edge_list(sutmats)$`Edge list`[[1]]
node_list(el)
}
