% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{outputs_unit_homogeneous}
\alias{outputs_unit_homogeneous}
\title{Tell whether industry outputs are unit-homogeneous}
\usage{
outputs_unit_homogeneous(.sutmats = NULL, V = "V",
  S_units = "S_units", keep_details = FALSE,
  outs_unit_homogeneous = ".outputs_unit_homogeneous")
}
\arguments{
\item{.sutmats}{a data frame of supply-use table matrices with matrices arranged in columns.}

\item{V}{a make (\code{V}) matrix or name of the column in \code{.sutmats} that contains same. Default is "\code{V}".}

\item{S_units}{an \code{S_units} matrix or name of a column in \code{.sutmats} that contains same. Default is "\code{S_units}".}

\item{keep_details}{if \code{TRUE}, per-industry results are returned;
if \code{FALSE}, per-ECC results are returned.}

\item{outs_unit_homogeneous}{the name of the output column
that tells whether each industry's outputs are unit-homogeneous.
Default is "\code{.outputs_unit_homogeneous}".}
}
\value{
a list or data frame containing
        \code{TRUE} if outputs from each energy conversion industry are unit-homogeneous,
        \code{FALSE} otherwise.
}
\description{
Returns \code{TRUE} if each industry's output are unit-homogeneous.
}
\details{
The \code{V_bar} matrix is queried for the number of non-zero entries in each row.
If the number of non-zero entries in each row is exactly 1,
industry outputs are unit-homogeneous.
Note that \code{V_bar = \link[matsbyname]{matrixproduct_byname}(V, S_units)}.
}
\examples{
library(tidyr)
UKEnergy2000mats \%>\%
  spread(key = "matrix.name", value = "matrix") \%>\%
  outputs_unit_homogeneous()
}
