% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PerfectSubtidy}
\alias{PerfectSubtidy}
\title{Example energy conversion chain to demonstrate perfect substitution}
\format{A data frame with 20 rows and 10 variables:
\describe{
  \item{Country}{country, (Example, only one country)}
  \item{Year}{year, 2000 (only one year as an example)}
  \item{Ledger.side}{Supply or Consumption}
  \item{Flow.aggregation.point}{tells where each row should be aggregated}
  \item{Energy.type}{E.ktoe (for energy) or X.ktoe (for exergy)}
  \item{Last.stage}{tells the final stage of the energy conversion chain: services is the only entry here}
  \item{Flow}{the Industry or Sector involved in this flow}
  \item{Product}{the energy product involved in this flow}
  \item{EX.ktoe}{value of the energy, exergy, or service flow in ktoe}
  \item{Unit}{unit in which quantity is expressed}
}}
\usage{
PerfectSubtidy
}
\description{
A dataset containing an example energy conversion chain
for the purpose of demonstrating changes to industry inputs
where the inputs are perfect substitutes.
}
\details{
\code{PerfectSubtidy} gives each non-zero entry in \code{PerfectSubmats}
as a single column in a data frame.
These data are in tidy format.
}
\keyword{datasets}
