% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{phi_vec}
\alias{phi_vec}
\title{A vector of phi (exergy-to-energy ratios) values}
\format{
A matrix with energy products in rows and phi values in a single column.
}
\usage{
phi_vec
}
\description{
Converting from energy to exergy requires vectors of phi values.
This object is a vector to assist that conversion process
for the \code{UKEnergy2000mats} energy conversion chain bundled with this package.
}
\keyword{datasets}
