% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency.R
\name{pivot_clean_complete_eta_pfd}
\alias{pivot_clean_complete_eta_pfd}
\title{Pivot, clean, and complete an efficiency data frame}
\usage{
pivot_clean_complete_eta_pfd(
  .eta_df,
  abbreviate_stage_names = TRUE,
  efficiency_name_suffix = Recca::efficiency_cols$efficiency_name_suffix,
  primary = "Primary",
  final = "Final",
  useful = "Useful",
  services = "Services",
  wellbeing = "Wellbeing",
  country = Recca::psut_cols$country,
  year = Recca::psut_cols$year,
  method = Recca::psut_cols$method,
  energy_type = Recca::psut_cols$energy_type,
  last_stage = Recca::psut_cols$last_stage,
  product_sector = Recca::aggregate_cols$product_sector,
  eta_pfd_gross = Recca::efficiency_cols$eta_pfd_gross,
  eta_pfd_net = Recca::efficiency_cols$eta_pfd_net,
  eta_pfd_gross_colname = paste0(eta_pfd_gross, efficiency_name_suffix),
  eta_pfd_net_colname = paste0(eta_pfd_net, efficiency_name_suffix),
  gross_net = Recca::efficiency_cols$gross_net,
  gross = Recca::efficiency_cols$gross,
  net = Recca::efficiency_cols$net,
  aggregate_primary = Recca::aggregate_cols$aggregate_primary,
  gross_aggregate_demand = Recca::aggregate_cols$gross_aggregate_demand,
  net_aggregate_demand = Recca::aggregate_cols$net_aggregate_demand,
  .eta = ".eta",
  .eta_type = ".eta_type",
  .eta_name = ".eta_name",
  .eta_stages = ".eta_stages"
)
}
\arguments{
\item{.eta_df}{A data frame of efficiencies, likely created by \code{calc_eta_pfd()}.}

\item{abbreviate_stage_names}{A boolean that tells whether stage names
are abbreviated to first letter.
E.g., "Primary" --> "p", "Final" --> "f", "Useful" --> "u", "Services" --> "s".
Default is \code{TRUE}.}
}
\value{
A cleaned version of \code{.eta_df}.
}
\description{
Efficiency data frames (created by \code{calc_eta_pfd()})
include (by default) columns of both
primary->final demand efficiencies and efficiency names.
The efficiency names are preferred to be columns to themselves.
The efficiency data frames can contain residual PSUT matrices,
which are cumbersome to carry around.
Furthermore, the efficiency data frames
contain only primary-to-final demand efficiencies,
not any intermediate efficiencies.
E.g., the data frame would contain
primary-to-final and primary-to-useful efficiencies
but not final-to-useful efficiencies.
This function cleans up these problems by
deleting columns of PSUT matrices (thereby cleaning the data frame),
pivoting to put all efficiencies in columns (thereby pivoting the data frame), and
calculating missing efficiencies (thereby completing the data frame).
}
\details{
This function knows about the following stages
in energy conversion chains: Primary, Final, Useful, Services, and Well-being.

The cleaning step eliminates all columns that contain exclusively matrices.
\code{matsindf::matrix_cols()} identifies the matrix columns.
}
