% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregates.R
\name{primary_aggregates_IEA}
\alias{primary_aggregates_IEA}
\title{Total primary aggregate energy from IEA tables}
\usage{
primary_aggregates_IEA(.ieadata, flow = "Flow", energy = "EX.ktoe",
  p_industries = c("Production", "Coal mines", "Oil and gas extraction",
  "Imports", "Exports", "International aviation bunkers",
  "International marine bunkers", "Stock changes"),
  aggregate_primary_colname = "EX_p_IEA.ktoe")
}
\arguments{
\item{.ieadata}{the data frame containing an IEA table}

\item{flow}{the name of the column that contains flow information. Default is "\code{Flow}".}

\item{energy}{the name of the column that contains energy information. Default is "\code{EX.ktoe}".}

\item{p_industries}{a vector of names of primary industries. Default is
"\code{c("Coal mines", "Oil and gas extraction",
"Production", "Imports", "Exports",
"International aviation bunkers", "International marine bunkers",
"Stock changes")}"}

\item{aggregate_primary_colname}{the name of the aggregate primary energy
column to be created in the output data frame.
Default is "\code{EX_p_IEA.ktoe}".}
}
\value{
a data frame containing the grouping columns of \code{.ieadata}
        as well as a column named with the value of \code{aggregate_primary_colname}.
}
\description{
Calculates total aggregate primary energy from a data frame of IEA data.
This function is named with "_IEA", because it is meant to operate on
tidy, IEA-style data frames.
The function \code{primary_aggregates} does the same thing,
but it is meant to operate on tidy SUT-style data frames.
}
\details{
This function works similar to \code{\link[dplyr]{summarise}}:
it distills \code{.ieadata} to many fewer rows
according to the grouping variables.
Thus, \code{.ieadata} should be grouped prior to sending into this function.
Grouping columns are preserved on output.
}
\examples{
library(magrittr)
p_ind <- primary_industries(UKEnergy2000mats \%>\%
  spread(key = matrix.name, value = matrix))[["p_industries"]][[1]]
UKEnergy2000tidy \%>\%
  group_by(Country, Year, Energy.type, Last.stage) \%>\%
  primary_aggregates_IEA(p_industries = p_ind)
}
