% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{primary_industries}
\alias{primary_industries}
\title{Primary industries}
\usage{
primary_industries(.sutdata = NULL, U = "U", V = "V",
  p_industries = "p_industries")
}
\arguments{
\item{.sutdata}{a list or data frame containing use matrix(ces) and make matrix(ces)}

\item{U}{identifier for the use matrix (a string). Default is "U".}

\item{V}{identifier for the make matrix (a string). Default is "Y".}

\item{p_industries}{name for the primary industries vector (a string). Default is "\code{p_industries}".}
}
\value{
\code{.sutdata} with an additional column (named with the value of the \code{p_industries} argument)
        containing the primary industries for each row
}
\description{
Identify primary industries
}
\details{
Primary industries are industries that make a product without using any products.
Primary industries are identified by interrogating
the use (\code{U}) and make (\code{V}) matrices.
Primary industries have all zeroes in their column of the use matrix (\code{U})
and at least one non-zero value in their row of the make (\code{V}) matrix.

Argument and value descriptions are written assuming that \code{.sutdata} is a data frame.
Alternatively, \code{.sutdata} can be unspecified, and \code{U} and \code{V} can be matrices.
In that case, the return value is a list with a single item: \code{p_industries}
which contains a vector of names of primary industries for the \code{U} and \code{V} matrices.
}
\examples{
library(magrittr)
library(tidyr)
primary_industries(UKEnergy2000mats \%>\% spread(key = matrix.name, value = matrix))
}
