% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prod_ind_names_colnames}
\alias{prod_ind_names_colnames}
\title{Names of data frame columns containing Product and Industry matrix row and column names}
\format{
A list with 2 entries.
\describe{
\item{product_names}{The name of a data frame column that contains Products. Default is "Product.names".}
\item{industry_names}{The name of a data frame column that contains Industries. Default is "Industry.names".}
}
}
\usage{
prod_ind_names_colnames
}
\description{
A list of data frame column names.
The data frame columns contain vectors of unique Product and Industry column names
from \strong{R}, \strong{U}, \strong{V}, and \strong{Y} matrices.
}
\keyword{datasets}
