% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{propulsion_eu.products}
\alias{propulsion_eu.products}
\title{A list of Propulsion useful work products}
\format{
A list with 3 entries.
\describe{
\item{MP}{The useful work product (Eu.product) Marine propulsion}
\item{RoP}{The useful work product (Eu.product) Road propulsion}
\item{RaP}{The useful work product (Eu.product) Rail propulsion}
}
}
\usage{
propulsion_eu.products
}
\description{
A list of Propulsion useful work products
}
\keyword{datasets}
