% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{read_ecc_from_excel}
\alias{read_ecc_from_excel}
\title{Read an energy conversion chain from a Excel file}
\usage{
read_ecc_from_excel(
  path,
  worksheets = NULL,
  R_region_name = Recca::psut_cols$R,
  U_region_name = Recca::psut_cols$U,
  V_region_name = Recca::psut_cols$V,
  Y_region_name = Recca::psut_cols$Y,
  r_eiou_region_name = Recca::psut_cols$r_eiou,
  U_eiou_region_name = Recca::psut_cols$U_eiou,
  U_feed_region_name = Recca::psut_cols$U_feed,
  S_units_region_name = Recca::psut_cols$S_units
)
}
\arguments{
\item{path}{The path to the Excel file.}

\item{R_region_name, U_region_name, V_region_name, Y_region_name, r_eiou_region_name, U_eiou_region_name, U_feed_region_name, S_units_region_name}{Named regions in the file at \code{path}.
See \code{Recca::psut_cols} for defaults.}
}
\value{
A data frame in \code{matsindf} format containing
the matrices from named regions in \code{path}.
}
\description{
Reads matrices from named regions in an Excel file
into \code{matsindf} format.
}
\details{
When \code{worksheets} is \code{NULL} (the default),
the named regions are assumed to be global.

Regions are assumed to include row and column names
in addition to numerical values.

This function is an inverse of \code{\link[=write_ecc_to_excel]{write_ecc_to_excel()}}.
}
\seealso{
\code{\link[=write_ecc_to_excel]{write_ecc_to_excel()}}
}
