% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregates.R
\name{region_aggregates}
\alias{region_aggregates}
\title{Aggregate by region}
\usage{
region_aggregates(
  .sut_data,
  aggregation_map,
  country = IEATools::iea_cols$country,
  year = IEATools::iea_cols$year,
  method = IEATools::iea_cols$method,
  energy_type = IEATools::iea_cols$energy_type,
  last_stage = IEATools::iea_cols$last_stage,
  s_units = Recca::psut_cols$s_units,
  U = Recca::psut_cols$U,
  r_EIOU = Recca::psut_cols$r_eiou,
  psut_colnames = Recca::psut_cols,
  matrix_names = Recca::psut_cols$matnames,
  matrix_values = Recca::psut_cols$matvals,
  continent = "Continent"
)
}
\arguments{
\item{.sut_data}{A wide-by-matrices \code{matsindf}-style data frame of PSUT matrices.}

\item{aggregation_map}{The recipe for aggregating to regions.}

\item{country}{The name of the country column in \code{.sut_data}. Default is \code{IEATools::iea_cols$country}.}

\item{continent}{The name of the continent column in the aggregated data frame.
Default is "Continent".}
}
\value{
A modified version of \code{.sut_data} wherein the \code{country} column is replaced
by region aggregates specified by \code{aggregation_map}.
}
\description{
Aggregates a data frame according to the regions given in an aggregation map.
}
\details{
An aggregation map is a named list of 3-letter country abbreviations.
The names are the regions into which the country data is to be aggregated.
}
