% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{replace_html_codes}
\alias{replace_html_codes}
\title{Unescape HTML codes in text}
\usage{
replace_html_codes(
  text,
  replacements = list(c("&amp;", "&"), c("&lt;", "<"), c("&gt;", ">"))
)
}
\arguments{
\item{text}{a vector (or one-dimensional list) of character strings}

\item{replacements}{a list of string pairs. Each pair consists of encoded string and unencoded string,
in that order. Default is
\code{list(c("&amp;", "&"), c("&lt;", "<"), c("&gt;", ">"))}}
}
\value{
If \code{text} is a vector, a vector of un-encoded strings.
If \code{text} is a list of strings, a list of un-encoded strings of same structure.
If possible, an outgoing list has simplified structure, even to the point of
conversion to vector.
}
\description{
Occasionally,
we need to un-escape HTML codes in text.
If \code{text} contains HTML codes, they are replaced with \code{replacements}, which,
by default, describe replacements for "\verb{&amp;}", "\verb{&lt;}", and "\verb{&gt;}"
("\code{&}", "\code{<}", and "\code{>}", respectively).
}
\details{
HTML codes can arrive in text read from an Excel file by the \code{openxlsx} package
due to a bug documented \href{https://github.com/awalker89/openxlsx/issues/393}{here}.
}
\examples{
replace_html_codes(list("a", "&amp;", "&lt;", "&gt;", "bcd"))
replace_html_codes(list(c("&amp;", "&amp;"), c("&lt;", "&lt;"), c("&gt;", "&gt;")))
}
