% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{row_col_types}
\alias{row_col_types}
\title{Names of row and column types in PSUT matrices}
\format{
A list with 2 entries.
\describe{
\item{product_type}{The name of the product row and column type. Default is "Product".}
\item{industry_type}{The name of the industry row and column type. Default is "Industry".}
}
}
\usage{
row_col_types
}
\description{
A list of row and column types.
}
\keyword{datasets}
