% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep_utilities.R
\name{S_units_from_tidy}
\alias{S_units_from_tidy}
\title{Extract an \code{S_units} matrix from a tidy data frame}
\usage{
S_units_from_tidy(.tidydf, Product = "Product", Unit = "Unit",
  S_units = "S_units")
}
\arguments{
\item{.tidydf}{the data frame from which an \code{S_units} matrix is to be formed}

\item{Product}{the name of the \code{Product} column in \code{.tidydf}. Default is "\code{Product}".}

\item{Unit}{the name of the \code{Unit} column in \code{.tidydf}. Default is "\code{Unit}".}

\item{S_units}{the name of the \code{S_units} column to be added to \code{.tidydf}.
Default is "\code{S_unit}".}
}
\value{
a data frame containing grouping variables and a new \code{S_unit} column
}
\description{
The \code{.tidydf} should be grouped as needed.
}
\examples{
library(dplyr)
S_units_from_tidy(UKEnergy2000tidy \%>\% group_by("Country", "Year", "Energy.type", "Last.stage"))
}
