% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sankey_cols}
\alias{sankey_cols}
\title{Sankey diagram data frame column names}
\format{
A string list with 1 entries.
\describe{
\item{sankey}{The name of a column in a wide-by-matrices data frame containing Sankey diagrams.}
}
}
\usage{
sankey_cols
}
\description{
A string list containing named names of columns in Sankey data frames.
Items in the list provide default values for column name function arguments
throughout the \code{Recca} package.
}
\examples{
sankey_cols
}
\keyword{datasets}
