% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{simplify_edge_list}
\alias{simplify_edge_list}
\title{Simplify an edge list}
\usage{
simplify_edge_list(edge_list, from = "From", to = "To", value = "Value",
  product = "Product")
}
\arguments{
\item{product}{}
}
\description{
An energy conversion chain edge list can be simplified if
there is only one source for a
}
