% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{test_against_file}
\alias{test_against_file}
\title{Execute \code{expect_known_value} interactively or during testing}
\usage{
test_against_file(actual_object, expected_file_name, update = FALSE,
  check.attributes = TRUE, expec_folder = file.path("tests",
  "expectations"))
}
\arguments{
\item{actual_object}{an object created during a test}

\item{expected_file_name}{the name of a file previously saved
against which \code{actual_object} will be tested.
This argument must be a string and should probably end in "\code{.rds}".
Do not include path information.
Example: "\code{expected_L.rds}".}

\item{update}{tells whether to update the saved object}

\item{check.attributes}{tells whether the test should also check attributes during the testing process}

\item{expec_folder}{a string giving the path to expected values stored as objects
on disk.
Default is "\code{tests/expectations}"}
}
\value{
NULL (invisibly). This function is called for its side effect of executing \code{expect_known_value}.
}
\description{
\code{testthat::expect_known_value} is difficult to use,
because working directories are different for interactive sessions and
during \code{R cmd check}.
This function abstracts those difficulties by using \code{testthat::is_testing()}
and other system checks.
The assumed directory for saved expected objects is "\code{tests/expectations}"
}
