% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UKEnergy2000mats}
\alias{UKEnergy2000mats}
\title{Energy consumption in the UK in 2000}
\format{
A data frame with 12 rows and 6 variables:
\describe{
\item{Country}{country, GB (Great Britain, only one country)}
\item{Year}{year, 2000 (only one year)}
\item{EnergyType}{E.ktoe (for energy) or X.ktoe (for exergy)}
\item{LastStage}{tells the final stage of the energy conversion chain: final, useful, or services}
\item{matrix.name}{gives the name of the matrix}
\item{matrix}{gives use (U), make (V), final demand (Y), r_EIOU, and S_units matrices}
}
}
\source{
\doi{10.1016/j.apenergy.2018.05.109}
}
\usage{
UKEnergy2000mats
}
\description{
A dataset containing approximations to
some of the energy flows in the UK in the year 2000.
These data first appeared as the example in
Figures 3, 4, 6, 7, 10, 11, B.1, and B.2 of
M.K. Heun, A. Owen, and P. E. Brockway.
A physical supply-use table framework for energy analysis on the energy conversion chain.
Applied Energy, 226:1134–1162, Sep 2018.
}
\details{
\code{UKEnergy2000mats} gives the use (\code{U}), make (\code{V}), and final demand (\code{Y}) matrices
associated with \code{UKEnergy2000tidy}.
These data are in \pkg{matsindf} format.
}
\keyword{datasets}
