% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{use_expected_data}
\alias{use_expected_data}
\title{Save an object for testing}
\usage{
use_expected_data(obj, dir = file.path("tests", "expectations"))
}
\arguments{
\item{obj}{an object to be used for testing}

\item{dir}{the directory in which to save expectation objects.
Default is file.path("tests", "expectations")}
}
\value{
This function is called for its side effect of saving data into the \code{tests/expectations} directory.
}
\description{
When developing tests,
it is convenient to perform an \code{R} calculation in a \code{test_that} function
and compare its result against an external calculation (from Excel or otherwise)
to verify correctness.
When correctness is verified,
you may want to
(a) capture the \code{R} object that represents the
result of the calculation and
(b) save it for later use with correctness tests in the \code{testthat} framework.
This function streamlines that process and works like
\code{\link[devtools]{use_data}}.
A call to \code{use_expected_data()} saves \code{obj} in a file named
"\code{expected_obj.rda}" file, where "\code{obj}" is replaced by the object's name in the environment.
Because the \code{testthat.R} file loads all previously-saved expectation objects,
they will be available for testing.
}
\examples{
a <- 5
use_expected_data(a)
}
