% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy_balance.R
\name{verify-balances}
\alias{verify-balances}
\alias{verify_inter_industry_balance}
\alias{verify_intra_industry_balance}
\title{Verify inter- and intra-industry balances for PSUT (\strong{RUVY}) matrices}
\usage{
verify_inter_industry_balance(
  .sutmats = NULL,
  balances = Recca::balance_cols$inter_industry_balance_colname,
  tol = 1e-06,
  balanced_colname = Recca::balance_cols$inter_industry_balanced_colname
)

verify_intra_industry_balance(
  .sutmats = NULL,
  balances = Recca::balance_cols$intra_industry_balance_colname,
  tol = 1e-06,
  balanced_colname = Recca::balance_cols$intra_industry_balanced_colname
)
}
\arguments{
\item{.sutmats}{An SUT-style data frame with a column
named \code{balances}.}

\item{balances}{The name of a column that contains balance vectors.
For \code{verify_inter_industry_balances()},
the default is \link{balance_cols}\verb{$inter_industry_balance_colname}
or "SUTInterIndustryBalance".
For \code{verify_intra_industry_balances()},
the default is \link{balance_cols}\verb{$intra_industry_balance_colname}
or "SUTIntraIndustryBalance".}

\item{tol}{The maximum amount by which products can be out of balance
and still be considered balanced.
Default is \code{1e-6}.}

\item{balanced_colname}{The name for booleans telling if balance is present.
For \code{verify_inter_industry_balances()},
the default is \link{balance_cols}\verb{$inter_industry_balanced_colname}
or "SUTInterIndustryBalanced".
For \code{verify_intra_industry_balances()},
the default is \link{balance_cols}\verb{$intra_industry_balanced_colname}
or "SUTIntraIndustryBalanced".}
}
\value{
A list or data frame with an additional value or column
saying whether \code{.sutmats} are in balance.
}
\description{
Balances can be verified in two ways,
inter-industry and intra-industry.
Inter-industry balances are between industries and performed by product.
Intra-industry balances are across industries and performed per industry.
}
\details{
\subsection{Inter-industry balances (by product)}{

In a PSUT description of an economy, all of every product
leaving one industry must arrive at another industry.
Inter-industry (between-industry) balances are verified by product
(within \code{tol}) for every row in \code{.sutmats}.
Inter-industry balances should be calculated via
\code{calc_inter_industry_balances()} before calling
\code{verify_inter_industry_balances()}.
See examples.
}

\subsection{Intra-industry balances (by industry)}{

Intra-industry (across-industry) balances are
verified by industry (within \code{tol}) for every row in \code{.sutmats}.
Intra-industry balances should be calculated via
\code{calc_intra_industry_balances()} before calling
\code{verify_intra_industry_balances()}.
See examples.

The calculation of inter-industry balances
will give non-zero vectors for
quantities that are not conserved (such as exergy)
and for conversion chains that do not include wastes or losses.
}

\subsection{Outputs}{

If every conversion chain in \code{.sutmats} is balanced,
\code{.sutmats} is returned with an additional column
(called \code{balanced_colname}), and
execution returns to the caller.
If balance is not observed for one or more of the rows
in \code{.sutmats},
a warning is emitted, and
the additional column (\code{balanced_colname})
indicates where the problem occurred, with \code{FALSE}
showing where products are not balanced
to within \code{tol}.

Typically, one would call \verb{calc_int*_industry_balance()}
before calling \verb{verify_int*_industry_balance()}.
See the examples.
}
}
\examples{
library(dplyr)
library(tidyr)
result_inter <- UKEnergy2000mats |>
  dplyr::filter(LastStage \%in\% c("Final", "Useful")) |>
  tidyr::pivot_wider(names_from = matrix.name,
                     values_from = matrix) |>
  calc_inter_industry_balance() |>
  verify_inter_industry_balance(tol = 1e-4)
result_inter
result_inter[[Recca::balance_cols$inter_industry_balanced_colname]]
}
\seealso{
\code{\link[=calc_inter_industry_balance]{calc_inter_industry_balance()}}, \code{\link[=calc_intra_industry_balance]{calc_intra_industry_balance()}}
}
