% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregates.R
\name{verify_footprint_aggregate_energy_balance}
\alias{verify_footprint_aggregate_energy_balance}
\title{Verify energy balance after footprint calculations}
\usage{
verify_footprint_aggregate_energy_balance(
  .sut_data = NULL,
  tol = 1e-04,
  R_mat,
  U_mat,
  U_feed_mat,
  V_mat,
  Y_mat,
  R_chop_list,
  U_chop_list,
  U_feed_chop_list,
  V_chop_list,
  Y_chop_list
)
}
\arguments{
\item{.sut_data}{An optional data frame of energy conversion chain matrices.}

\item{tol}{The tolerance within which energy balance is assumed to be OK. Default is \code{1e-4}.}

\item{R_mat, U_mat, U_feed_mat, V_mat, Y_mat}{The matrices of the original ECC.}

\item{R_chop_list, U_chop_list, U_feed_chop_list, V_chop_list, Y_chop_list}{Lists of matrices from different upstream swims corresponding to different rows or columns of \strong{Y}.}
}
\value{
\code{TRUE} if energy balance is observed, \code{FALSE} otherwise.
}
\description{
Footprint calculations involve
isolating rows or columns of the \strong{Y} matrix (chopping),
performing upstream swims (with \code{new_Y()}), and
creating the ECC portions that support the creation of the row or column of \strong{Y}.
After performing that upstream swim, the sum of the
isolated (chopped) ECCs should equal the original ECC.
This function performs that energy balance verification.
}
\details{
The various \verb{*_chop_list} arguments should be lists of matrices
formed by isolating (chopping) different parts of \strong{Y}.
The matrices in \code{R_chop_list}, \code{U_chop_list}, \code{U_feed_chop_list}
\code{U_eiou_chop_list}, \code{V_chop_list}, and \code{Y_chop_list} should sum to
\code{R}, \code{U}, \code{U_feed}, \code{U_eiou}, \code{V}, and \code{Y}, respectively.

This is not a public function.
It is an internal helper function
for \code{footprint_aggregates()}.
}
