% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy_balance.R
\name{verify_IEATable_energy_balance}
\alias{verify_IEATable_energy_balance}
\title{Confirm that an IEA-style data frame conserves energy.}
\usage{
verify_IEATable_energy_balance(.ieatidydata, ledger.side = "Ledger.side",
  energy = "E.ktoe", supply = "Supply", consumption = "Consumption",
  err = ".err", tol = 1e-06)
}
\arguments{
\item{.ieatidydata}{an IEA-style data frame containing the following columns:
\code{Country}, \code{Year}, \code{Ledger.side}, \code{Product}, \code{E.ktoe}.}

\item{ledger.side}{the name of the column in \code{.ieatidydata}
that contains ledger side information (a string). Default is "\code{Ledger.side}".}

\item{energy}{the name of the column in \code{.ieatidydata}
that contains energy data (a string). Default is "\code{E.ktoe}".}

\item{supply}{the identifier for supply data in the \code{ledger.side} column (a string).
Default is "\code{Supply}".}

\item{consumption}{the identifier for consumption data in the \code{ledger.side} column (a string).
Default is "\code{Consumption}".}

\item{tol}{the maximum amount by which Supply and Consumption can be out of balance}
}
\value{
Nothing is returned.
This function should be called
for its side-effect of testing whether energy is balanced in \code{.ieatidydata}.
}
\description{
Energy balances are confirmed (within \code{tol}) for every combination of
grouping variables in \code{.ieatidydata}.
}
\details{
Be sure to group \code{.ieatidydata} prior to calling this function,
as shown in the example.
}
\examples{
UKEnergy2000tidy \%>\%
  group_by(Country, Year, Energy.type, Last.stage) \%>\%
  verify_IEATable_energy_balance(energy = "EX.ktoe")
}
