% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy_balance.R
\name{verify_inter_industry_balance}
\alias{verify_inter_industry_balance}
\title{Confirm that an SUT-style data frame conserves energy between industries}
\usage{
verify_inter_industry_balance(
  .sutmats_with_balances = NULL,
  balances = Recca::balance_cols$inter_industry_balance_colname,
  tol = 1e-06,
  balanced_colname = Recca::balance_cols$inter_industry_balanced_colname
)
}
\arguments{
\item{tol}{The maximum amount by which Supply and Consumption can be out of balance.
Default is \code{1e-6}.}

\item{balanced_colname}{The name for booleans telling if balance is present.
Default is \code{Recca::balance_cols$inter_industry_balanced_colname}.}

\item{.sutmats}{an SUT-style data frame with columns of matrices,
including \code{R}, \code{U}, \code{V}, and \code{Y} columns.}

\item{R}{Resource (\strong{R}) matrix or name of the column in \code{.sutmats} that contains same. Default is "R".}

\item{U}{Use (\strong{U}) matrix or name of the column in \code{.sutmats} that contains same. Default is "U".}

\item{V}{Make (\strong{V}) matrix or name of the column in \code{.sutmats}that contains same. Default is "V".}

\item{Y}{Final demand (\strong{Y}) matrix or name of the column in \code{.sutmats} that contains same. Default is "Y".}
}
\value{
A list or data frame saying whether \code{.sutmats_with_balances} are in balance.
}
\description{
Inter-industry energy balances are verified by Product
(within \code{tol}) for every row in \code{.sutmats}.
}
\details{
In a PSUT description of an economy, all of every flow
leaving one industry must arrive at another industry.

If energy is in balance for every row, \code{.sutmats} is returned with an additional column, and
execution returns to the caller.
If energy balance is not observed for one or more of the rows,
a warning is emitted, and
the additional column (\code{SUT_energy_blance})
indicates where the problem occurred, with \code{FALSE}
showing where energy is not in balance.

Typically, one needs to call \code{\link[=calc_inter_industry_balance]{calc_inter_industry_balance()}}
before calling \code{\link[=verify_inter_industry_balance]{verify_inter_industry_balance()}}.
}
\examples{
library(dplyr)
library(tidyr)
verify_SUT_energy_balance(UKEnergy2000mats \%>\%
                            dplyr::filter(LastStage \%in\% c("Final", "Useful")) \%>\%
                            tidyr::spread(key = matrix.name, value = matrix),
                          tol = 1e-4)
}
