% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy_balance.R
\name{verify_SUT_energy_balance}
\alias{verify_SUT_energy_balance}
\title{Verifies SUT inter-industry energy balances}
\usage{
verify_SUT_energy_balance(
  .sutmats = NULL,
  R = "R",
  U = "U",
  V = "V",
  Y = "Y",
  tol = 1e-06,
  SUT_energy_balance = ".SUT_energy_balance"
)
}
\arguments{
\item{.sutmats}{an SUT-style data frame with columns of matrices,
including \code{U}, \code{V}, and \code{Y} columns.}

\item{R}{resources (\code{R}) matrix or name of the column in \code{.sutmats} that contains same. Default is "R".}

\item{U}{use (\code{U}) matrix or name of the column in \code{.sutmats} that contains same. Default is "U".}

\item{V}{make (\code{V}) matrix or name of the column in \code{.sutmats}that contains same. Default is "V".}

\item{Y}{final demand (\code{Y}) matrix or name of the column in \code{.sutmats} that contains same. Default is "Y".}

\item{tol}{the maximum amount by which Supply and Consumption can be out of balance.
Default is \code{1e-6}.}

\item{SUT_energy_balance}{The name of the column in output containing logical
values representing the balance status of that row.}
}
\value{
a list or data frame saying whether \code{.sutmats} are in balance.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\details{
\code{\link[=verify_SUT_energy_balance]{verify_SUT_energy_balance()}} is deprecated in favor of the combination of
\code{\link[=calc_inter_industry_balance]{calc_inter_industry_balance()}} and
\code{\link[=verify_inter_industry_balance]{verify_inter_industry_balance()}}.
Pipe the first into the second.

Energy balances are confirmed by Product (within \code{tol}) for every row in \code{.sutmats}.

If energy is in balance for every row, \code{.sutmats} is returned with an additional column, and
execution returns to the caller.
If energy balance is not observed for one or more of the rows,
a warning is emitted, and
the additional column (\code{SUT_energy_blance})
indicates where the problem occurred, with \code{FALSE}
showing where energy is not in balance.
}
\examples{
library(dplyr)
library(tidyr)
# This function is deprecated.
# Instead of this:
UKEnergy2000mats |>
  dplyr::filter(LastStage \%in\% c("Final", "Useful")) |>
  tidyr::spread(key = matrix.name, value = matrix) |>
  verify_SUT_energy_balance(tol = 1e-4)
# Do this:
UKEnergy2000mats |>
  dplyr::filter(LastStage \%in\% c("Final", "Useful")) |>
  tidyr::spread(key = matrix.name, value = matrix) |>
  calc_inter_industry_balance() |>
  verify_inter_industry_balance(tol = 1e-4)
}
