% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy_balance.R
\name{verify_SUT_industry_production}
\alias{verify_SUT_industry_production}
\title{Confirm that all Industries in an SUT-style data frame produce energy.}
\usage{
verify_SUT_industry_production(.sutdata = NULL, U_colname = "U",
  V_colname = "V", industry_production_OK = ".industry_production_OK")
}
\arguments{
\item{.sutdata}{an SUT-style data frame containing metadata columns
(typically \code{Country}, \code{Year}, \code{Ledger.side}, \code{Product}, etc.)
and columns of SUT matrices, including \code{U} and \code{V}.}

\item{U_colname}{the name of the column of use matrices}

\item{V_colname}{the name of the column of make matrices}

\item{industry_production_OK}{the name of the column in \code{.sutdata} that
tells whether all industries produce something. Default is "\code{.industry_production_OK}".}
}
\value{
\code{.sutdata} with added column named with the value of \code{industry_production_OK}.
}
\description{
If an Industry consumes energy (in the \strong{U} matrix)
but does not make energy (in the \strong{V} matrix),
it is most certainly an error.
(In contrast, there can be Industries that make energy but do not consume it,
such as Industries involved in Production.)
This function errors if an Industry is found that consumes energy but does not make energy.
But it will also place an object named \code{problems_data_frame_name}
into the global environment.
\code{problems_data_frame_name} can assist debugging this problem
First, look at the \code{industry_production_OK_colname} column of \code{problems_data_frame_name}.
It shows which rows of \code{.sutdata} had a problem (FALSE in that column).
If you find a row with a problem, look at the \code{check_colname} column.
Any \code{0} entries indicate industries that consume energy but do not make energy.
}
\examples{
verify_SUT_industry_production(UKEnergy2000mats)
}
