% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{waste_edges}
\alias{waste_edges}
\title{Create waste energy edges for an edge map}
\usage{
waste_edges(Umat, Vmat, from = "From", to = "To", value = "Value",
  product = "Product", waste = "Rejected energy")
}
\arguments{
\item{waste}{}
}
\value{

}
\description{
Create waste energy edges for an edge map
}
