% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{add_row_col_types}
\alias{add_row_col_types}
\title{Infer and set row and column types}
\usage{
add_row_col_types(
  .df = NULL,
  matvals = Recca::psut_cols$matvals,
  matnames = Recca::psut_cols$matnames,
  with_row_col_types_colname = "WithRCTypes",
  R = Recca::psut_cols$R,
  U = Recca::psut_cols$U,
  V = Recca::psut_cols$V,
  Y = Recca::psut_cols$Y,
  U_feed = Recca::psut_cols$U_feed,
  U_eiou = Recca::psut_cols$U_eiou,
  r_eiou = Recca::psut_cols$r_eiou,
  S_units = Recca::psut_cols$S_units,
  industry_type = Recca::row_col_types$industry_type,
  product_type = Recca::row_col_types$product_type,
  unit_type = Recca::row_col_types$unit_type
)
}
\arguments{
\item{matvals}{Matrices to have their row and column types set.}

\item{matnames}{Names for the matrices. Default is \code{names(matvals)}.}

\item{R, U, V, Y, U_feed, U_eiou, r_eiou, S_units}{Names of matrices.
Defaults from \code{Recca::psut_cols}.}

\item{industry_type, product_type, unit_type}{Names of row and column types.
Defaults from \code{Recca::row_col_types}.}
}
\value{
\code{.mats} with row and product types set.
}
\description{
Given a list of matrices and their names,
this function sets the row and column types
according to the following algorithm:
\itemize{
\item \strong{R} and \strong{V}: Rows are \code{industry_type}; Columns are \code{product_type}.
\item \strong{U}, \strong{Y}, \strong{U_feed}, \strong{U_EIOU}, \strong{r_EIOU}: Rows are \code{product_type}; Colums are \code{industry_type}.
\item \strong{S_units}: Rows are \code{product_type}; Columns are \code{unit_type}.
}
}
\details{
Note that matrix names are matched via \code{\link[=startsWith_any_of]{startsWith_any_of()}}.

This function is vectorized, so
\code{matvals} and \code{matnames} can be lists.
}
