% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balances.R
\name{balances}
\alias{balances}
\alias{calc_inter_industry_balance}
\alias{calc_intra_industry_balance}
\title{Calculate inter- and intra-industry balances for PSUT (\strong{RUVY}) matrices2}
\usage{
calc_inter_industry_balance(
  .sutmats = NULL,
  R = "R",
  U = "U",
  V = "V",
  Y = "Y",
  balance_colname = Recca::balance_cols$inter_industry_balance_colname
)

calc_intra_industry_balance(
  .sutmats = NULL,
  U = "U",
  V = "V",
  balance_colname = Recca::balance_cols$intra_industry_balance_colname
)
}
\arguments{
\item{.sutmats}{A named list of matrices or
an SUT-style, wide-by-matrices data frame
with columns of matrices,
including \code{R}, \code{U}, \code{V}, and \code{Y}.}

\item{R}{Resources (\strong{R}) matrix or name of the column in \code{.sutmats}
that contains same.
Default is "R".}

\item{U}{Use (\strong{U}) matrix or name of the column in \code{.sutmats}
that contains same.
Default is "U".}

\item{V}{Make (\strong{V}) matrix or name of the column in \code{.sutmats}
that contains same.
Default is "V".}

\item{Y}{Final demand (\strong{Y}) matrix or name of the column in \code{.sutmats}
that contains same.
Default is "Y".}

\item{balance_colname}{The name of the column containing energy balance vectors.
Defaults are
\link{balance_cols}\verb{$inter_industry_balance_colname} or
"SUTInterIndustryBalance"
for
\code{calc_inter_industry_balances()} and
\link{balance_cols}\verb{$intra_industry_balance_colname} or
"SUTIntraIndustryBalance"
for
\code{calc_intra_industry_balances()}.}
}
\value{
A list or data frame containing balances.
}
\description{
Balances are an important aspect of analyzing mass and energy conversion chains
in the PSUT framework with the \strong{RUVY} matrices.
Often, balances are calculated on energy or exergy flows.
Balances can be calculated for mass flows or monetary flows, too.
}
\details{
\subsection{Inter-industry balances}{

In a PSUT description of a mass or energy conversion chain,
all of every product leaving one industry must arrive at another industry.
\code{calc_inter_industry_balances()}
calculates these between-industry balances via
((\strong{R} + \strong{V})^T^ - (\strong{U} + \strong{Y}))\strong{i}.
Inter-industry balances are calculated for products (not industries).
}

\subsection{Intra-industry balances}{

\code{calc_intra_industry_balances()} calculates across-industry
balances via
(\strong{U}^T^ - \strong{V})\strong{i}.
Inter-industry balances are calculated for industries (not products),
and the result is a column vector of industry balances.
}

\subsection{Outputs}{

The argument \code{balance_colname} specifies the name of the
result.
By default, the output of \code{calc_inter_industry_balances()} is
\link{balance_cols}\verb{$inter_industry_balance_colname} or
"SUTInterIndustryBalance".
By default, the output of \code{calc_intra_industry_balances()} is
\link{balance_cols}\verb{$intra_industry_balance_colname} or
"SUTIntraIndustryBalance".

For inter-industry (between-industry) balances,
the result is a column vector of product balances.
The result should \strong{always} be the \strong{0} vector.

In a PSUT description of a mass or energy conversion chain,
the meaning of intra-industry (across-industry) balances
depends on the construction of the matrices.
When all losses are accounted in the matrices themselves,
the calculation of intra-industry balances for
mass and energy conversion chains should give the \strong{0} vector
with industries in rows,
because mass and energy are conserved.
When losses are \strong{not} accounted in the matrices,
the calculation of intra-industry balances gives losses.

For exergy conversion chains
(mass, energy, or both),
when losses are accounted in the matrices,
the intra-industry balance gives exergy destruction.
When losses are \strong{not} accounted in the matrices,
the intra-industry balance gives the sum of
exergy destruction and exergy losses.
}
}
\examples{
# Inter-industry balances
inter <- UKEnergy2000mats |>
  dplyr::filter(LastStage \%in\% c("Final", "Useful")) |>
  tidyr::pivot_wider(names_from = matrix.name, values_from = matrix) |>
  calc_inter_industry_balance()
inter[[Recca::balance_cols$inter_industry_balance_colname]][[1]]
inter[[Recca::balance_cols$inter_industry_balance_colname]][[2]]
# Intra-industry balances
intra <- UKEnergy2000mats |>
  dplyr::filter(LastStage \%in\% c("Final", "Useful")) |>
  tidyr::pivot_wider(names_from = matrix.name, values_from = matrix) |>
  calc_intra_industry_balance()
intra[[Recca::balance_cols$intra_industry_balance_colname]][[1]]
intra[[Recca::balance_cols$intra_industry_balance_colname]][[2]]
}
\seealso{
\code{\link[=verify_inter_industry_balance]{verify_inter_industry_balance()}} and \code{\link[=verify_intra_industry_balance]{verify_intra_industry_balance()}}
}
