% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{calc_A}
\alias{calc_A}
\title{Calculate \code{Z}, \code{D}, \code{C}, and \code{A} matrices}
\usage{
calc_A(.sutdata = NULL, U = "U", V = "V", q = "q", f = "f",
  g = "g", Z = "Z", K = "K", C = "C", D = "D", A = "A")
}
\arguments{
\item{.sutdata}{a data frame of supply-use table matrices with matrices arranged in columns.}

\item{U}{use (\code{U}) matrix or name of the column in \code{.sutmats} that contains same. Default is "\code{U}".}

\item{V}{make (\code{V}) matrix or name of the column in \code{.sutmats}that contains same. Default is "\code{V}".}

\item{q}{\code{q} vector or name of the column in \code{.sutmats} that contains same. Default is "\code{q}".}

\item{f}{\code{f} vector or name of the column in \code{.sutmats} that contains same. Default is "\code{r}".}

\item{g}{\code{g} vector or name of the column in \code{.sutmats} that contains same. Default is "\code{g}".}

\item{Z}{\code{Z} matrix or name of the column in \code{.sutmats} that contains same. Default is "\code{Z}".
\code{W} is calculated by \code{U * g_hat_inv}.}

\item{K}{\code{K} matrix or name of the column in \code{.sutmats} that contains same. Default is "\code{K}".
\code{K} is calculated by \code{U * f_hat_inv}.}

\item{C}{\code{C} matrix or name of the column in \code{.sutmats} that contains same. Default is "\code{C}".
\code{C} is calculated by \code{transpose(V) * g_hat_inv}.}

\item{D}{\code{D} matrix or name of the column in \code{.sutmats} that contains same. Default is "\code{D}".
\code{D} is calculated by \code{V * q_hat_inv}.}

\item{A}{\code{A} matrix or name of the column in \code{.sutmats} that contains same. Default is "\code{A}".
\code{A} is calculated by \code{Z * D}.}
}
\value{
a list or data frame containing \code{Z},
\code{K}, \code{C}, \code{D}, and \code{A} matrices.
}
\description{
Calculate \code{Z}, \code{D}, \code{C}, and \code{A} matrices
}
