% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{calc_A}
\alias{calc_A}
\title{Calculate \strong{Z}, \strong{K}, \strong{D}, \strong{C}, \strong{A}, and \strong{O} matrices}
\usage{
calc_A(
  .sutdata = NULL,
  R = "R",
  U = "U",
  V = "V",
  q = "q",
  f = "f",
  g = "g",
  r = "r",
  h = "h",
  Z = "Z",
  K = "K",
  C = "C",
  D = "D",
  A = "A",
  O = "O"
)
}
\arguments{
\item{.sutdata}{a data frame of supply-use table matrices with matrices arranged in columns.}

\item{R}{resources (\strong{R}) matrix or name of the column in \code{.sutmats} that contains same. Default is "R".
\code{R} is an optional argument.
If all of \strong{R} is added to \strong{V}, this argument can be left unspecified.}

\item{U}{Use (\strong{U}) matrix or name of the column in \code{.sutmats} that contains same. Default is "U".}

\item{V}{Make (\strong{V}) matrix or name of the column in \code{.sutmats}that contains same. Default is "V".}

\item{q}{A \strong{q} vector or name of the column in \code{.sutmats} that contains same. Default is "q".}

\item{f}{An \strong{f} vector or name of the column in \code{.sutmats} that contains same. Default is "r".}

\item{g}{A \strong{g} vector or name of the column in \code{.sutmats} that contains same. Default is "g".}

\item{r}{An \strong{r} vector or name of the column in \code{.sutmats} that contains same. Default is "r".}

\item{h}{An \strong{h} vector or name of the column in \code{.sutmats} that contains same. Default is "h".}

\item{Z}{The name for the \strong{Z} matrix on output. Default is "Z".
\strong{Z} is calculated by \code{U * g_hat_inv}.}

\item{K}{The name for the \strong{K} matrix on output. Default is "K".
\strong{K} is calculated by \code{U * f_hat_inv}.}

\item{C}{The name for the \strong{C} matrix on output. Default is "C".
\strong{C} is calculated by \code{transpose(V) * g_hat_inv}.}

\item{D}{The name for the \strong{D} matrix on output. Default is "D".
\strong{D} is calculated by \code{V * q_hat_inv}.}

\item{A}{The name for the \strong{A} matrix on output. Default is "A".
\strong{A} is calculated by Z * D`.}

\item{O}{The name for the \strong{O} matrix on output. Default is "O".
\strong{O} is calculated by \code{r_hat_inv * R}.}
}
\value{
A list or data frame containing \strong{Z},
\strong{K}, \strong{C}, \strong{D}, and \strong{A} matrices.
}
\description{
Calculate \strong{Z}, \strong{K}, \strong{D}, \strong{C}, \strong{A}, and \strong{O} matrices
}
