% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chop.R
\name{calc_aggregates_from_ecc_prime}
\alias{calc_aggregates_from_ecc_prime}
\title{Calculate aggregates from list of reconstructed ECCs}
\usage{
calc_aggregates_from_ecc_prime(
  ecc_prime,
  calc_pfd_aggs,
  p_industries,
  fd_sectors,
  pattern_type,
  product_sector,
  aggregates_df,
  aggregate_primary,
  gross_aggregate_demand,
  net_aggregate_demand,
  R_prime_colname,
  U_prime_colname,
  U_feed_prime_colname,
  U_eiou_prime_colname,
  r_eiou_prime_colname,
  V_prime_colname,
  Y_prime_colname
)
}
\arguments{
\item{ecc_prime}{A list of reconstructed energy conversion chains.}

\item{calc_pfd_aggs}{Tells whether to calculate and add primary and final demand aggregates
to the nested data frame.}

\item{p_industries}{A vector of names of industries to be aggregated as "primary."
See \code{footprint_aggregates()} for details.}

\item{fd_sectors}{A vector of names of sectors in final demand.
See \code{footprint_aggregates()} for details.}

\item{pattern_type}{One of "exact", "leading", "trailing", or "anywhere" which specifies
how matches are made for \code{p_industries}.
See \code{footprint_aggregates()} for details.}

\item{product_sector}{The name of the output column that contains the product, industry, or sector
for which footprint aggregates are given.}

\item{aggregates_df, aggregate_primary, net_aggregate_demand, gross_aggregate_demand}{Names of output columns.
See \code{Recca::aggregate_cols}.}

\item{R_prime_colname, U_prime_colname, U_feed_prime_colname, U_eiou_prime_colname, r_eiou_prime_colname, V_prime_colname, Y_prime_colname}{Names of output matrices in the return value.
Default values are constructed from}
}
\value{
A data frame containing reconstructed (prime) matrices and
primary and final demand aggregates in a list suitable for use in \code{matsindf::matsindf_apply()}.
}
\description{
This is a helper function for \code{footprint_aggregates()} and \code{effects_aggregates()}.
It calculates the primary and final demand aggregates for a list of
reconstructed energy conversion chains (ECCs) in \code{ecc_prime}.
}
\details{
This is not a public function.
It is an internal helper function
for \code{footprint_aggregates()} and \code{effects_aggregates()}.
}
