% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embodied.R
\name{calc_E}
\alias{calc_E}
\title{Calculate the \code{E} matrix for embodied energy calculations}
\usage{
calc_E(
  .iomats = NULL,
  R = "R",
  V = "V",
  U_feed = "U_feed",
  g = "g",
  r = "r",
  E = "E"
)
}
\arguments{
\item{.iomats}{a data frame containing matrices that describe the Input-Output structure of an Energy Conversion Chain.
\code{.iomats} will likely have been obtained from the \code{\link{calc_io_mats}} function.}

\item{R}{The name of the \code{R} matrix column.
Default is "R".}

\item{V}{The name of the \code{V} matrix column.
Default is "V".}

\item{U_feed}{The name of the \code{U_feed} matrix column.
Default is "U_feed".}

\item{g}{The name of the \code{g} vector in \code{.iomats}. Default is "g".}

\item{r}{The name of the \code{r} vector in the \code{iomats}. Default is "r".}

\item{E}{the name for the \code{E} matrix on output. Default is "E".
\code{E} is calculated by \code{W * g_hat_inv}.}
}
\value{
list or data frame containing \code{E} matrices
}
\description{
Calculate the \code{E} matrix for embodied energy calculations
}
