% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embodied.R
\name{calc_E}
\alias{calc_E}
\title{Calculate the \code{E} matrix for embodied energy calculations}
\usage{
calc_E(
  .iomats = NULL,
  R = "R",
  V = "V",
  U_feed = "U_feed",
  g = "g",
  r = "r",
  E = "E"
)
}
\arguments{
\item{.iomats}{a data frame containing matrices that describe the Input-Output structure of an Energy Conversion Chain.
\code{.iomats} will likely have been obtained from the \code{\link{calc_io_mats}} function.}

\item{R}{A resources (\strong{R}) matrix or name of column in \code{.iomats} containing same. Default is "R".}

\item{V}{A make (\strong{V}) matrix or name of column in \code{.iomats} containing same. Default is "V".}

\item{U_feed}{A feedstock use (\strong{U_feed}) matrix or name of column in \code{.iomats} containing same. Default is "U_feed".}

\item{g}{A \strong{g} vector or name of column in \code{.iomats} containing same. Default is "g".}

\item{r}{An \strong{r} vector or name of column in \code{.iomats} containing same. Default is "r".}

\item{E}{The name for the \strong{E} matrix on output. \strong{E} is calculated by \code{W * g_hat_inv}. Default is "E".}
}
\value{
A list or data frame containing \code{E} matrices.
}
\description{
Calculate the \code{E} matrix for embodied energy calculations
}
