% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embodied.R
\name{calc_E}
\alias{calc_E}
\title{Calculate the E matrix for embodied energy calculations}
\usage{
calc_E(.iodata = NULL, g_colname = "g", W_colname = "W",
  U_EIOU_colname = "U_EIOU", E_colname = "E")
}
\arguments{
\item{.iodata}{a data frame containing matrices that describe the Input-Output structure of an Energy Conversion Chain.
\code{.iodata} will likely have been obtained from the \code{calc_io_mats} function.}

\item{g_colname}{the name of the column in \code{.iodata} containing final demand (\code{g}) vectors}

\item{W_colname}{the name of the column in \code{.iodata} containing Product-by-Industry
value added (\code{W}) matrices}

\item{U_EIOU_colname}{the name of the column in \code{.iodata} containing energy industry own use matrices}

\item{E_colname}{the name of the output column containing \code{E} matrices.
\code{E} is calculated by \code{W * g_hat_inv}.}
}
\value{
\code{.iodata} with column \code{E_colname} added
}
\description{
Calculate the E matrix for embodied energy calculations
}
