% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eroi.R
\name{calc_E_EIOU}
\alias{calc_E_EIOU}
\title{Calculate various embodied energy matrices}
\usage{
calc_E_EIOU(
  .iomats = NULL,
  g = "g",
  U_EIOU = "U_EIOU",
  E_EIOU = "E_EIOU",
  e_EIOU = "e_EIOU"
)
}
\arguments{
\item{.iomats}{a data frame containing matrices that describe the Input-Output structure
(using the supply-use table format) of an Energy Conversion Chain.
\code{.iomats} will likely have been obtained from the \code{calc_io_mats()} function.}

\item{g}{name of the \code{g} vector on output. Default is "\code{g}".}

\item{U_EIOU}{name of the \code{U_EIOU} matrices on output. Default is "\code{U_EIOU}".}

\item{E_EIOU}{name for the \code{E_EIOU} matrix on output. Default is "\code{E_EIOU}".
\code{E} is calculated by \code{U_EIOU * g_hat_inv}.}

\item{e_EIOU}{name for the \code{e_EIOU} vector on output. Default is "\code{e_EIOU}".
\code{e_EIOU} is calculated by \code{transpose(i) * e_EIOU}.}
}
\value{
list or data frame containing the \code{E_EIOU} matrix and \code{e_EIOU} vector.
}
\description{
\subsection{Change title and add description.}{
}
}
