% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embodied.R
\name{calc_embodied_EIOU}
\alias{calc_embodied_EIOU}
\title{Calculate various embodied EIOU matrices}
\usage{
calc_embodied_EIOU(
  .iomats = NULL,
  e_EIOU = "e_EIOU",
  Y = "Y",
  y = "y",
  L_ixp = "L_ixp",
  L_ixp_feed = "L_ixp_feed",
  Q_EIOU_s = "Q_EIOU_s",
  Q_EIOU_p = "Q_EIOU_p",
  Q_EIOU_feed_s = "Q_EIOU_feed_s",
  Q_EIOU_feed_p = "Q_EIOU_feed_p"
)
}
\arguments{
\item{.iomats}{A wide-by-matrices data frame containing matrices that describe the Input-Output structure
(using the supply-use table format) of an Energy Conversion Chain.
\code{.iomats} will likely have been obtained combining the \code{calc_io_mats()} and \code{calc_E_EIOU()} functions.
See the example.?c}

\item{e_EIOU}{The name of the column containing the direct energy use extension vector.
Default is "e_EIOU".}

\item{Y}{The name of the column containing final demand \code{Y} matrices in \code{.iomats}.
Default is "Y".}

\item{y}{The name of the column containing total final demand by product \code{y} vectors in \code{iomats}.
Default is "y".}

\item{L_ixp}{The name of the column containing \code{L_ixp} matrices in \code{.iomats}.
Default is "L_ixp".}

\item{L_ixp_feed}{The name of the column containing \code{L_ixp_feed} matrix in \code{.iomats}.
Default is "L_ixp_feed".}

\item{Q_EIOU_s}{The name of the output column containing the EIOU embodied by final demand sectors, including both energy use for feedstock and EIOU production.
Default is "Q_EIOU_s".}

\item{Q_EIOU_p}{The name of the output column containing the EIOU embodied by final demand products, including both energy use for feedstock and EIOU production.
Default is "Q_EIOU_p".}

\item{Q_EIOU_feed_s}{The name of the output column containing the EIOU embodied by final demand sectors, including only energy use for feedstock production.
Default is "Q_EIOU_feed_s".}

\item{Q_EIOU_feed_p}{The name of the output column containing the EIOU embodied by final demand products, including only energy use for feedstock production.
Default is "Q_EIOU_feed_p".}
}
\value{
A data frame that contains several embodied EIOU matrices in added columns.
See description for details.
}
\description{
This function calculates different embodied Energy Industry Own Use (EIOU) matrices (see details) for a given energy conversion chain
and final demand, using a data frame of input-output matrices in the physical supply-use table (PSUT) format.
}
\details{
The argument \code{.iomats} should be a wide-by-matrices data frame, obtained combining the \code{calc_iomats()} and \code{calc_E_EIOU()}
functions as described in the example.

This function adds many additional columns to \code{.iomats}, each one containing particular embodied EIOU matrices.

The embodied EIOU matrices are calculated either:
\itemize{
\item by final demand sector (subscript "_s" appears in the name);
\item by final demand products (subscript "_p" appears in the name);
\item including only EIOU required for feedstock inputs production (subscript "_feed" appears in the name);
\item including both EIOU required for feedstock and EIOU inputs production (no additional subscript).
}

Note: All matrix multiplication (\code{\%*\%}) is performed "by name" using
\code{matsbyname::matrixproduct_byname()}.

Output columns include:
\itemize{
\item \code{Q_EIOU_s}: matrix of embodied EIOU by final demand sectors, including both energy use for feedstock and EIOU production.
\code{Q_EIOU_s} is calculated by \code{e_EIOU_hat \%*\% L_ixp \%*\% Y}.
\item \code{Q_EIOU_p}: matrix of embodied EIOU by final demand products, including both energy use for feedstock and EIOU production.
\code{Q_EIOU_p} is calculated by \code{e_EIOU_hat \%*\% L_ixp \%*\% y_hat}.
\item \code{Q_EIOU_feed_s}: matrix of embodied EIOU by final demand sectors, including only energy use for feedstock production.
\code{Q_EIOU_feed_s} is calculated by \code{e_EIOU_hat \%*\% L_ixp_feed \%*\% Y}.
\item \code{Q_EIOU_feed_p}: matrix of embodied EIOU by final demand products, including only energy use for feedstock production.
\code{Q_EIOU_feed_p} is calculated by \code{e_EIOU_hat \%*\% L_ixp_feed \%*\% y_hat}.
}
}
\examples{
library(IEATools)
UKEnergy2000mats \%>\%
  tidyr::pivot_wider(names_from = "matrix.name", values_from = "matrix") \%>\%
  calc_io_mats() \%>\%
  calc_E_EIOU() \%>\%
  calc_embodied_EIOU()
}
