% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embodied.R
\name{calc_embodied_etas}
\alias{calc_embodied_etas}
\title{Embodied energy efficiencies}
\usage{
calc_embodied_etas(
  .embodiedmats = NULL,
  primary_machine_names,
  Y = "Y",
  G = "G",
  H = "H",
  eta_p = "eta_p",
  eta_s = "eta_s"
)
}
\arguments{
\item{.embodiedmats}{a data frame containing columns of \code{Y}, \code{G}, and \code{H} matrices}

\item{primary_machine_names}{a list of strings representing names of Industries whose output is counted in Total Primary Energy Supply (TPES)}

\item{Y}{final demand \code{Y} matrix or name of a column in \code{.embodiedmats} containing same. Default is "\code{Y}".}

\item{G}{\code{G} matrix or name of a column in \code{.embodiedmats} containing same. Default is "\code{G}".}

\item{H}{\code{H} matrix or name of a column in \code{.embodiedmats} containing same. Default is "\code{H}".}

\item{eta_p}{the name for product-based efficiencies on output. Default is "\code{eta_p}".}

\item{eta_s}{the name for final-demand-sector-based efficiencies on output. Default is "\code{eta_s}".}
}
\value{
a list or data frame containing embodied energy efficiencies
}
\description{
Embodied energy efficiencies are based on the total upstream primary energy demand
for a product produced by the ECC or
for the energy consumed by a final demand sector of the ECC.
This function calculates both.
\code{eta_s} gives sector-based embodied energy efficiency, and
\code{eta_p} gives product-based embodied energy efficiency.
}
\details{
Note that these efficiencies (\code{eta_s} and \code{eta_p}) are different from
energy conversion industry efficiencies.
To calculate energy conversion industry efficiencies, use the
\code{\link{calc_eta_i}} function.
}
