% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embodied.R
\name{calc_embodied_etas}
\alias{calc_embodied_etas}
\title{Embodied energy efficiencies}
\usage{
calc_embodied_etas(.embodiedmats = NULL, primary_machine_names,
  Y_colname = "Y", G_colname = "G", H_colname = "H",
  eta_p_colname = "eta_p", eta_s_colname = "eta_s")
}
\arguments{
\item{.embodiedmats}{a data frame containing columns of \strong{Y}, \strong{G}, and \strong{H} matrices}

\item{primary_machine_names}{a list of strings representing names of Industries whose output is counted in TPES}

\item{Y_colname}{a string for the name of a column of Y matrices in \code{.embodiedmats} (default is \code{Y})}

\item{G_colname}{a string for the name of a column of G matrices in \code{.embodiedmats} (default is \code{G})}

\item{H_colname}{a string for the name of a column of H matrices in \code{.embodiedmats} (default is \code{H})}

\item{eta_p_colname}{a string for the name of the output column containing vectors of product-based efficiencies}

\item{eta_s_colname}{a string for the name of the output column containing vectors of final-demand-sector-based efficiencies}
}
\value{
\code{.embodiedmats} with columns \code{eta_p_colname} and \code{eta_s_colname} added
}
\description{
Embodied energy efficiencies
}
