% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embodied.R
\name{calc_embodied_mats}
\alias{calc_embodied_mats}
\title{Calculate various embodied energy matrices}
\usage{
calc_embodied_mats(
  .iomats = NULL,
  q = "q",
  g = "g",
  r = "r",
  L_ixp = "L_ixp",
  A = "A",
  R = "R",
  V = "V",
  U_feed = "U_feed",
  Y = "Y",
  G = "G",
  H = "H",
  E = "E",
  M_p = "M_p",
  M_s = "M_s",
  F_footprint_p = "F_footprint_p",
  F_effects_p = "F_effects_p",
  F_footprint_s = "F_footprint_s",
  F_effects_s = "F_effects_s"
)
}
\arguments{
\item{.iomats}{a data frame containing matrices that describe the Input-Output structure
(using the supply-use table format)
of an Energy Conversion Chain.
\code{.iomats} will likely have been obtained from the \code{calc_io_mats()} function.}

\item{q}{Final demand (\strong{q}) vector or name of the column in \code{.iomats} containing same. Default is "q".}

\item{g}{A \strong{g} vector or name of the column in \code{.iomats} containing same. Default is "g".}

\item{r}{An \strong{r} vector or name of the column in \code{.iomats} containing same. Default is "r".}

\item{L_ixp}{Industry-by-product Leontief (\strong{L_ixp}) matrix or name of the column in \code{.iomats} containing same. Default is "L_ixp".}

\item{A}{An \strong{A} matrix or name of the column in \code{.iomats} containing same. Default is "A".}

\item{R}{A resources (\strong{R}) matrix or name of the column in \code{.iomats} containing same. Default is "R".}

\item{V}{A make (\strong{V}) matrix or name of the column in \code{.iomats} containing same. Default is "V".}

\item{U_feed}{A feedstock use (\strong{U_feed}) matrix or name of the column in \code{.iomats} containing same. Default is "U_feed".}

\item{Y}{A final demand (\strong{Y}) matrix or name of the column in \code{.iomats} containing same. Default is "Y".}

\item{G}{The name of the \strong{G} matrix on output. \strong{G} is calculated by \code{L_ixp * y_hat}. Default is "G".}

\item{H}{The name of the \strong{H} matrix on output. \strong{H} is calculated by \code{L_ixp * Y}. Default is "H".}

\item{E}{The name of the \strong{E} matrix on output. \strong{E} is calculated by \code{W * g_hat_inv}. Default is "E".}

\item{M_p}{The name of the \strong{M_p} matrix on output. \strong{M_p} is formed from column sums of positive entries in the various \strong{Q}_x matrices. Default is "M_p".}

\item{M_s}{The name of the \strong{M_s} matrix on output. \strong{M_s} is calculated by \code{M_p * q_hat_inv * Y}. Default is "M_s".}

\item{F_footprint_p}{The name of the \strong{F_footprint_p} matrix on output. \strong{F_footprint_p} is calculated by \verb{M_p * (M_p^T * i)_hat_inv}. Default is "F_footprint_p".}

\item{F_effects_p}{The name of the \strong{F_effects_p} matrix on output. \strong{F_effects_p} is calculated by \verb{(M_p * i)_hat_inv * M_p}. Default is "F_effects_p".}

\item{F_footprint_s}{The name of the \strong{F_footprint_s} matrix on output. \strong{F_footprint_s} is calculated by \verb{M_s * (M_s^T *i)_hat_inv}. Default is "F_footprint_s".}

\item{F_effects_s}{The name of the \strong{F_effects_s} matrix on output. \strong{F_effects_s} is calculated by \verb{(M_s * i)_hat_inv * M_s}. Default is "F_effects_s".}
}
\value{
A list or data frame containing embodied energy matrices.
}
\description{
Calculate various embodied energy matrices
}
