% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eroi.R
\name{calc_erois}
\alias{calc_erois}
\title{Calculate EROIs}
\usage{
calc_erois(
  .iomats = NULL,
  e_EIOU = "e_EIOU",
  L_ixp = "L_ixp",
  L_ixp_feed = "L_ixp_feed",
  D = "D",
  C = "C",
  eroi_g_p = "eroi_g_p",
  eroi_g_i = "eroi_g_i",
  eroi_g_p_feed = "eroi_g_p_feed",
  eroi_g_i_feed = "eroi_g_i_feed"
)
}
\arguments{
\item{.iomats}{A wide-by-matrices data frame containing matrices that describe the Input-Output structure
(using the supply-use table format) of an Energy Conversion Chain.
\code{.iomats} will likely have been obtained combining the \code{calc_io_mats()} and \code{calc_E_EIOU()} functions.
See the example.}

\item{e_EIOU}{The name of the column containing \code{e_EIOU} vectors in \code{.iomats}. Default is "e_EIOU".}

\item{L_ixp}{The name of the column containing \code{L_ixp} matrices in \code{.iomats}. Default is "L_ixp".}

\item{L_ixp_feed}{The name of the column containing \code{L_ixp_feed} matrix in \code{.iomats}. Default is "L_ixp_feed".}

\item{D}{The name of the column containing \code{D} matrices in \code{.iomats}. Default is "D".}

\item{C}{The name of the column containing \code{C} matrix in \code{.iomats}. Default is "C".}

\item{eroi_g_p}{The name of the output column containing vectors of product-level gross EROIs, including both energy use for feedstock and EIOU production.
Default is "eroi_g_p".}

\item{eroi_g_i}{The name of the output column containing vectors of industry-level gross EROIs,
including both energy use for feedstock and EIOU production.
Default is "eroi_g_i".}

\item{eroi_g_p_feed}{The name of the output column containing vectors of product-level gross EROIs,
including only energy use for feedstock production.
Default is "eroi_g_p_feed".}

\item{eroi_g_i_feed}{The name of the output column containing vectors of industry-level gross EROIs,
including only energy use for feedstock production.
Default is "eroi_g_i_feed".}
}
\value{
A data frame that includes several additional EROIs vectors in added columns.
See description for details.
}
\description{
This function calculates energy return on investment (EROI) values
given a data frame of input-output matrices
in the physical supply-use table (PSUT) format for an energy conversion chain.
}
\details{
The argument \code{.iomats} should be a wide-by-matrices data frame.

Other input columns are named by the matrices they contain.

This function adds many additional columns to \code{.iomats}, each one containing a particular type of EROI.
The default column names use the following naming convention:
\itemize{
\item names of EROIs calculated for products include the string "_p";
\item names of EROIs calculated for industries include the string "_i";
\item names of gross EROIs include the string "_g";
\item names of net EROIs include the string "_n".
}

In addition, calculations are made based on inclusion of either
\itemize{
\item only EIOU required for feedstock inputs production (in which case "_feed" appears in the name);
\item both EIOU required for feedstock and EIOU inputs production (no additional string in the name).
}

Output columns include:
\itemize{
\item \code{eroi_g_p}: vector of product-level gross EROIs, including both EIOU required for feedstock and EIOU inputs production.
The inverse of \code{eroi_g_p} is calculated by \code{transpose(i) \%*\% e_EIOU_hat \%*\% L_ixp}.
\item \code{eroi_g_i}: vector of industry-level gross EROIs, including both EIOU required for feedstock and EIOU inputs production.
The inverse of \code{eroi_i_p} is calculated by \code{transpose(C) * eroi_g_p_inv}.
\item \code{eroi_g_p_feed}: vector of product-level gross EROIs, including only EIOU required for feedstock inputs production.
The inverse of \code{eroi_g_p_feed} is calculated by \code{transpose(i) \%*\% e_EIOU_hat \%*\% L_ixp_feed}.
\item \code{eroi_g_i_feed}: vector of industry-level gross EROIs, including only EIOU required for feedstock inputs production.
The inverse of \code{eroi_i_p} is calculated by \code{transpose(C) \%*\% eroi_g_p_feed_inv}.
}

Note: All matrix multiplication (\code{\%*\%}) is performed "by name" using
\code{matsbyname::matrixproduct_byname()}.
}
\examples{
library(IEATools)
UKEnergy2000mats \%>\%
  dplyr::filter(LastStage == "Final", EnergyType == "E") \%>\%
  tidyr::pivot_wider(names_from = "matrix.name", values_from = "matrix") \%>\%
  calc_io_mats() \%>\%
  calc_E_EIOU() \%>\%
  calc_erois()
}
