% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency.R
\name{calc_eta_fu}
\alias{calc_eta_fu}
\title{Calculate final-to-useful efficiencies}
\usage{
calc_eta_fu(
  .c_mats_eta_phi_vecs = NULL,
  C_Y = Recca::alloc_cols$C_Y,
  C_eiou = Recca::alloc_cols$C_eiou,
  eta_i = Recca::efficiency_cols$eta_i,
  phi = Recca::psut_cols$phi,
  energy_type = Recca::energy_types$energy_type,
  eta_fu = Recca::efficiency_cols$eta_fu,
  energy = Recca::energy_types$e,
  exergy = Recca::energy_types$x,
  eta_fu_Y_e = paste0(eta_fu, "_Y_", energy),
  eta_fu_Y_x = paste0(eta_fu, "_Y_", exergy),
  eta_fu_eiou_e = paste0(eta_fu, "_EIOU_", energy),
  eta_fu_eiou_x = paste0(eta_fu, "_EIOU_", exergy),
  notation = RCLabels::arrow_notation
)
}
\arguments{
\item{.c_mats_eta_phi_vecs}{A data frame containing allocation matrices (\code{C_Y} and \code{C_eiou}),
vectors of machine efficiencies (\code{eta_i}), and
exergy-to-energy ratio vectors (\code{phi}).
Default is \code{NULL}, in which case individual matrices or vectors
can be passed to \code{C_Y}, \code{C_eiou}, \code{eta_i}, and \code{phi}.}

\item{C_Y}{The name of the column in \code{.c_mats_eta_phi_vecs} containing allocation
matrices for final demand (the \code{Y} in \code{C_Y}).
Or a single \strong{C_Y} matrix.
Or a list of \strong{C_Y} matrices.
Default is \code{Recca::alloc_cols$C_Y}.}

\item{C_eiou}{The name of the column in \code{.c_mats_eta_phi_vecs} containing allocation
matrices for energy industry own use (the \code{eiou} in \code{C_eiou}).
Or a single \strong{C_EIOU} matrix.
Or a list of \strong{C_EIOU} matrices.
Default is \code{Recca::alloc_cols$C_eiou}.}

\item{eta_i}{The name of the column in \code{.c_mats_eta_phi_vecs} containing machine efficiencies.
Or a single \strong{eta_i} vector.
Or a list of \strong{eta_i} vectors.
Default is \code{Recca::efficiency_cols$eta_i}.}

\item{phi}{The name of the column in \code{.c_mats_eta_phi_vecs} containing exergy-to-energy ratios.
Or a single \strong{phi} vector.
Default is \code{Recca::psut_cols$phi}.}

\item{energy_type, energy, exergy}{See \code{Recca::energy_types}.}

\item{eta_fu}{The base name of the output columns.
Default is \code{Recca::efficiency_cols$eta_fu}.}

\item{eta_fu_Y_e}{The name of the energy efficiency output column
for energy flowing into final demand (\strong{Y}).
Default is \code{paste0(eta_fu, "_Y_", energy)}.}

\item{eta_fu_Y_x}{The name of the exergy efficiency output column
for energy flowing into final demand (\strong{Y}).
Default is \code{paste0(eta_fu, "_Y_", exergy)}.}

\item{eta_fu_eiou_e}{The name of the energy efficiency output column
for energy flowing into the energy industry (\strong{U_EIOU}).
Default is \code{paste0(eta_fu, "_EIOU_", energy)}.}

\item{eta_fu_eiou_x}{The name of the exergy efficiency output column
for energy flowing into the energy industry (\strong{U_EIOU}).
Default is \code{paste0(eta_fu, "_EIOU_", energy)}.}

\item{notation}{The notation for the row and column labels of the matrices and vectors.
Default is \code{RCLabels::arrow_notation}.}
}
\value{
A data frame or list containing final-to-useful efficiencies.
}
\description{
Final-to-useful efficiencies can be calculated from
allocations (\code{C_Y} and \code{C_eiou}),
efficiencies (\code{eta_i}), and
(for exergetic efficiencies) exergy-to-energy ratios (\code{phi}).
This function performs those calculations.
}
\details{
The matrix formula for calculating energy efficiencies
is \strong{eta_fu_E} \code{=} \strong{C} \code{*} \strong{eta_i},
where \strong{C} is one of \strong{C_Y} or \strong{C_EIOU}.
The matrix formula for calculating exergy efficiencies
from allocations and machine energy efficiencies is
\strong{eta_fu_X} \code{=} (\strong{phi_u_hat_inv} \code{*} (\strong{C} \code{*} \strong{eta_fu_hat})) \code{*} \strong{phi_u}.

The \strong{C_Y} matrix is assumed to have rows named
with prefixes of final energy carriers and
suffixes of the final demand sector
into which the final energy carrier flows.
The \strong{C_EIOU} matrix is similar, except that
its rows are named with suffixes of the
energy industry sector into which the final energy carrier flows.
The columns of the \strong{C_Y} and \strong{C_EIOU} matrices
are named with prefixes of machine names and
suffixes of useful energy product made by the machine.
See examples.

The \strong{eta_i} vector of machine efficiencies
has rows named with
prefixes same as \strong{C_Y} and \strong{C_EIOU} columns.
The name of the \strong{eta_i} column is not used.
See examples.

The \strong{phi} vector of exergy-to-energy ratios
has rows named with energy carriers
that correspond to the prefixes of
\strong{C_Y} and \strong{C_EIOU} rows and the suffixes of
\strong{C_Y} and \strong{C_EIOU} columns.
See examples.

This function uses \code{\link[matsbyname:vec_from_store_byname]{matsbyname::vec_from_store_byname()}}
to construct the \code{eta_i} and \code{phi} vectors before multiplying, thereby
eliminating unnecessary growth of the output vectors.
}
\examples{
C_Y <- matrix(c(0.7, 0.3, 0,   0,   0,
                0,   0,   0.2, 0.5, 0.3), byrow = TRUE, nrow = 2, ncol = 5,
              dimnames = list(c("Electricity -> Non-ferrous metals",
                                "PSB -> Residential"),
                              c("Electric arc furnaces -> HTH.600.C",
                                "Electric lights -> L",
                                "Wood stoves -> LTH.20.C",
                                "Wood stoves -> LTH.50.C",
                                "Wood stoves -> MTH.100.C")))
C_Y
eta_i <- matrix(c(0.9, 0.2, 0.4, 0.4, 0.3), nrow = 5, ncol = 1,
                  dimnames = list(c("Electric arc furnaces -> HTH.600.C",
                                    "Electric lights -> L",
                                    "Wood stoves -> LTH.20.C",
                                    "Wood stoves -> LTH.50.C",
                                    "Wood stoves -> MTH.100.C"),
                                  "eta_i"))
eta_i
phi <- matrix(c(1, 1.1, 1 - 298.15/(600+273.15), 0.95,
                1 - (20 + 273.15)/298.15,
                1 - 298.15/(50+273.15),
                1 - 298.15/(100+273.15)),
              nrow = 7, ncol = 1,
              dimnames = list(c("Electricity", "PSB", "HTH.600.C", "L",
                                "LTH.20.C", "LTH.50.C", "MTH.100.C"),
                              "phi"))
phi
res <- calc_eta_fu(C_Y = C_Y, C_eiou = C_Y, eta_i = eta_i, phi = phi)
res$eta_fu_Y_E
res$eta_fu_EIOU_E # Same because C_Y and C_EIOU are same
res$eta_fu_Y_X
res$eta_fu_EIOU_X # Same because C_Y and C_EIOU are same
}
