% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency.R
\name{calc_eta_i}
\alias{calc_eta_i}
\title{Calculate industry efficiencies}
\usage{
calc_eta_i(.sutmats, U = "U", V = "V", S_units = "S_units",
  eta_i = "eta_i")
}
\arguments{
\item{.sutmats}{a data frame containing columns for \code{U}, \code{V}, and \code{S_units} matrices}

\item{U}{a string for the name of a column of \code{U} matrices in \code{.sutmats}. (Default is "\code{U}".)}

\item{V}{a string for the name of a column of \code{V} matrices in \code{.sutmats}. (Default is "\code{V}".)}

\item{S_units}{a string for the name of a column of \code{S_units} matrices in \code{.sutmats}. (Default is "\code{S_units}".)}

\item{eta_i}{the name of the industry efficiency column in output. Default is "\code{eta_i}".}
}
\value{
\code{.sutmats} with an additional column "\code{eta_i}"
}
\description{
Calculates industry efficiencies for all energy conversion industries in the ECC.
The efficiency for a given industry is calculated iff the units for inputs and outputs for that industry are unit-homogeneous.
If units for inputs and outputs are heterogeneous for an industry, \code{NA} is the result.
}
\details{
Note that these efficiencies (\code{eta}) are different from
final demand sector and product efficiencies (\code{eta_s} and \code{eta_p}, respectively).
Both final demand sector and product efficiencies
(\code{eta_s} and \code{eta_p}) are based on embodied energy, whereas
industry efficiencies (\code{eta}) is based on direct inputs consumed and outputs produced
by the energy conversion industry.

To calculate energy conversion final demand sector and product efficiencies, use the
\code{\link{calc_embodied_etas}} function.
}
\examples{
library(tidyr)
UKEnergy2000mats \%>\%
  spread(key = "matrix.name", value = "matrix") \%>\%
  calc_eta_i()
}
