% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency.R
\name{calc_eta_i}
\alias{calc_eta_i}
\title{Calculate industry efficiencies}
\usage{
calc_eta_i(.sutdata, U_colname = "f", V_colname = "g",
  S_units_colname = "S_units", eta_i_colname = "eta_i")
}
\arguments{
\item{U_colname}{a string for the name of a column of \code{U} matrices in \code{.sutmats}. (Default is "\code{U}".)}

\item{V_colname}{a string for the name of a column of \code{V} matrices in \code{.sutmats}. (Default is "\code{V}".)}

\item{S_units_colname}{a string for the name of a column of \code{S_units} matrices in \code{.sutmats}. (Default is "\code{S_units}".)}

\item{eta_i_colname}{the name of the industry efficiency column in output. Default is "\code{eta_i}".}

\item{.sutmats}{a data frame containing columns for \code{U}, \code{V}, and \code{S_units} matrices}
}
\value{
\code{.sutmats} with an additional column "\code{eta}"
}
\description{
Calculates industry efficiencies for all energy conversion industries in the ECC.
The efficiency for a given industry is calculated iff the units for inputs and outputs for that industry are unit-homogeneous.
If units for inputs and outpus are heterogeneous for an industry, \code{NA} is the result.
}
\details{
Note that these efficiencies (\code{eta}) are different from
final demand sector and product efficiencies (\code{eta_s} and \code{eta_p}, respectively).
Both final demand sector and product efficiencies
(\code{eta_s} and \code{eta_p}) are based on embodied energy, whereas
industry efficiencies (\code{eta}) is based on direct inputs consumed and outputs produced
by the energy conversion industry.

To calculate energy conversion final demand sector and product efficiencies, use the
\code{\link{calc_embodied_etas}} function.
}
