% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency.R
\name{calc_eta_i}
\alias{calc_eta_i}
\title{Calculate industry efficiencies}
\usage{
calc_eta_i(
  .sutmats,
  U = Recca::psut_cols$U,
  V = Recca::psut_cols$V,
  S_units = Recca::psut_cols$S_units,
  eta_i = Recca::efficiency_cols$eta_i
)
}
\arguments{
\item{.sutmats}{A data frame containing columns for \strong{U}, \strong{V}, and \strong{S_units} matrices.}

\item{U}{A string for the name of a column of \strong{U} matrices in \code{.sutmats}. Default is \code{Recca::psut_cols$U}.}

\item{V}{A string for the name of a column of \strong{V} matrices in \code{.sutmats}. Default is \code{Recca::psut_cols$V}.}

\item{S_units}{A string for the name of a column of \strong{S_units} matrices in \code{.sutmats}. Default is \code{Recca::psut_cols$S_units}.)}

\item{eta_i}{The name of the industry efficiency column in output. Default is \code{Recca::psut_cols$S_units}.}
}
\value{
\code{.sutmats} with an additional column \code{eta_i}
}
\description{
Calculates industry efficiencies for all energy conversion industries in the ECC.
Calculations are performed as shown in Equation 11 in
Heun, Owen, and Brockway. 2018.
A physical supply-use table framework for energy analysis on the energy conversion chain.
Applied Energy, vol 226, pp. 1134-1162.
}
\details{
The efficiency for a given industry is calculated iff the units for inputs and outputs for that industry are unit-homogeneous.
If units for inputs and outputs are heterogeneous for an industry, \code{NA} is the result.

Note that these efficiencies (\code{eta}) are different from
final demand sector and product efficiencies (\code{eta_s} and \code{eta_p}, respectively).
Both final demand sector and product efficiencies
(\code{eta_s} and \code{eta_p}) are based on embodied energy, whereas
industry efficiencies (\code{eta}) is based on direct inputs consumed and outputs produced
by the energy conversion industry.

To calculate energy conversion final demand sector and product efficiencies, use the
\code{calc_embodied_etas()} function.
}
\examples{
library(tidyr)
UKEnergy2000mats \%>\%
  spread(key = "matrix.name", value = "matrix") \%>\%
  calc_eta_i()
}
