% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency.R
\name{calc_eta_pfd}
\alias{calc_eta_pfd}
\title{Calculate aggregate (total) efficiencies}
\usage{
calc_eta_pfd(
  .aggregate_df = NULL,
  add_eta_names = TRUE,
  abbreviate_stage_name = TRUE,
  efficiency_name_suffix = Recca::efficiency_cols$efficiency_name_suffix,
  aggregate_primary_colname = Recca::aggregate_cols$aggregate_primary,
  gross_aggregate_demand_colname = Recca::aggregate_cols$gross_aggregate_demand,
  net_aggregate_demand_colname = Recca::aggregate_cols$net_aggregate_demand,
  energy_type = Recca::psut_cols$energy_type,
  last_stage = Recca::psut_cols$last_stage,
  eta_pfd_gross = Recca::efficiency_cols$eta_pfd_gross,
  eta_pfd_net = Recca::efficiency_cols$eta_pfd_net,
  eta_pfd_gross_colname = paste0(eta_pfd_gross, efficiency_name_suffix),
  eta_pfd_net_colname = paste0(eta_pfd_net, efficiency_name_suffix)
)
}
\arguments{
\item{.aggregate_df}{A data frame or list containing columns
\code{aggregate_primary_colname},
\code{net_aggregate_demand_colname},
\code{gross_aggregate_demand_colname},
probably the result of calling \code{footprint_aggregates()}.}

\item{add_eta_names}{A boolean that tells whether to add columns for the names of
the calculated efficiencies.
Default is \code{TRUE}.}

\item{abbreviate_stage_name}{A boolean that tells whether stage names
are abbreviated to first letter.
E.g., "Primary" --> "p", "Final" --> "f", "Useful" --> "u", "Services" --> "s".
Default is \code{TRUE}.}

\item{efficiency_name_suffix}{The suffix for efficiency names.
Default is \code{Recca::efficiency_cols$efficiency_name_suffix}.}

\item{aggregate_primary_colname}{The name of the column in \code{p_aggregates} that contains primary energy or exergy aggregates.
Default is \code{Recca::aggregate_cols$aggregate_primary}.}

\item{gross_aggregate_demand_colname}{The name of the column in \code{finaldemand_aggregates}
that contains gross final demand aggregates.
Default is \code{Recca::aggregate_cols$gross_aggregate_demand}.}

\item{net_aggregate_demand_colname}{The name of the column in \code{finaldemand_aggregates}
that contains net final demand aggregates.
Default is \code{Recca::aggregate_cols$net_aggregate_demand}.}

\item{eta_pfd_gross}{The name of the output column containing efficiencies
of converting primary energy into gross final demand energy.
Default is \code{Recca::efficiency_cols$eta_pfd_gross}.}

\item{eta_pfd_net}{The name of the output column containing efficiencies
of converting primary energy into net final demand energy.
Default is \code{Recca::efficiency_cols$eta_pfd_net}.}
}
\value{
A data frame of aggregate efficiencies.
}
\description{
Calculates aggregate (total) primary-to-final demand (gross and net) efficiencies
across the energy conversion chain.
\code{.aggregate_df} is probably formed by joining the results from
\code{primary_aggregates()} and \code{finaldemand_aggregates()}
or by calling \code{footprint_aggregates()}.
See examples.
}
\details{
The argument \code{add_eta_names} controls whether,
in addition to calculating the efficiencies,
names for the efficiencies are provided in columns.
\code{add_eta_names = TRUE} (the default) makes it easier to \code{tidyr::pivot_wider()}
on the efficiencies later.
}
\examples{
wide <- primary_total_aggregates_sut <- UKEnergy2000mats \%>\%
  tidyr::pivot_wider(names_from = matrix.name, values_from = matrix)
# Define primary industries
p_industries <- c("Resources - Crude", "Resources - NG")
primary_total_aggregates <- wide \%>\%
  Recca::primary_aggregates(p_industries = p_industries, by = "Total") \%>\%
  # Don't need the matrices
  dplyr::select(IEATools::iea_cols$country,
                IEATools::iea_cols$year,
                IEATools::iea_cols$energy_type,
                IEATools::iea_cols$last_stage,
                Recca::aggregate_cols$aggregate_primary)
# Define final demand sectors
fd_sectors <- c("Residential", "Transport", "Oil fields")
finaldemand_total_aggregates <- wide \%>\%
  Recca::finaldemand_aggregates(fd_sectors = fd_sectors, by = "Total") \%>\%
  # Don't need the matrices
  dplyr::select(IEATools::iea_cols$country,
                IEATools::iea_cols$year,
                IEATools::iea_cols$energy_type,
                IEATools::iea_cols$last_stage,
                Recca::aggregate_cols$gross_aggregate_demand,
                Recca::aggregate_cols$net_aggregate_demand)
dplyr::full_join(primary_total_aggregates,
                 finaldemand_total_aggregates,
                 by = c(IEATools::iea_cols$country,
                        IEATools::iea_cols$year,
                        IEATools::iea_cols$energy_type,
                        IEATools::iea_cols$last_stage)) \%>\%
  calc_eta_pfd()
}
