% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency.R
\name{calc_eta_pfd}
\alias{calc_eta_pfd}
\title{Calculate aggregate (total) efficiencies}
\usage{
calc_eta_pfd(
  .aggregate_df = NULL,
  aggregate_primary_colname = Recca::aggregate_cols$aggregate_primary,
  gross_aggregate_demand_colname = Recca::aggregate_cols$gross_aggregate_demand,
  net_aggregate_demand_colname = Recca::aggregate_cols$net_aggregate_demand,
  eta_pfd_gross = Recca::efficiency_cols$eta_pfd_gross,
  eta_pfd_net = Recca::efficiency_cols$eta_pfd_net
)
}
\arguments{
\item{.aggregate_df}{A data frame or list containing columns
\code{aggregate_primary_colname},
\code{net_aggregate_demand_colname},
\code{gross_aggregate_demand_colname},
probably formed by joining the results from}

\item{aggregate_primary_colname}{The name of the column in \code{p_aggregates} that contains primary energy or exergy aggregates.
Default is \code{Recca::aggregate_cols$aggregate_primary}.}

\item{gross_aggregate_demand_colname}{The name of the column in \code{finaldemand_aggregates}
that contains gross final demand aggregates.
Default is \code{Recca::aggregate_cols$gross_aggregate_demand}.}

\item{net_aggregate_demand_colname}{The name of the column in \code{finaldemand_aggregates}
that contains net final demand aggregates.
Default is \code{Recca::aggregate_cols$net_aggregate_demand}.}

\item{eta_pfd_gross}{The name of the output column containing efficiencies
of converting primary energy into gross final demand energy.
Default is \code{Recca::efficiency_cols$eta_pfd_gross}.}

\item{eta_pfd_net}{The name of the output column containing efficiencies
of converting primary energy into net final demand energy.
Default is \code{Recca::efficiency_cols$eta_pfd_net}.}
}
\value{
A data frame of aggregate efficiencies.
}
\description{
Calculates aggregate (total) primary-to-final demand (gross and net) efficiencies
across the energy conversion chain.
}
\details{
\code{.aggregate_df} is probably formed by joining the results from
\code{primary_aggregates()} and \code{finaldemand_aggregates()}.
See examples.
}
\examples{
wide <- primary_total_aggregates_sut <- UKEnergy2000mats \%>\%
  tidyr::pivot_wider(names_from = matrix.name, values_from = matrix)
# Define primary industries
p_industries <- c("Resources - Crude", "Resources - NG")
primary_total_aggregates <- wide \%>\%
  Recca::primary_aggregates(p_industries = p_industries, by = "Total") \%>\%
  # Don't need the matrices
  dplyr::select(IEATools::iea_cols$country,
                IEATools::iea_cols$year,
                IEATools::iea_cols$energy_type,
                IEATools::iea_cols$last_stage,
                Recca::aggregate_cols$aggregate_primary)
# Define final demand sectors
fd_sectors <- c("Residential", "Transport", "Oil fields")
finaldemand_total_aggregates <- wide \%>\%
  Recca::finaldemand_aggregates(fd_sectors = fd_sectors, by = "Total") \%>\%
  # Don't need the matrices
  dplyr::select(IEATools::iea_cols$country,
                IEATools::iea_cols$year,
                IEATools::iea_cols$energy_type,
                IEATools::iea_cols$last_stage,
                Recca::aggregate_cols$gross_aggregate_demand,
                Recca::aggregate_cols$net_aggregate_demand)
dplyr::full_join(primary_total_aggregates,
                 finaldemand_total_aggregates,
                 by = c(IEATools::iea_cols$country,
                        IEATools::iea_cols$year,
                        IEATools::iea_cols$energy_type,
                        IEATools::iea_cols$last_stage)) \%>\%
  calc_eta_pfd()
}
