% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/losses_irrev.R
\name{calc_exergy_losses_irrev}
\alias{calc_exergy_losses_irrev}
\title{Calculate exergy losses and irreversibility}
\usage{
calc_exergy_losses_irrev(
  .sutmats = NULL,
  R = Recca::psut_cols$R,
  U = Recca::psut_cols$U,
  V = Recca::psut_cols$V,
  Y = Recca::psut_cols$Y,
  U_feed = Recca::psut_cols$U_feed,
  U_eiou = Recca::psut_cols$U_eiou,
  r_eiou = Recca::psut_cols$r_eiou,
  intra_industry_balance = Recca::balance_cols$intra_industry_balance_colname,
  losses_alloc = Recca::balance_cols$losses_alloc_colname,
  loss_sector = Recca::balance_cols$losses_sector,
  phi_vec = Recca::psut_cols$phi,
  replace_cols = FALSE,
  clean = FALSE,
  tol = 1e-06,
  exergy_loss = Recca::psut_cols$exergy_loss,
  irreversibility = Recca::psut_cols$irreversibility
)
}
\arguments{
\item{.sutmats}{An optional wide-by-matrices data frame
containing conversion chains quantified
in conserved quantities (such as mass or energy).
Losses may be included in the matrices,
but that is not required.}

\item{R, U, V, Y, U_feed, U_eiou, r_eiou}{PSUT matrices that describe the conversion chain.}

\item{irreversibility}{}
}
\description{
Given a conversion chain quantified in conserved quantities,
such as mass or energy,
calculate exergy losses and irreversibility
(exergy destruction)
}
\details{
Additional inputs are the losses allocation matrix
(\code{losses_alloc}) and
the phi vector (\code{phi_vec}).
Internally, this function uses
\code{\link[=endogenize_losses]{endogenize_losses()}} and
\code{\link[=extend_to_exergy]{extend_to_exergy()}}.
The algorithm for performing these calculations is:
\itemize{
\item step 1
\item step 2
\item step 3
}
}
