% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embodied.R
\name{calc_F_footprint_effects}
\alias{calc_F_footprint_effects}
\title{Upstream footprint and downstream effects matrices}
\usage{
calc_F_footprint_effects(.Mdata, M_p_colname = "M_p", M_s_colname = "M_s",
  F_footprint_p_colname = "F_footprint_p",
  F_effects_p_colname = "F_effects_p",
  F_footprint_s_colname = "F_footprint_s",
  F_effects_s_colname = "F_effects_s")
}
\arguments{
\item{.Mdata}{a data frame containing a column of embodied matrices}

\item{M_p_colname}{the name of the column in \code{.Mdata} containing embodied product  matrices (default is \strong{"M_p"})}

\item{M_s_colname}{the name of the column in \code{.Mdata} containing embodied sector matrices (default is \strong{"M_s"})}

\item{F_footprint_p_colname}{the name of the column in the output containing \strong{F_footprint_p} matrices (as a string)}

\item{F_effects_p_colname}{the name of the column in the output containing \strong{F_effects_p} matrices (as a string)}

\item{F_footprint_s_colname}{the name of the column in the output containing \strong{F_footprint_s} matrices (as a string)}

\item{F_effects_s_colname}{the name of the column in the output containing \strong{F_effects_s} matrices (as a string)}
}
\value{
\code{.Mdata} with columns \code{F_footprint_p}, \code{F_effects_p},
\code{F_footprint_s}, and \code{F_effects_s} added
}
\description{
Calculates upstream footprint matrices (\strong{F_footprint_p}, \strong{F_footprint_s})
and downstream effects matrices (\strong{F_effects_p}, \strong{F_effects_s})
given an embodied matries \strong{M_p} and \strong{M_s}.
Column sums of \strong{F_footprint} are 1.
Row sums of \strong{F_effects} are 1.
}
