% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embodied.R
\name{calc_F_footprint_effects}
\alias{calc_F_footprint_effects}
\title{Upstream footprint and downstream effects matrices}
\usage{
calc_F_footprint_effects(
  .Mmats = NULL,
  M_p = "M_p",
  M_s = "M_s",
  F_footprint_p = "F_footprint_p",
  F_effects_p = "F_effects_p",
  F_footprint_s = "F_footprint_s",
  F_effects_s = "F_effects_s"
)
}
\arguments{
\item{.Mmats}{A data frame containing a column of embodied matrices.}

\item{M_p}{An embodied product matrix or name of the column in \code{.Mmats} containing same. Default is "M_p".}

\item{M_s}{An embodied sector matrix or name of the column in \code{.Mmats} containing same. Default is "M_s".}

\item{F_footprint_p}{The name for \strong{F_footprint_p} matrices on output. Default is "F_footprint_p".}

\item{F_effects_p}{The name for \strong{F_effects_p} matrices on output. Default is "F_effects_p".}

\item{F_footprint_s}{The name for \strong{F_footprint_s} matrices on output. Default is "F_footprint_s".}

\item{F_effects_s}{The name for \strong{F_effects_s} matrices on output. Default is "F_effects_s".}
}
\value{
A list or data frame containing \strong{F_footprint_p}, \strong{F_effects_p},
\strong{F_footprint_s}, and \strong{F_effects_s} matrices.
}
\description{
Calculates upstream footprint matrices (\strong{F_footprint_p} and \strong{F_footprint_s})
and downstream effects matrices (\strong{F_effects_p} and \strong{F_effects_s})
given embodied matrices \strong{M_p} and \strong{M_s}.
Column sums of \strong{F_footprint} are \code{1}.
Row sums of \strong{F_effects} are \code{1}.
}
