% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embodied.R
\name{calc_GH}
\alias{calc_GH}
\title{Calculate the G and H matrices for embodied energy calculations}
\usage{
calc_GH(.iodata = NULL, Y_colname = "Y", L_ixp_colname = "L_ixp",
  G_colname = "G", H_colname = "H")
}
\arguments{
\item{.iodata}{a data frame containing matrices that describe the Input-Output structure of an Energy Conversion Chain.
\code{.iodata} will likely have been obtained from the \code{calc_io_mats} function.}

\item{Y_colname}{the name of the column in \code{.iodata} containing final demand (\code{Y}) matrices.}

\item{L_ixp_colname}{the name of the column in \code{.iodata} containing Industry-by-Product
Leontief (\code{L_ixp}) matrices.}

\item{G_colname}{the name of the output column containing \code{G} matrices.
\code{G} is calculated by \code{L_ixp * y_hat}.}

\item{H_colname}{the name of the output column containing \code{H} matrices.
\code{G} is calculated by \code{L_ixp * Y}.}
}
\value{
\code{.iodata} with columns \code{G_colname} and \code{H_colname} added.
}
\description{
Calculate the G and H matrices for embodied energy calculations
}
