% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embodied.R
\name{calc_GH}
\alias{calc_GH}
\title{Calculate the G and H matrices for embodied energy calculations}
\usage{
calc_GH(.iodata, y_colname = "y", Y_colname = "Y",
  L_ixp_colname = "L_ixp", keep_cols = NULL, G_colname = "G",
  H_colname = "H")
}
\arguments{
\item{.iodata}{a data frame containing matrices that describe the Input-Output structure of an Energy Conversion Chain.
\code{.iodata} will likely have been obtained from the \code{calc_io_mats} function.}

\item{y_colname}{the name of the column in \code{.iodata} containing final demand (\code{y}) vectors.}

\item{Y_colname}{the name of the column in \code{.iodata} containing final demand (\code{Y}) matrices.}

\item{L_ixp_colname}{the name of the column in \code{.iodata} containing Industry-by-Product
Leontief (\code{L_ixp}) matrices.}

\item{keep_cols}{a vector of names of columns of \code{.iodata} to return with the output}

\item{G_colname}{the name of the output column containing \code{G} matrices.
\code{G} is calculated by \code{L_ixp * y_hat}.}

\item{H_colname}{the name of the output column containing \code{H} matrices.
\code{G} is calculated by \code{L_ixp * Y}.}
}
\value{
a data frame containing columns specified in \code{keep_cols} and \code{G_colname}.
}
\description{
Calculate the G and H matrices for embodied energy calculations
}
