% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embodied.R
\name{calc_GH}
\alias{calc_GH}
\title{Calculate the \code{G} and \code{H} matrices for embodied energy calculations}
\usage{
calc_GH(.iomats = NULL, Y = "Y", L_ixp = "L_ixp", G = "G",
  H = "H")
}
\arguments{
\item{.iomats}{a data frame containing matrices that describe the Input-Output structure of an Energy Conversion Chain.
\code{.iomats} will likely have been obtained from the \code{\link{calc_io_mats}} function.}

\item{Y}{final demand (\code{Y}) matrix or name of the column in \code{.iodata} containing same. Default is "\code{Y}".}

\item{L_ixp}{industry-by-product Leontief (\code{L_ixp}) matrix or name of the column in \code{.iodata} containing same. Default is "\code{L_ixp}".}

\item{G}{name for the \code{G} matrix on output. Default is "\code{G}".
\code{G} is calculated by \code{L_ixp * y_hat}.}

\item{H}{name for the \code{H} matrix on output. Default is "\code{H}".
\code{G} is calculated by \code{L_ixp * Y}.}
}
\value{
a list or data frame containing \code{G} and \code{H} matrices.
}
\description{
Calculate the \code{G} and \code{H} matrices for embodied energy calculations
}
