% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{calc_io_mats}
\alias{calc_io_mats}
\title{Calculate several input-output matrices}
\usage{
calc_io_mats(.sutdata = NULL, U = "U", V = "V", Y = "Y",
  S_units = "S_units", y = "y", q = "q", f = "f", g = "g",
  W = "W", K = "K", Z = "Z", C = "C", D = "D", A = "A",
  L_ixp = "L_ixp", L_pxp = "L_pxp")
}
\arguments{
\item{.sutdata}{a data frame of supply-use table matrices with matrices arranged in columns.}

\item{U}{use (\code{U}) matrix or name of the column in \code{.sutmats} that contains same. Default is "\code{U}".}

\item{V}{make (\code{V}) matrix or name of the column in \code{.sutmats}that contains same. Default is "\code{V}".}

\item{Y}{final demand (\code{Y}) matrix or name of the column in \code{.sutmats} that contains same. Default is "\code{Y}".}

\item{S_units}{\code{S_units} matrix or name of the column in \code{.sutmats} that contains same. Default is "\code{S_units}".}

\item{y}{name for \code{y} vector on output. Default is "\code{y}".
\code{y} is calculated by \code{rowsums(Y)}.}

\item{q}{name for \code{q} vector on output. Default is "\code{q}".
\code{q} is calculated by \code{rowsums(U) + y}.}

\item{f}{name for \code{f} vector on output. Default is "\code{f}".
\code{f} is calculated by \code{colsums(U)}.}

\item{g}{name for \code{g} vector on output. Default is "\code{g}".
\code{g} is calculated by \code{rowsums(V)}.}

\item{W}{name for \code{W} matrix on output. Default is "\code{W}".
\code{W} is calculated by \code{transpose(V) - U}.}

\item{K}{name for \code{K} matrix on output. Default is "\code{K}".
\code{K} is calculated by \code{U * f_hat_inv}.}

\item{Z}{name for \code{Z} matrix on output. Default is "\code{Z}".
\code{Z} is calculated by \code{U * g_hat_inv}.}

\item{C}{name for \code{C} matrix on output. Default is "\code{C}".
\code{C} is calculated by \code{transpose(V) * g_hat_inv}.}

\item{D}{name for \code{D} matrix on output. Default is "\code{D}".
\code{D} is calculated by \code{V * q_hat_inv}.}

\item{A}{name for \code{A} matrix on output. Default is "\code{A}".
\code{A} is calculated by \code{Z * D}.}

\item{L_ixp}{name for \code{L_ixp} matrix on output. Default is "\code{L_ixp}".
\code{L_ixp} is calculated by \code{D * L_pxp}.}

\item{L_pxp}{name for \code{L_pxp} matrix on output. Default is "\code{L_pxp}".
\code{L_pxp} is calculated by \code{(I - Z*D)^-1}.}
}
\value{
a list or data frame containing input-output matrices
}
\description{
Calculate several input-output matrices
}
\examples{
library(dplyr)
library(tidyr)
UKEnergy2000mats \%>\%
  spread(key = matrix.name, value = matrix) \%>\%
  select(Country, Year, Energy.type, Last.stage, U, V, Y, r_EIOU, S_units) \%>\%
  calc_io_mats()
}
