% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{calc_io_mats}
\alias{calc_io_mats}
\title{Calculate several input-output matrices}
\usage{
calc_io_mats(.sutdata = NULL, U = "U", V = "V", Y = "Y",
  S_units = "S_units", y = "y", q = "q", f = "f", g = "g",
  W = "W", K = "K", Z = "Z", C = "C", D = "D", A = "A",
  L_ixp = "L_ixp", L_pxp = "L_pxp")
}
\arguments{
\item{.sutdata}{a data frame of supply-use table matrices with matrices arranged in columns.}

\item{U}{use (\code{U}) matrix or name of the column in \code{.sutmats} that contains same. Default is "\code{U}".}

\item{V}{make (\code{V}) matrix or name of the column in \code{.sutmats}that contains same. Default is "\code{V}".}

\item{Y}{final demand (\code{Y}) matrix or name of the column in \code{.sutmats} that contains same. Default is "\code{Y}".}

\item{S_units}{\code{S_units} matrix or name of the column in \code{.sutmats} that contains same. Default is "\code{S_units}".}

\item{y}{final demand (\code{y}) vector or name of the column in \code{.sutmats} that contains same. Default is "\code{y}".
\code{y} is calculated by \code{rowsums(Y)}.}

\item{q}{\code{q} vector or name of the column in \code{.sutmats} that contains same. Default is "\code{q}".
\code{q} is calculated by \code{rowsums(U) + y}.}

\item{f}{\code{f} vector or name of the column in \code{.sutmats} that contains same. Default is "\code{r}".
\code{f} is calculated by \code{colsums(U)}.}

\item{g}{\code{g} vector or name of the column in \code{.sutmats} that contains same. Default is "\code{g}".
\code{g} is calculated by \code{rowsums(V)}.}

\item{W}{\code{W} matrix or name of the column in \code{.sutmats} that contains same. Default is "\code{W}".
\code{W} is calculated by \code{transpose(V) - U}.}

\item{K}{\code{K} matrix or name of the column in \code{.sutmats} that contains same. Default is "\code{K}".
\code{K} is calculated by \code{U * f_hat_inv}.}

\item{Z}{\code{Z} matrix or name of the column in \code{.sutmats} that contains same. Default is "\code{Z}".
\code{W} is calculated by \code{U * g_hat_inv}.}

\item{C}{\code{C} matrix or name of the column in \code{.sutmats} that contains same. Default is "\code{C}".
\code{C} is calculated by \code{transpose(V) * g_hat_inv}.}

\item{D}{\code{D} matrix or name of the column in \code{.sutmats} that contains same. Default is "\code{D}".
\code{D} is calculated by \code{V * q_hat_inv}.}

\item{A}{\code{A} matrix or name of the column in \code{.sutmats} that contains same. Default is "\code{A}".
\code{A} is calculated by \code{Z * D}.}

\item{L_ixp}{\code{L_ixp} matrix or name of the column in \code{.sutmats} that contains same. Default is "\code{L_ixp}".
\code{L_ixp} is calculated by \code{D * L_pxp}.}

\item{L_pxp}{\code{L_pxp} matrix or name of the column in \code{.sutmats} that contains same. Default is "\code{L_pxp}".
\code{L_pxp} is calculated by \code{(I - Z*D)^-1}.}
}
\value{
a list or data frame containing input-output matrices
}
\description{
Calculate several input-output matrices
}
