% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{calc_L}
\alias{calc_L}
\title{Calculates total requirements matrices (\strong{L_pxp} and \strong{L_ixp})}
\usage{
calc_L(
  .sutdata = NULL,
  method = c("solve", "QR", "SVD"),
  tol = .Machine$double.eps,
  D = "D",
  A = "A",
  L_pxp = "L_pxp",
  L_ixp = "L_ixp"
)
}
\arguments{
\item{.sutdata}{A data frame of supply-use table matrices with matrices arranged in columns.
Default is \code{NULL}, meaning that matrices will be taken from the \code{D} and \code{A} arguments.
Set to a list or data frame to pull matrices from its store.}

\item{method}{One of "solve", "QR", or "SVD". Default is "solve". See details.}

\item{tol}{The tolerance for detecting linear dependencies during matrix inversion.
Default is \code{.Machine$double.eps}.}

\item{D}{The \strong{D} matrix or name of the column in \code{.sutmats} that contains same. Default is "D".}

\item{A}{The \strong{A} matrix or name of the column in \code{.sutmats} that contains same. Default is "A".}

\item{L_pxp}{The name for the \strong{L_pxp} matrix on output. Default is "L_pxp".
\code{L_pxp} is calculated by \code{inverse(I - A)}.}

\item{L_ixp}{The name for the \strong{L_ixp} matrix on output. Default is "L_ixp".
\strong{L_ixp} is calculated by \code{D * L_pxp}.}
}
\value{
A list or data frame containing \strong{L_pxp} and \strong{L_ixp} matrices.
}
\description{
\strong{L_pxp} tells how much of a product (in a row) is required to make another product (in a column).
\strong{L_ixp} tells how much of an industry's output (in a row) is required to make another product (in a column).
}
\details{
Calculating some matrices requires
a matrix inversion operation.
The \code{method} argument specifies which method should be used for
calculating the inverse.
See \code{matsbyname::invert_byname()}.

Both \code{tol} and \code{method} should be single values and apply to all matrices being inverted.
}
