% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embodied.R
\name{calc_M}
\alias{calc_M}
\title{Add embodied matrices colums to a data frame}
\usage{
calc_M(.YqGHEdata = NULL, Y_colname = "Y", q_colname = "q",
  G_colname = "G", E_colname = "E", tol = 1e-04, M_p_colname = "M_p",
  M_s_colname = "M_s")
}
\arguments{
\item{.YqGHEdata}{a data frame containing columns with \strong{q} vectors
and \strong{Y}, \strong{G}, \strong{H}, and \strong{E} matrices.
\code{.YqGEdata} will likely have been obtained from the \code{calc_G} and \code{calc_E} functions.}

\item{Y_colname}{the name of the output column containing \strong{Y} matrices.
\strong{Y} is the final demand matrix.}

\item{q_colname}{the name of the output column containing \strong{q} column vectors.
\strong{q} is calculated by \code{\strong{Ui} + \strong{y}}.}

\item{G_colname}{the name of the output column containing \code{G} matrices.
\code{G} is calculated by \code{\strong{L_ixp} * \strong{y_hat}}.}

\item{E_colname}{the name of the output column containing \strong{E} matrices.
\code{E} is calculated by \code{\strong{W} * \strong{g_hat_inv}}.}

\item{tol}{the allowable energy balance error.
\code{Q} is calculated by \code{e_hat * G},
but the e_hat column contains lists of matrices,
so the \code{Q} column will also contain lists of matrices.
In each list, there is one Q matrix for each Product in the Energy Conversion Chain.}

\item{M_p_colname}{the name of the output column containing matrices of embodied energy in products.
These matrices contain embodied products in rows and embodying products in columns.}

\item{M_s_colname}{the name of the output column containing matrices of embodied energy consumed by final demand sectors.
These matrices contain embodied products in rows and consuming final demand sectors in columns.}

\item{Q_colname}{the name of the output column containing lists of \strong{Q} matrices.}
}
\value{
\code{.YqGHEdata} with columns \code{Q_colname}, \code{M_p_colname}, and \code{M_s_colname} added
}
\description{
Add embodied matrices colums to a data frame
}
