% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embodied.R
\name{calc_M}
\alias{calc_M}
\title{Add embodied matrices colums to a data frame}
\usage{
calc_M(.YqGHEdata, Y_colname = "Y", q_colname = "q", G_colname = "G",
  E_colname = "E", M_p_colname = "M_p", M_s_colname = "M_s",
  e_colname = ".evectors", e_hat_colname = ".ehatmatrices",
  Q_colname = ".Qmatrices", Qpos_colname = ".Qposmatrices",
  Qposcolsums_colname = ".Qposcolsums")
}
\arguments{
\item{.YqGHEdata}{a data frame containing columns with \strong{q} vectors
and \strong{Y}, \strong{G}, \strong{H}, and \strong{E} matrices.
\code{.YqGEdata} will likely have been obtained from the \code{calc_G} and \code{calc_E} functions.}

\item{Y_colname}{the name of the output column containing \strong{Y} matrices.
\strong{Y} is the final demand matrix.}

\item{q_colname}{the name of the output column containing \strong{q} column vectors.
\strong{q} is calculated by \code{\strong{Ui} + \strong{y}}.}

\item{G_colname}{the name of the output column containing \code{G} matrices.
\code{G} is calculated by \code{\strong{L_ixp} * \strong{y_hat}}.}

\item{E_colname}{the name of the output column containing \strong{E} matrices.
\code{E} is calculated by \code{\strong{W} * \strong{g_hat_inv}}.}

\item{M_p_colname}{the name of the output column containing matrices of embodied energy in products.
These matrices contain embodied products in rows and embodying products in columns.}

\item{M_s_colname}{the name of the output column containing matrices of embodied energy consumed by final demand sectors.
These matrices contain embodied products in rows and consuming final demand sectors in columns.}

\item{e_colname}{the name of the column containing \strong{e} vectors.
Each \strong{e} vector is a row of the \strong{E} matrix.
Thus, column \code{e_colname} contains lists of \strong{e} vectors,
one \strong{e} vector for each row in the corresponding \strong{E} matrix.
\code{e_colname} is a column of intermediate results that is not included in the output.}

\item{e_hat_colname}{the name of the column containing \code{e_hat} matrices.
Each \strong{e} vector is converted to an \strong{e_hat} matrix.
Thus the \code{e_hat_colname} column contains lists of \strong{e_hat} matrices.
\code{e_hat_colname} is a column of intermediate results that is not included in the output.}

\item{Q_colname}{the name of the output column containing lists of \strong{Q} matrices.
\strong{Q} is calculated by \code{\strong{e_hat} * \strong{G}},
but the e_hat column contains lists of matrices,
so the \code{Q} column will also contain lists of matrices.
In each list, there is one Q matrix for each Product in the Energy Conversion Chain.}

\item{Qpos_colname}{the name of the column containing \strong{Qpos} matrices.
Embodied energy entries in \strong{Q} are positive.}

\item{Qposcolsums_colname}{the name of the column containing column sums of \strong{Qpos} matrices.}

\item{keep_cols}{a vector of names of columns of \code{.iodata} (if any) to return with the output}
}
\value{
a data frame containing columns specified in \code{keep_cols},
\code{Q_colname}, \code{M_p_colname}, and \code{M_s_colname}.
}
\description{
Add embodied matrices colums to a data frame
}
