% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{calc_yqfgW}
\alias{calc_yqfgW}
\title{Calculate \code{y}, \code{f}, \code{g}, and \code{q} vectors and the \code{W} matrix}
\usage{
calc_yqfgW(.sutdata = NULL, U = "U", V = "V", Y = "Y",
  S_units = "S_units", y = "y", q = "q", f = "f", g = "g",
  W = "W")
}
\arguments{
\item{.sutdata}{a data frame of supply-use table matrices with matrices arranged in columns.}

\item{U}{use (\code{U}) matrix or name of the column in \code{.sutmats} that contains same. Default is "\code{U}".}

\item{V}{make (\code{V}) matrix or name of the column in \code{.sutmats}that contains same. Default is "\code{V}".}

\item{Y}{final demand (\code{Y}) matrix or name of the column in \code{.sutmats} that contains same. Default is "\code{Y}".}

\item{S_units}{\code{S_units} matrix or name of the column in \code{.sutmats} that contains same. Default is "\code{S_units}".}

\item{y}{final demand (\code{y}) vector or name of the column in \code{.sutmats} that contains same. Default is "\code{y}".
\code{y} is calculated by \code{rowsums(Y)}.}

\item{q}{\code{q} vector or name of the column in \code{.sutmats} that contains same. Default is "\code{q}".
\code{q} is calculated by \code{rowsums(U) + y}.}

\item{f}{\code{f} vector or name of the column in \code{.sutmats} that contains same. Default is "\code{r}".
\code{f} is calculated by \code{colsums(U)}.}

\item{g}{\code{g} vector or name of the column in \code{.sutmats} that contains same. Default is "\code{g}".
\code{g} is calculated by \code{rowsums(V)}.}

\item{W}{\code{W} matrix or name of the column in \code{.sutmats} that contains same. Default is "\code{W}".
\code{W} is calculated by \code{transpose(V) - U}.}
}
\value{
a list or data frame containing \code{y}, \code{q},
         \code{f}, \code{g}, and \code{W}.
}
\description{
Note that a necessary condition for calculating the \code{f} and \code{g} vectors is that
the U_bar and V_bar matrices should have only one entry per column and row, respectively,
meaning that all products entering a given industry need to be unit homogeneous
before we can calculate the \code{f} vector and
all products of a given industry are measured in the same units
before we can calculate the \code{g} vector.
If the unit homogeneity assumptions above are violated, we will return NA
for violating industries in the \code{f} and \code{g} vectors.
The checks for unit homogenity are performed only when an \code{S_units} matrix is present.
}
