% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{calc_yqfgW}
\alias{calc_yqfgW}
\title{Calculate \strong{y}, \strong{f}, \strong{g}, \strong{q}, \strong{h}, and \strong{r} vectors and the \strong{W} matrix}
\usage{
calc_yqfgW(
  .sutdata = NULL,
  method_q_calculation = c("sum_U_Y_rows", "sum_R_V_cols"),
  R = "R",
  U = "U",
  V = "V",
  Y = "Y",
  S_units = "S_units",
  y = "y",
  q = "q",
  f = "f",
  g = "g",
  h = "h",
  r = "r",
  W = "W"
)
}
\arguments{
\item{.sutdata}{a data frame of supply-use table matrices with matrices arranged in columns.}

\item{method_q_calculation}{Specifies the method with which the q vector should be calculated. See details.}

\item{R}{The resources (\strong{R}) matrix or name of the column in \code{.sutmats} that contains same. Default is "R".}

\item{U}{The use (\strong{U}) matrix or name of the column in \code{.sutmats} that contains same. Default is "U".}

\item{V}{The make (\strong{V}) matrix or name of the column in \code{.sutmats} that contains same. Default is "V".}

\item{Y}{The final demand (\strong{Y}) matrix or name of the column in `.sutmats`` that contains same. Default is "Y".}

\item{S_units}{The \strong{S_units} matrix or name of the column in \code{.sutmats} that contains same. Default is "S_units".}

\item{y}{The name for the \strong{y} vector on output. Default is "y".
\strong{y} is calculated by \code{rowsums(Y)}.}

\item{q}{The name for the \strong{q} vector on output. Default is "q".
\strong{q} is calculated by \code{rowsums(U) + y}.}

\item{f}{The name for the \strong{f} vector on output. Default is "f".
\strong{f} is calculated by \code{colsums(U)}.}

\item{g}{The name for the \strong{g} vector on output. Default is "g".
\strong{g} is calculated by \code{rowsums(V)}.}

\item{h}{The name for the \strong{h} vector on output. Default is "h".
\strong{h} is calculated by \code{colsums(transpose(R))}.}

\item{r}{The name for the \strong{r} vector on output. Default is "r".
\strong{r} is calculated by \code{rowsums(R)}.}

\item{W}{The name for the \strong{W} matrix on output. Default is "W".
\strong{W} is calculated by \code{transpose(V) - U}.}
}
\value{
A list or data frame containing \strong{y}, \strong{q},
\strong{f}, \strong{g}, \strong{h}, and \strong{r} vectors and the \strong{W} matrix.
}
\description{
Note that a necessary condition for calculating the \strong{f}, \strong{g}, and \strong{r} vectors is that
the \strong{R_bar}, \strong{U_bar}, and \strong{V_bar} matrices
should have only one entry per column and row, respectively,
meaning that all products entering a given industry need to be unit homogeneous
before we can calculate the \strong{f} vector and
all products of a given industry are measured in the same units
before we can calculate the \strong{g} vector.
If the unit homogeneity assumptions above are violated, we will return \code{NA}
for violating industries in the \strong{f} and \strong{g} vectors.
The checks for unit homogeneity are performed only when an \strong{S_units} matrix is present.
}
\details{
\code{method_q_calculation} specifies the method with which the q vector should be calculated.
Default is "sum_U_Y_rows", corresponding to a demand-sided view of \strong{q}.
Alternatively, an analyst can choose to use the "sum_R_V_cols" method,
corresponding to a supply-sided view of \strong{q}.
In the case of a balanced ECC, the method does not matter.
}
