% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{calc_yqfgW}
\alias{calc_yqfgW}
\title{Calculate \code{y}, \code{f}, \code{g}, and \code{q} vectors and the \code{W} matrix}
\usage{
calc_yqfgW(
  .sutdata = NULL,
  method_q_calculation = c("sum_U_Y_rows", "sum_R_V_cols"),
  R = "R",
  U = "U",
  V = "V",
  Y = "Y",
  S_units = "S_units",
  y = "y",
  q = "q",
  f = "f",
  g = "g",
  h = "h",
  r = "r",
  W = "W"
)
}
\arguments{
\item{.sutdata}{a data frame of supply-use table matrices with matrices arranged in columns.}

\item{method_q_calculation}{Specifies the method which with the q vector should be calculated.
Default is \code{sum_U_Y_rows}.
Alternatively, an analyst can choose to use the \code{sum_R_V_cols} method.
In the case of a balanced ECC, the method does not matter.}

\item{R}{The resources (\code{R}) matrix or name of the column in \code{.sutmats} that contains same. Default is "R".}

\item{U}{The use (\code{U}) matrix or name of the column in \code{.sutmats} that contains same. Default is "U".}

\item{V}{The make (\code{V}) matrix or name of the column in \code{.sutmats} that contains same. Default is "V".}

\item{Y}{The final demand (\code{Y}) matrix or name of the column in `.sutmats`` that contains same. Default is "Y".}

\item{S_units}{The \code{S_units} matrix or name of the column in \code{.sutmats} that contains same. Default is "S_units".}

\item{y}{The name for the \code{y} vector on output. Default is "y".
\code{y} is calculated by \code{rowsums(Y)}.}

\item{q}{The name for the \code{q} vector on output. Default is "q".
\code{q} is calculated by \code{rowsums(U) + y}.}

\item{f}{The name for the \code{f} vector on output. Default is "f".
\code{f} is calculated by \code{colsums(U)}.}

\item{g}{The name for the \code{g} vector on output. Default is "g".
\code{g} is calculated by \code{rowsums(V)}.}

\item{h}{The name for the \code{h} vector on output. Default is "h".
\code{h} is calculated by \code{colsums(transpose(R))}.}

\item{r}{The name for the \code{r} vector on output. Default is "r".
\code{r} is calculated by \code{rowsums(R)}.}

\item{W}{The name for the \code{W} matrix on output. Default is "W".
\code{W} is calculated by \code{transpose(V) - U}.}
}
\value{
A list or data frame containing \code{y}, \code{q},
\code{f}, \code{g}, and \code{W}.
}
\description{
Note that a necessary condition for calculating the \code{f} and \code{g} vectors is that
the U_bar and V_bar matrices should have only one entry per column and row, respectively,
meaning that all products entering a given industry need to be unit homogeneous
before we can calculate the \code{f} vector and
all products of a given industry are measured in the same units
before we can calculate the \code{g} vector.
If the unit homogeneity assumptions above are violated, we will return NA
for violating industries in the \code{f} and \code{g} vectors.
The checks for unit homogeneity are performed only when an \code{S_units} matrix is present.
}
