% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{check_worksheet_name_violations}
\alias{check_worksheet_name_violations}
\title{Develop a warning message for malformed Excel worksheet names}
\usage{
check_worksheet_name_violations(candidate_sheet_names)
}
\arguments{
\item{candidate_worksheet_names}{Worksheet names to be checked.}
}
\value{
\code{NULL} invisibly and a warning if any problems are detected.
}
\description{
\code{write_ecc_to_excel()} can include worksheet names, but
it is important that they are legal names.
This function emits a warning when \code{candidate_worksheet_names}
is mal-formed.
}
