% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{combine_RV}
\alias{combine_RV}
\title{Combine resource (\code{R}) and make (\code{V}) matrices into a make plus resource (\code{R_plus_V}) matrix}
\usage{
combine_RV(.sutmats = NULL, R = "R", V = "V", R_plus_V = "R_plus_V")
}
\arguments{
\item{.sutmats}{a list or data frame containing use matrix(ces) and make matrix(ces)}

\item{R}{an \code{R} matrix or name of a column in \code{.sutmats} that contains same. Default is "\code{R}".}

\item{V}{a make (\code{V}) matrix or name of a column in \code{.sutmats} that contains same. Default is "\code{V}".}

\item{R_plus_V}{name for \code{R_plus_V} matrix on output. Default is "\code{R_plus_V}".}
}
\value{
a list or data frame containing \code{R_plus_V}
}
\description{
\code{\link{combine_RV}} is the inverse of \code{\link{separate_RV}}.
}
\examples{
library(dplyr)
library(tidyr)
UKEnergy2000mats \%>\%
  spread(key = "matrix.name", value = "matrix") \%>\%
  combine_RV()
}
