% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregates.R
\name{despecified_aggregates}
\alias{despecified_aggregates}
\title{Despecify and aggregate PSUT matrices}
\usage{
despecified_aggregates(
  .sut_data = NULL,
  piece_to_keep = "pref",
  R = Recca::psut_cols$R,
  U = Recca::psut_cols$U,
  V = Recca::psut_cols$V,
  Y = Recca::psut_cols$Y,
  r_EIOU = Recca::psut_cols$r_eiou,
  U_EIOU = Recca::psut_cols$U_eiou,
  U_feed = Recca::psut_cols$U_feed,
  inf_notation = TRUE,
  notation = list(RCLabels::notations_list),
  margin = list(c(1, 2)),
  choose_most_specific = TRUE,
  prepositions = list(RCLabels::prepositions_list),
  R_colname = Recca::psut_cols$R,
  U_colname = Recca::psut_cols$U,
  V_colname = Recca::psut_cols$V,
  Y_colname = Recca::psut_cols$Y,
  r_EIOU_colname = Recca::psut_cols$r_eiou,
  U_EIOU_colname = Recca::psut_cols$U_eiou,
  U_feed_colname = Recca::psut_cols$U_feed
)
}
\arguments{
\item{.sut_data}{A data frame of matrices to be despecified and aggregated.}

\item{piece_to_keep}{The piec of the label which is aggregated.
Default is "noun".}

\item{R, U, V, Y, r_EIOU}{Names of matrices to be despecified and aggregated. See \code{Recca::psut_cols}.}
}
\value{
A modified version of \code{.sut_data} where rows and columns of matrices
have been aggregated to their despecified parts.
}
\description{
PSUT matrices often have row and column names that been specified to
contain more information than simply the industry or product.
Examples include
"Resources [of Coal]" and
"Automobiles -> RoP".
It is sometimes helpful to despecify and aggregate these rows and columns
so that all "Resources" are summed together,
all "Automobiles" are summed together, etc.
This function performs that aggregation.
}
\details{
By default, the aggregation is made to the nouns of row and column names
as defined by the \code{RCLabels} package.
}
