% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecc_layout.R
\name{ecc_layout}
\alias{ecc_layout}
\title{Layout for graph representation of an energy conversion chain}
\usage{
ecc_layout(Industries, Products, industry_colname = "Industry",
  product_colname = "Product", stage_colname = "Stage",
  storage_stagename = "Storage", group_colname = "Group",
  node_name_colnames = "Node_name", x_colname = "x", y_colname = "y")
}
\arguments{
\item{Industries}{a data frame consisting of columns named
\code{industry_colname} and \code{stage_colname}.
\code{industry_colname} must contain names of industries.
\code{stage_colname} must contain the name of the ECC stage
in which the corresponding industry belongs.
\code{stage_colname} is 1-based, i.e., the leftmost industry stage should be stage 1;
the second industry stage should be stage 2, and the rightmost ( final demand)
stage should have the highest number.
Industries in the final demand industry must be included in the \code{Industries} data frame.
The left-to-right order of stages in the network layout generated by this function
is taken from the order of appearance of stages in the \code{stage_colname} column.
Do not include any "storage" industries in the Industries data frame;
rather, include "storage" indusries in the Storage data frame.
Storage industries
(such as Stock changes, International aviation and marine bunkers, and Statistical differences)
should be included in the \code{Industries} data frame
with corresponding \code{stage_colname} of \code{storage_stagename}.
All industries with \code{storage_stagename} are laid out along the top of the
network.
The left-to-right order of storage industries is given by
the appearance order of the industries in \code{industry_colname}.}

\item{Products}{a data frame consisting of columns named
\code{product_colname} and \code{stage_colname}.
\code{product_colname} must contain names of products.
\code{stage_colname} must contain the name of the ECC stage
in which the corresponding product belongs.
\code{stage_colname} is 1-based, i.e., the leftmost product stage should be stage 1;
the second product stage should be stage 2, and the rightmost stage
should have the highest number.}

\item{industry_colname}{the name of the column in \code{Industries} containing
names of industries (a string).
Default is "Industries".}

\item{product_colname}{the name of the column in \code{Products} containing
names of products (a string).
Default is "Products".}

\item{stage_colname}{the name of the columns in \code{Industries} and \code{Products}
containing names of stages (a string).
The left-to-right order of stages in the network layout
is determined by the top-to-bottom
order in which stage names appear in the
\code{Industries} and \code{Products} data frames.}

\item{storage_stagename}{the name of the stage in \code{stage_colname} of \code{Industries}
that identifies a "storage" industry (a string).
Default is "Storage".}

\item{group_colname}{the name of the column in \code{Industries} and \code{Products}
containing industries and products that should be grouped together vertically
at a stage (a string).
The top-to-bottom order of groups at a stage is give by the order of appearance
in \code{group_colname}.}

\item{node_name_colnames}{the name of the output column containing node names
(industries and products).}

\item{x_colname}{the name of the output column containing x coordinates for each node.}

\item{y_colname}{the name of the output column containing y coordinates for each node.}
}
\value{
a data frame with three columns.
        The first column contains node names and is named \code{node_name_colname}.
        The second column contains x coordinates for the nodes and is named \code{x_colname}.
        The third column contains y coordinates for the nodes and is named \code{y_colname}.
}
\description{
Industries and products are interleaved from left to right,
starting with the first industry, followed by the first product.
Industries and products alternate until all products are exhausted.
Finally, the last industry (which should be final demand) is placed at the far right.
}
