% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecc_layout.R
\name{ecc_layout}
\alias{ecc_layout}
\title{Layout for a \code{qgraph} network representation of an energy conversion chain}
\usage{
ecc_layout(Industries, Products, industry_colname = "Industry",
  product_colname = "Product", stage_colname = "Stage",
  storage_stagename = "Storage", g, group_colname = "Group",
  node_name_colname = "Node_name", x_colname = "x", y_colname = "y")
}
\arguments{
\item{Industries}{a data frame consisting of columns named
\code{industry_colname}, \code{stage_colname}, and \code{group_colname}.
Entries in \code{industry_colname} must be names of industries.
Entries in \code{stage_colname} must be names of left-to-right stages in the ECC.
Industries (or sectors) in final demand must be included in the \code{Industries}
data frame.
Industries that provide "storage" must also be included in the \code{Industries}
data frame (such as Stock changes, International aviation and marine bunkers, and
Statistical differences), their \code{stage_colname} value being \code{storage_stagename}.
All industries with \code{storage_stagename} are laid out
along the top of the network.
The left-to-right order of stages in the network layout generated by this function
is taken from the order of appearance of stages in the \code{stage_colname} column.
The left-to-right order of storage industries is given by
the appearance order of the industries in \code{industry_colname}.
Entries in \code{group_colname} give groups of industries or products to keep together
at each stage. The top-to-bottom order of groups is taken from the appearance order
in \code{group_colname}.}

\item{Products}{a data frame consisting of same columns as \code{Industries},
except that \code{product_colname} takes the place of \code{industry_colname}.
Note that the number of stages in \code{Products} must be one less than than
number of stages in \code{Industries}, not counting storage industries.}

\item{industry_colname}{the name of the column in \code{Industries} containing
names of industries (a string).
Default is "\code{Industry}".}

\item{product_colname}{the name of the column in \code{Products} containing
names of products (a string).
Default is "\code{Product}".}

\item{stage_colname}{the name of the columns in \code{Industries} and \code{Products}
containing names of stages (a string).
The left-to-right order of stages in the network layout
is determined by the top-to-bottom
order in which stage names appear in the
\code{Industries} and \code{Products} data frames.
Default is "\code{Stage}".}

\item{storage_stagename}{the name of the stage in \code{stage_colname} of \code{Industries}
that identifies a "storage" industry (a string).
Default is "\code{Storage}".
Typical storage industries are bunkers, stock changes, and statistical differences.
This layout puts storage industries across the top of the ECC graph.}

\item{g}{an optional \code{qgraph} object from which the order of node names is extracted,
if present.
It is assumed that the layout calculated by this function will be applied to \code{g},
so the row order of the outgoing layout is set to the row order expected by \code{g}.
Furthermore, additional error checking is performed:
if the names of nodes in \code{Industries} and \code{Products}
don't exactly match the names of nodes that \code{g} expects,
a data frame of mismatching names is printed and execution is halted.
An easy way to obtain a \code{qgraph} object for this argument is to call
\code{g <- qgraph(somedata, DoNotPlot = TRUE)}.}

\item{group_colname}{the name of the column in \code{Industries} and \code{Products}
containing industries and products that should be grouped together vertically
at a stage (a string).
The top-to-bottom order of groups at a stage is give by the order of appearance
in \code{group_colname}.
Default is "\code{Group}".
\code{group_colname} is optional in \code{Industries} and \code{Products}.
If \code{group_colname} is missing, a column named "\code{group_colname}"
will be created and filled with "\code{group_colname}".}

\item{node_name_colname}{the name of the output column containing names of nodes
representing industries and products (a string).
Node names are taken from the \code{industry_colname} of \code{Industries}
and the \code{product_colname} of \code{Products}.
Default is "\code{Node_name}".}

\item{x_colname}{the name of the output column containing x coordinates for each node
(a string).
Default is "\code{x}".}

\item{y_colname}{the name of the output column containing y coordinates for each node
(a string).
Default is "\code{y}".}
}
\value{
a matrix with two columns.
        The first column contains x coordinates for the nodes and is named \code{x_colname}.
        The second column contains y coordinates for the nodes and is named \code{y_colname}.
        Node names are the row names for the outgoing matrix.
}
\description{
Industries and products are interleaved from left to right,
starting with the first industry, followed by the first product.
Industries and products alternate until all products are exhausted.
Finally, the last industry (which should be final demand) is placed at the far right.
Storage industries are arranged across the top of the network.
}
