% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{edge_list}
\alias{edge_list}
\title{Title}
\usage{
edge_list(.sutdata = NULL, U = "U", V = "V", Y = "Y",
  edge_list = "edge_list", from = "From", to = "To", value = "Value",
  product = "Product", rejected_energy = "Rejected Energy",
  simplify_edges = TRUE)
}
\arguments{
\item{simplify_edges}{}
}
\description{
Title
}
