% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{efficiency_cols}
\alias{efficiency_cols}
\title{Efficiency data frame column names}
\format{
A string list with 14 entries.
\describe{
\item{eta_i}{The name of a column in a wide-by-matrices data frame containing industry efficiencies.}
\item{eta_pfd_gross}{The name of a column in a wide-by-matrices data frame containing  efficiencies between the primary and gross final demand stages.}
\item{eta_pfd_net}{The name of a column in a wide-by-matrices data frame containing efficiencies between the primary and net final demand stages.}
\item{eta_pf}{The string name for primary-to-final efficiency.}
\item{eta_fu}{The string name for final-to-useful efficiency.}
\item{eta_pu}{The string name for primary-to-useful efficiency.}
\item{eta_ps}{The string name for primary-to-services efficiency.}
\item{eta_us}{The string name for useful-to-services efficiency.}
\item{eta_pw}{The string name for primary-to-well-being efficiency.}
\item{eta_sw}{The string name for services-to-well-being efficiency.}
\item{efficiency_name_suffix}{The suffix for names of columns containing efficiency names.}
\item{gross_net}{The name of a column that contains "Gross" or "Net" for the type of efficiency.}
\item{gross}{The entry in the \code{gross_net} column that identifies a gross efficiency.}
\item{net}{The entry in the \code{gross_net} column that identifies a net efficiency.}
}
}
\usage{
efficiency_cols
}
\description{
A string list containing named names of columns in efficiency data frames.
Items in the list provide default values for column name function arguments
to efficiency functions throughout the \code{Recca} package.
}
\examples{
efficiency_cols
}
\keyword{datasets}
