% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy_balance.R
\name{endogenize_losses}
\alias{endogenize_losses}
\title{Endogenize losses into PSUT matrices}
\usage{
endogenize_losses(
  .sutmats = NULL,
  loss_product = Recca::balance_cols$waste_heat,
  loss_sector = Recca::balance_cols$losses_sector,
  V = "V",
  Y = "Y",
  balance_colname = Recca::balance_cols$intra_industry_balance_colname,
  V_prime = "V_prime",
  Y_prime = "Y_prime"
)
}
\arguments{
\item{.sutmats}{A \code{matsindf} data frame, wide by matrices.}

\item{loss_product}{The string name of the loss product.
Default is \link{balance_cols}\verb{$waste_heat} or
"Waste heat".}

\item{loss_sector}{The string name of the sector
that will absorb losses.
Default is \link{balance_cols}\verb{$losses_sector} or
"Losses".}

\item{V}{Make (\strong{V}) matrix or name of the column in \code{.sutmats}
that contains same.
Default is "V".}

\item{Y}{Final demand (\strong{Y}) matrix or name
of the column in \code{.sutmats} that contains same.
Default is "Y".}

\item{balance_colname}{The name of the column containing
energy balance vectors.
Default is
\link{balance_cols}\verb{$intra_industry_balance_colname} or
"SUTIntraIndustryBalance".}

\item{Y_prime}{}
}
\value{
A version of \code{.sutmats} with losses endogenized.
}
\description{
When a conversion chain does \emph{not} include losses in the
\strong{RUVY} matrices of the PSUT framework,
it may be helpful to endogenize the losses.
This function performs the endogenization.
}
\details{
Intra-industry balances need to be calculated
(most easily vai \code{\link[=calc_intra_industry_balance]{calc_intra_industry_balance()}})
prior to calling this function.
}
