% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy_balance.R
\name{endogenize_losses}
\alias{endogenize_losses}
\title{Endogenize losses into PSUT matrices}
\usage{
endogenize_losses(
  .sutmats = NULL,
  loss_product = "Waste heat",
  loss_sector = "Losses",
  V = "V",
  Y = "Y",
  balance_colname = Recca::balance_cols$intra_industry_balance_colname
)
}
\arguments{
\item{.sutmats}{A \code{matsindf} data frame, wide by matrices.}

\item{loss_product}{The string name of the product for losses.
Default is "Waste heat".}

\item{loss_sector}{The string name of the sector that will absorb losses.
Default is "Losses".}

\item{V}{Make (\strong{V}) matrix or name of the column in \code{.sutmats}
that contains same.
Default is "V".}

\item{Y}{Final demand (\strong{Y}) matrix or name
of the column in \code{.sutmats} that contains same.
Default is "Y".}

\item{balance_colname}{The name of the column containing
energy balance vectors.
Default is
\link{balance_cols}\verb{$intra_industry_balance_colname} or
"SUTIntraIndustryBalance".}
}
\value{
A version of \code{.sutmats} with losses endogenized.
}
\description{
When a conversion chain does \emph{not} include losses in the
\strong{RUVY} matrices of the PSUT framework,
it may be helpful to endogenize the losses.
This function performs the endogenization.
}
