% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy_balance.R
\name{endogenize_losses}
\alias{endogenize_losses}
\title{Endogenize losses into PSUT matrices}
\usage{
endogenize_losses(
  .sutmats = NULL,
  V = "V",
  Y = "Y",
  loss_product = Recca::balance_cols$waste_heat,
  loss_sector = Recca::balance_cols$losses_sector,
  replace_cols = FALSE,
  balance_colname = Recca::balance_cols$intra_industry_balance_colname,
  V_prime = "V_prime",
  Y_prime = "Y_prime"
)
}
\arguments{
\item{.sutmats}{A \code{matsindf} data frame, wide by matrices.}

\item{V}{Make (\strong{V}) matrix or name of the column in \code{.sutmats}
that contains same.
Default is "V".}

\item{Y}{Final demand (\strong{Y}) matrix or name
of the column in \code{.sutmats} that contains same.
Default is "Y".}

\item{loss_product}{The string name of the loss product.
Default is \link{balance_cols}\verb{$waste_heat} or
"Waste heat".}

\item{loss_sector}{The string name of the sector
that will absorb losses.
Default is \link{balance_cols}\verb{$losses_sector} or
"Losses".}

\item{replace_cols}{A boolean that tells whether to
(a) replace
the \code{V} and \code{Y} columns with
\code{V_prime} and \code{Y_prime} columns, respectively and
(b) delete the \code{V_prime} and \code{Y_prime}
columns
after endogenizing the losses
when \code{.sutmats} is a data frame or a list.
Default is \code{FALSE}.}

\item{balance_colname}{The name of the column containing
energy balance vectors.
Default is
\link{balance_cols}\verb{$intra_industry_balance_colname} or
"SUTIntraIndustryBalance".}

\item{V_prime}{The name of the \strong{V} matrix with endogenized losses.}

\item{Y_prime}{The name of the \strong{Y} matrix with endogenized losses.}
}
\value{
A version of \code{.sutmats} with losses endogenized.
}
\description{
When a conversion chain does \emph{not} include losses in the
\strong{RUVY} matrices of the PSUT framework,
it may be helpful to endogenize the losses.
This function performs the endogenization.
}
\details{
Intra-industry balances need to be calculated
(most easily via \code{\link[=calc_intra_industry_balance]{calc_intra_industry_balance()}})
prior to calling this function.
See the example.
}
\examples{
mats <- UKEnergy2000mats |>
  tidyr::pivot_wider(names_from = matrix.name,
                     values_from = matrix) |>
  dplyr::filter(.data[[IEATools::iea_cols$last_stage]] \%in\%
                  c(IEATools::last_stages$final, IEATools::last_stages$useful)) |>
  calc_intra_industry_balance()
mats |>
  endogenize_losses() |>
  dplyr::glimpse()
  # Replace original matrices with endogenized matrices=
mats |>
  endogenize_losses(replace_cols = TRUE) |>
  dplyr::glimpse()
}
