% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy_to_exergy.R
\name{extend_to_exergy}
\alias{extend_to_exergy}
\title{Extend an ECC in PSUT format from energy to exergy}
\usage{
extend_to_exergy(
  .sutmats = NULL,
  clean_up_df = TRUE,
  endogenize_losses_irrev = FALSE,
  R = Recca::psut_cols$R,
  U = Recca::psut_cols$U,
  V = Recca::psut_cols$V,
  Y = Recca::psut_cols$Y,
  U_feed = Recca::psut_cols$U_feed,
  U_eiou = Recca::psut_cols$U_eiou,
  r_eiou = Recca::psut_cols$r_eiou,
  intra_industry_balance = Recca::balance_cols$intra_industry_balance_colname,
  losses_alloc = Recca::balance_cols$losses_alloc_colname,
  losses_sector = Recca::balance_cols$losses_sector,
  phi = Recca::psut_cols$phi,
  irrev_alloc = Recca::balance_cols$irrev_alloc_colname,
  irrev_sector = Recca::balance_cols$irrev_sector,
  clean = FALSE,
  mat_piece = "all",
  phi_piece = "all",
  notation = RCLabels::bracket_notation,
  prepositions = RCLabels::prepositions_list,
  energy_type = Recca::psut_cols$energy_type,
  exergy = Recca::energy_types$x,
  tol = 1e-06,
  .exergy_suffix = "_exergy",
  R_exergy = paste0(Recca::psut_cols$R, .exergy_suffix),
  U_exergy = paste0(Recca::psut_cols$U, .exergy_suffix),
  V_exergy = paste0(Recca::psut_cols$V, .exergy_suffix),
  Y_exergy = paste0(Recca::psut_cols$Y, .exergy_suffix),
  U_feed_exergy = paste0(Recca::psut_cols$U_feed, .exergy_suffix),
  U_eiou_exergy = paste0(Recca::psut_cols$U_eiou, .exergy_suffix),
  r_eiou_exergy = paste0(Recca::psut_cols$r_eiou, .exergy_suffix)
)
}
\arguments{
\item{.sutmats}{An optional data frame of mass or energy
(conserved quantities)
conversion chains in wide-by-matrix format.}

\item{clean_up_df}{When \code{.sutmats} is a data frame, tells whether to \code{tidyr::pivot_longer()} the result,
remove no-longer-needed input column \code{phi}, and
fill the \code{energy_type} column with "X" for the exergy versions of the ECC matrices.
Default is \code{TRUE}.}

\item{endogenize_losses_irrev}{A boolean that tells whether to endogenize
losses of the conserved quantity and
calculate irreversibility
(exergy destruction) of the industries
in the exergy version of the
conversion chain.
Default is \code{FALSE}.}

\item{R}{Resources (\strong{R}) matrix or name of the column in \code{.sutmats}
that contains same.
Default is \link{psut_cols}\verb{$R} or
"R".}

\item{U}{Use (\strong{U}) matrix or name of the column in \code{.sutmats}
that contains same.
Necessary for verifying calculating losses.
Default is \link{psut_cols}\verb{$U} or
"U".}

\item{V}{Make (\strong{V}) matrix or name of the column in \code{.sutmats}
that contains same.
Default is \link{psut_cols}\verb{$V} or
"V".}

\item{Y}{Final demand (\strong{Y}) matrix or name
of the column in \code{.sutmats} that contains same.
Default is \link{psut_cols}\verb{$Y} or
"Y".}

\item{U_feed}{Feedstock use (\strong{U_feed}) matrix or name of the column in \code{.sutmats}
that contains same.
Default is \link{psut_cols}\verb{$U_feed} or
"U_feed".}

\item{U_eiou}{Energy industry own use (\strong{U_EIOU}) matrix or name of the column in \code{.sutmats}
that contains same.
Default is \link{psut_cols}\verb{$U_eiou} or
"U_EIOU".}

\item{r_eiou}{A matrix of the ratio of energy industry own use
to total use for the use (\strong{U}) matrix
or name of the column in \code{.sutmats} that contains same.
Default is \link{psut_cols}\verb{$r_eiou} or
"r_EIOU".}

\item{intra_industry_balance}{A vector or the name of the column containing
intra-industry balance vectors for the conserved quantity.
If missing, losses are calculated internally
with \code{\link[=calc_intra_industry_balance]{calc_intra_industry_balance()}}
before endogenizing.
Required only if \code{endogenize_losses_irrev} is \code{TRUE}.
Default is
\link{balance_cols}\verb{$intra_industry_balance_colname} or
"SUTIntraIndustryBalance".}

\item{losses_alloc}{A matrix or the name of the column containing
loss allocation matrices.
See details for structure of this matrix.
Required only if \code{endogenize_losses_irrev} is \code{TRUE}.
Default is \link{balance_cols}\verb{$losses_alloc_colname} or
"LossesAlloc".}

\item{losses_sector}{The string name of the sector
that will absorb losses in the \strong{Y} matrix for the conserved quantity.
Required only if \code{endogenize_losses_irrev} is \code{TRUE}.
Default is \link{balance_cols}\verb{$losses_sector}
or "Transformation losses".}

\item{phi}{A vector of exergy-to-energy ratios (phi)
or the string name of a column in \code{.sutmats} containing same.
Default is \link{psut_cols}\verb{$phi} or
"phi".}

\item{irrev_alloc}{An irreversibility allocation matrix or the name of a column in
\code{.sutmats} containing same.
Required only if \code{endogenize_losses_irrev} is \code{TRUE}.
Default is \link{balance_cols}\verb{$irrev_alloc_colname} or
"IrrevAlloc".}

\item{irrev_sector}{The string name of the sector that will absorb irreversibilities
(exergy destruction)
in the \strong{Y} matrix.
Required only if \code{endogenize_losses_irrev} is \code{TRUE}.
Default is \link{balance_cols}\verb{$irrev_sector}
or "Irreversibilities".}

\item{clean}{A boolean that tells whether the outgoing
\strong{V} and \strong{Y} matrices with endogenized losses and
irreversibilities should have
\code{0} rows and columns removed.
Affects results only when \code{endogenize_irrev_losses} is \code{TRUE}.
Default is \code{FALSE}.}

\item{mat_piece}{The piece of row and column names for matrices
\code{R}, \code{U}, \code{V}, \code{Y}, \code{U_feed}, and \code{U_EIOU}
against which row names of the \code{phi} vector is to be matched.
Default is "all", meaning that entire names are to be matched.}

\item{phi_piece}{The piece of row names in the \code{phi} vector against which
row and column names for matrices
\code{R}, \code{U}, \code{V}, \code{Y}, \code{U_feed}, and \code{U_EIOU}
is to be matched.
Default is "all", meaning that entire names are to be matched.}

\item{notation}{The nomenclature for the row and column labels.
Default is \code{RCLabels::bracket_notation}.}

\item{prepositions}{The prepositions to be used row and column notation.
Default is \code{RCLabels::prepositions_list}.}

\item{exergy}{A string type to be given for the results.
See \link{energy_types}.
Default is \link{energy_types}\verb{$x} or
"X".}

\item{tol}{The maximum allowable difference from \code{1} for the rowsums of
loss allocation matrices.
Also the maximum allowable different from \code{0} for the
inter-industry balances and intra-industry balances.
Default is \code{1e-6}.}

\item{.exergy_suffix}{The string suffix to be appended to exergy versions of ECC matrices.}

\item{R_exergy, U_exergy, U_feed_exergy, U_eiou_exergy, r_eiou_exergy, V_exergy, Y_exergy, energy_type}{Names of output matrices}
}
\value{
A data frame or list of matrices that represents the exergy version of the ECC.
}
\description{
A conversion chain can be represented in energy or mass terms (conserved quantities)
or in exergy terms (a non-conserved quantity).
This function moves from an energy or mass quantification to an exergy quantification,
given the matrices for the mass/energy quantification and
phi (exergy-to-energy ratio) vectors.
}
\details{
Internally, this function uses \code{matsindf::apply()}, so
the ECC matrices can be provided
as individual matrices,
in a named list (passed in \code{.sutmats}), or
or in a data frame (also passed to \code{.sutmats}),
in which case the arguments should
give the string names in the list or
of columns in the \code{.sutmats} data frame.
The default is strings for most arguments,
thereby assuming \code{.sutmats} will be supplied
with a list or data frame.

The vector \code{phi} is considered to be a store of values
to be applied to each type of energy carrier.
To determine which entry in the \code{phi} vector  is matched against which energy carrier,
\code{mat_piece} and \code{phi_piece} are consulted.
\code{mat_piece} and \code{phi_piece} can be any of
"all", "pref", "suff", "noun", or one of many prepositions.
Consult the \code{RCLabels} package for details.

The inputs to this function should all be conserved quantities
that can be converted to exergy via \code{phi} vectors.
}
\examples{
sutmats <- UKEnergy2000mats |>
  # Put in wide-by-matrix format.
  tidyr::spread(key = matrix.name, value = matrix) |>
  # Eliminate services ECCs.
  dplyr::filter(LastStage \%in\% c("Final", "Useful"))  |>
  dplyr::mutate(
    phi = RCLabels::make_list(Recca::phi_vec, n = dplyr::n(), lenx = 1)
  )
extend_to_exergy(sutmats)
}
