% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregates.R
\name{finaldemand_aggregates}
\alias{finaldemand_aggregates}
\title{Final demand aggregate energy}
\usage{
finaldemand_aggregates(.sutdata, fd_sectors, U_colname = "U",
  V_colname = "V", Y_colname = "Y", r_EIOU_colname = "r_EIOU",
  by = c("Total", "Product", "Sector"), keep_cols = NULL,
  net_aggregate_demand_colname, gross_aggregate_demand_colname)
}
\arguments{
\item{.sutdata}{a data frame with columns of matrices from a supply-use analysis.}

\item{fd_sectors}{a vector of names of sectors in final demand.}

\item{U_colname}{the name of the column in \code{.sutdata} containing Use (\code{U}) matrices.}

\item{V_colname}{the name of the column in \code{.sutdata} containing Make (\code{V}) matrices.}

\item{Y_colname}{the name of the column in \code{.sutdata} containing final demand (\code{Y}) matrices.}

\item{r_EIOU_colname}{the name of the colum that holds ratios of EIOU to total input for each Machine and Product.}

\item{by}{one of "Product", "Sector", or "Total" to indicate the desired aggregation:
"Product" for aggregation by energy carrier (Crude oil, Primary solid biofuels, etc.),
"Sector" for aggregation by final demand sector (Agriculture/forestry, Domestic navigation, etc.), or
"Total" for aggregation over both Product and Sector (the default).}

\item{keep_cols}{a vector of names of columns of \code{.sutdata} to return with the output}

\item{net_aggregate_demand_colname}{the name of the output column containing aggregates of net energy demand.
Each entry in this column is \code{sumall(Y_fd)}.}

\item{gross_aggregate_demand_colname}{the name of the output column containing aggregates of gross energy demand.
Each entry in this column is calculated by \code{sumall(Y_fd)} + \code{sumall(U_EIOU)}.}
}
\value{
a data frame containing the columns in \code{keep_cols}, gross aggregate energy demand
and net aggregate energy demand for each row of \code{.sutdata}.
}
\description{
Calculates aggregate final demand energy from a data frame of Supply-Use matrices.
The calculation includes non-energy uses if they are present in the final demand matrix.
The calculation does not include balancing items (Losses and Statistical differences).
}
