% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregates.R
\name{finaldemand_aggregates}
\alias{finaldemand_aggregates}
\title{Final demand aggregate energy}
\usage{
finaldemand_aggregates(
  .sutdata,
  fd_sectors,
  U = "U",
  Y = "Y",
  r_EIOU = "r_EIOU",
  by = c("Total", "Product", "Sector"),
  net_aggregate_demand = "EX_fd_net.ktoe",
  gross_aggregate_demand = "EX_fd_gross.ktoe"
)
}
\arguments{
\item{.sutdata}{a data frame with columns of matrices from a supply-use analysis.}

\item{fd_sectors}{a vector of names of sectors in final demand.}

\item{U}{use (\code{U}) matrix or name of the column in \code{.sutdata} containing same}

\item{Y}{final demand (\code{Y}) matrix or name of the column in \code{.sutdata} containing same}

\item{r_EIOU}{matrix of ratios of EIOU for the make (\code{U}) matrix or name of the column in \code{.sutdata} containing same}

\item{by}{one of "Product", "Sector", or "Total" to indicate the desired aggregation:
"Product" for aggregation by energy carrier (Crude oil, Primary solid biofuels, etc.),
"Sector" for aggregation by final demand sector (Agriculture/forestry, Domestic navigation, etc.), or
"Total" for aggregation over both Product and Sector (the default).}

\item{net_aggregate_demand}{the name of net energy demand on output.
Each entry is \code{sumall(Y_fd)}.}

\item{gross_aggregate_demand}{the name of gross energy demand on output.
Each entry is calculated by \code{sumall(Y_fd)} + \code{sumall(U_EIOU)}.}
}
\value{
a list or data frame containing \code{net_aggregate_demand} and \code{gross_aggregate_demand}
}
\description{
Calculates aggregate final demand energy from a data frame of Supply-Use matrices.
The calculation includes non-energy uses if they are present in the final demand matrix.
The calculation does not include balancing items (Losses and Statistical differences).
}
