% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregates.R
\name{finaldemand_aggregates}
\alias{finaldemand_aggregates}
\title{Final demand aggregate energy}
\usage{
finaldemand_aggregates(
  .sutdata,
  fd_sectors,
  U = Recca::psut_cols$U,
  Y = Recca::psut_cols$Y,
  r_EIOU = Recca::psut_cols$r_eiou,
  by = c("Total", "Product", "Sector"),
  net_aggregate_demand = Recca::aggregate_cols$net_aggregate_demand,
  gross_aggregate_demand = Recca::aggregate_cols$gross_aggregate_demand
)
}
\arguments{
\item{.sutdata}{A data frame with columns of matrices from a supply-use analysis.}

\item{fd_sectors}{A vector of names of sectors in final demand.}

\item{U, Y, r_EIOU}{See \code{Recca::psut_cols}.}

\item{by}{One of "Product", "Sector", or "Total" to indicate the desired aggregation:
"Product" for aggregation by energy carrier (Crude oil, Primary solid biofuels, etc.),
"Sector" for aggregation by final demand sector (Agriculture/forestry, Domestic navigation, etc.), or
"Total" for aggregation over both Product and Sector (the default).}

\item{net_aggregate_demand, gross_aggregate_demand}{See \code{Recca::aggregate_cols}.
Net energy demand is calculated by \code{sumall(Y_fd)}.
Gross energy demand is calculated by \code{sumall(Y_fd) + sumall(U_EIOU)}.}
}
\value{
A list or data frame containing \code{net_aggregate_demand} and \code{gross_aggregate_demand} columns.
}
\description{
Calculates aggregate final demand energy from a data frame of Supply-Use matrices.
The calculation includes non-energy uses if they are present in the final demand matrix.
The calculation does not include balancing items (Losses and Statistical differences).
If \code{.sutdata} is a data frame, \code{fd_sectors} should be the name of a column in the data frame.
If \code{.sutdata} is \code{NULL}, \code{fd_sectors} can be a single vector of industry names.
}
\examples{
library(matsbyname)
UKEnergy2000mats \%>\%
  tidyr::pivot_wider(names_from = "matrix.name", values_from = "matrix") \%>\%
  dplyr::mutate(
    fd_sectors = rep(list(c("Residential", "Transport")), times = nrow(.))
  ) \%>\%
  dplyr::filter(Last.stage \%in\% c(IEATools::last_stages$final,
                                  IEATools::last_stages$useful)) \%>\%
  finaldemand_aggregates(fd_sectors = "fd_sectors", by = "Sector")
}
