% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregates.R
\name{finaldemand_aggregates_with_units}
\alias{finaldemand_aggregates_with_units}
\title{Final demand aggregate energy with units}
\usage{
finaldemand_aggregates_with_units(.sutdata, fd_sectors, U = "U", Y = "Y",
  r_EIOU = "r_EIOU", S_units = "S_units", by = c("Total", "Product",
  "Sector"), net_aggregate_demand_colname, gross_aggregate_demand_colname)
}
\arguments{
\item{.sutdata}{a data frame with columns of matrices from a supply-use analysis.}

\item{fd_sectors}{a vector of names of sectors in final demand.}

\item{U}{the name of the column in \code{.sutdata} containing Use (\code{U}) matrices.}

\item{Y}{the name of the column in \code{.sutdata} containing final demand (\code{Y}) matrices.}

\item{r_EIOU}{the name of the column that holds ratios of EIOU to total input for each Machine and Product.}

\item{S_units}{the name of the column in \code{.sutdata} containing \code{S_units} matrices.}

\item{by}{one of "Product", "Sector", or "Total" to indicate the desired aggregation:
"Product" for aggregation by energy carrier (Crude oil, Primary solid biofuels, etc.),
"Sector" for aggregation by final demand sector (Agriculture/forestry, Domestic navigation, etc.), or
"Total" for aggregation over both Product and Sector (the default).}

\item{gross_aggregate_demand_colname}{the name of the output column containing aggregates of gross energy demand.
This column includes energy industry own use.}

\item{net_aggregate_demand}{the name of the output column containing aggregates of net energy demand.
This column excludes energy industry own use.}
}
\value{
a data frame containing net aggregate energy demand
and gross aggregate energy demand for each row of \code{.sutdata}.
}
\description{
Calculates aggregate final demand energy and services from a data frame of Supply-Use matrices.
The calculation includes non-energy uses if they are present in the final demand matrix.
The calculation does not include balancing items (Losses and Statistical differences).
}
