% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregates.R
\name{finaldemand_aggregates_with_units}
\alias{finaldemand_aggregates_with_units}
\title{Final demand aggregate energy with units}
\usage{
finaldemand_aggregates_with_units(.sutdata, fd_sectors, U = "U",
  Y = "Y", r_EIOU = "r_EIOU", S_units = "S_units", by = c("Total",
  "Product", "Sector"), net_aggregate_demand, gross_aggregate_demand)
}
\arguments{
\item{.sutdata}{a data frame with columns of matrices from a supply-use analysis.}

\item{fd_sectors}{a vector of names of sectors in final demand.}

\item{U}{use (\code{U}) matrix or name of the column in \code{.sutdata} containing same}

\item{Y}{final demand (\code{Y}) matrix or name of the column in \code{.sutdata} containing same}

\item{r_EIOU}{matrix of ratios of EIOU for the make (\code{U}) matrix or name of the column in \code{.sutdata} containing same}

\item{S_units}{the name of the column in \code{.sutdata} containing \code{S_units} matrices.}

\item{by}{one of "Product", "Sector", or "Total" to indicate the desired aggregation:
"Product" for aggregation by energy carrier (Crude oil, Primary solid biofuels, etc.),
"Sector" for aggregation by final demand sector (Agriculture/forestry, Domestic navigation, etc.), or
"Total" for aggregation over both Product and Sector (the default).}

\item{net_aggregate_demand}{the name of net energy demand (which excludes energy industry own use) on output.}

\item{gross_aggregate_demand}{the name of gross energy demand (which includes energy industry own use) on output.}
}
\value{
a list or data frame containing net aggregate energy demand
and gross aggregate energy demand
}
\description{
Calculates aggregate final demand energy and services from a data frame of Supply-Use matrices.
The calculation includes non-energy uses if they are present in the final demand matrix.
The calculation does not include balancing items (Losses and Statistical differences).
}
