% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{flows_unit_homogeneous}
\alias{flows_unit_homogeneous}
\title{Tell whether industry flows (inputs and outputs) are unit-homogeneous}
\usage{
flows_unit_homogeneous(
  .sutmats = NULL,
  U = "U",
  V = "V",
  S_units = "S_units",
  keep_details = FALSE,
  flows_unit_homogeneous = ".flows_unit_homogeneous"
)
}
\arguments{
\item{.sutmats}{a data frame of supply-use table matrices with matrices arranged in columns.}

\item{U}{a use (\code{U}) matrix or name of the column in \code{.sutmats} that contains same. Default is "\code{U}".}

\item{V}{a make (\code{V}) matrix or name of the column in \code{.sutmats} that contains same. Default is "\code{V}".}

\item{S_units}{an \code{S_units} matrix or name of a column in \code{.sutmats} that contains same. Default is "\code{S_units}".}

\item{keep_details}{if \code{TRUE}, per-industry results are returned;
if \code{FALSE}, per-ECC results are returned.
Default is "\code{FALSE}".}

\item{flows_unit_homogeneous}{the name of the output column
that tells whether each industry's outputs are unit-homogeneous.
Default is "\code{.flows_unit_homogeneous}".}
}
\value{
\code{.sutdata} with additional column "\code{flows_unit_homogeneous}" containing
\code{TRUE} if each industry's flows are unit-homogeneous,
\code{FALSE} if each industry's flows are unit-heterogeneous.
}
\description{
Returns \code{TRUE} if each industry's flows (all inputs and outputs) are unit-homogeneous.
}
\details{
The \code{V_bar} matrix is queried for the number of non-zero entries in each row.
If the number of non-zero entries in each row is exactly 1,
industry outputs are unit-homogeneous.
Note that \code{V_bar = \link[matsbyname]{matrixproduct_byname}(V, S_units)}.
}
\examples{
library(tidyr)
UKEnergy2000mats \%>\%
  spread(key = "matrix.name", value = "matrix") \%>\%
  flows_unit_homogeneous()
}
