% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_all_products_and_industries}
\alias{get_all_products_and_industries}
\title{Create lists of all products and industries}
\usage{
get_all_products_and_industries(
  .sutdata,
  R = IEATools::psut_cols$R,
  U = IEATools::psut_cols$U,
  V = IEATools::psut_cols$V,
  Y = IEATools::psut_cols$Y,
  products_col = Recca::prod_ind_names_colnames$product_names,
  industries_col = Recca::prod_ind_names_colnames$industry_names
)
}
\arguments{
\item{.sutdata}{A data frame or list of \code{matsindf} matrices.}

\item{R, U, V, Y}{The names of PSUT matrices. See \code{IEAtools::psut_cols}.}

\item{products_col}{The name of the products column in the output list or data frame.
Default is \code{Recca::prod_ind_names_colnames$product_names}.}

\item{industries_col}{The name of the products column in the output list or data frame.
Default is \code{Recca::prod_ind_names_colnames$industry_names}.}
}
\value{
\code{.sutdata} with two new columns containing the names of products and industries.
}
\description{
From matrices that describe an energy conversion chain (\strong{R}, \strong{U}, \strong{V}, and \strong{Y}),
create a list of unique products (energy carriers) and industries (processing stages)
in the energy conversion chain.
}
\details{
This function is a \code{matsindf::matsindf_apply()} style function.
It can accept a \code{matsindf} data frame in the \code{.df} argument
and string for the \code{R}, \code{U}, \code{V}, and \code{Y} (as column names) in the arguments.
}
