% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{inputs_unit_homogeneous}
\alias{inputs_unit_homogeneous}
\title{Tell whether each industry's inputs are unit-homogeneous}
\usage{
inputs_unit_homogeneous(
  .sutmats = NULL,
  U = "U",
  S_units = "S_units",
  keep_details = FALSE,
  ins_unit_homogeneous = ".inputs_unit_homogeneous"
)
}
\arguments{
\item{.sutmats}{a data frame of supply-use table matrices with matrices arranged in columns.}

\item{U}{a use (\code{U}) matrix or name of the column in \code{.sutmats} that contains same. Default is "\code{U}".}

\item{S_units}{an \code{S_units} matrix or name of a column in \code{.sutmats} that contains same. Default is "\code{S_units}".}

\item{keep_details}{if \code{TRUE}, per-product results are returned;
if \code{FALSE}, per-ECC results are returned. Default is \code{FALSE}.}

\item{ins_unit_homogeneous}{name of the output boolean that
tells whether each industry's inputs are unit-homogeneous.
Default is "\code{.inputs_unit_homogeneous}".}
}
\value{
a list or data frame containing
\code{TRUE} if inputs to each energy conversion industry are unit-homogeneous,
\code{FALSE} otherwise.
}
\description{
Returns \code{TRUE} if each industry's inputs are unit-homogeneous.
}
\details{
The \code{U_bar} matrix is queried for the number of non-zero entries in each column.
If the number of non-zero entries in each column is exactly 1,
industry inputs are unit-homogeneous.
Note that \code{U_bar = \link[matsbyname]{matrixproduct_byname}(\link[matsbyname]{transpose_byname}(S_units), U)}.
}
\examples{
library(tidyr)
UKEnergy2000mats \%>\%
  spread(key = "matrix.name", value = "matrix") \%>\%
  inputs_unit_homogeneous()
}
