% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sankey.R
\name{make_sankey}
\alias{make_sankey}
\title{Make a Sankey diagram}
\usage{
make_sankey(
  .sutmats = NULL,
  R = Recca::psut_cols$R,
  U = Recca::psut_cols$U,
  V = Recca::psut_cols$V,
  Y = Recca::psut_cols$Y,
  simplify_edges = TRUE,
  sankey = Recca::sankey_cols$sankey,
  ...
)
}
\arguments{
\item{.sutmats}{an optional wide-by-matrices data frame}

\item{R, U, V, Y}{See \code{Recca::psut_cols}.}

\item{simplify_edges}{a boolean which tells whether edges should be simplified.
Applies to every row of \code{.sutmats} if \code{.sutmats} is specified.}

\item{sankey}{See \code{Recca::sankey_cols}.}

\item{...}{Arguments passed to \code{networkD3::sankeyNetwork()}, mostly for formatting purposes.}
}
\value{
a Sankey diagram
}
\description{
A Sankey diagram is a flow diagram in which the width of the lines is proportional
to the rate of energy flow.
Sankey diagrams are a helpful way to visualize energy flows in an energy conversion chain (ECC).
This function takes a matrix description of an ECC and produces a Sankey diagram.
}
\details{
At present, this function uses the \code{networkD3} package to draw the Sankey diagram.

If any of \code{R}, \code{U}, \code{V}, or \code{Y} is \code{NA}, \code{NA} is returned.
}
\examples{
library(dplyr)
library(magrittr)
library(networkD3)
library(tidyr)
UKEnergy2000mats \%>\%
  spread(key = "matrix.name", value = "matrix") \%>\%
  make_sankey() \%>\%
  extract2("Sankey") \%>\%
  extract2(1)
}
