% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sankey.R
\name{make_sankey}
\alias{make_sankey}
\title{Make a Sankey diagram}
\usage{
make_sankey(
  .sutmats = NULL,
  R = "R",
  U = "U",
  V = "V",
  Y = "Y",
  simplify_edges = TRUE,
  sankey = "Sankey"
)
}
\arguments{
\item{.sutmats}{an optional data frame}

\item{R}{a resource matrix or the name of the column in \code{.sutmats} containing \code{R} matrices}

\item{U}{a use matrix or the name of the column in \code{.sutmats} containing \code{U} matrices}

\item{V}{a make matrix or the name of the column in \code{.sutmats} containing \code{V} matrices}

\item{Y}{a final demand matrix or the name of the column in \code{.sutmats} containing \code{Y} matrices}

\item{simplify_edges}{a boolean which tells whether edges should be simplified.
Applies to every row of \code{.sutmats} if \code{.sutmats} is specified.}

\item{sankey}{the name of the output Sankey diagram or the name of the column in \code{.sutmats} containing Sankey diagrams}
}
\value{
a Sankey diagram
}
\description{
A Sankey diagram is a flow diagram in which the width of the lines is proportional
to the rate of energy flow.
Sankey diagrams are a helpful way to visualize energy flows in an energy conversion chain (ECC).
This function takes a matrix description of an ECC and produces a Sankey diagram.
}
\details{
At present, this function uses the `networkD3` package to draw the Sankey diagram.

If any of \code{R}, \code{U}, \code{V}, or \code{Y} is \code{NA}, \code{NA} is returned.
}
\examples{
library(dplyr)
library(magrittr)
library(networkD3)
library(tidyr)
UKEnergy2000mats \%>\%
  spread(key = "matrix.name", value = "matrix") \%>\%
  rename(
    R_plus_V = V
  ) \%>\%
  separate_RV() \%>\%
  make_sankey() \%>\%
  extract2("Sankey") \%>\%
  extract2(1)
}
