% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruct.R
\name{new_k_ps}
\alias{new_k_ps}
\title{Assess the effect of changing perfectly substitutable intermediate inputs to an industry}
\usage{
new_k_ps(
  .sutmats = NULL,
  k_prime = "k_prime",
  U = "U",
  V = "V",
  Y = "Y",
  K = "K",
  L_ixp = "L_ixp",
  L_pxp = "L_pxp",
  Z = "Z",
  D = "D",
  f = "f",
  U_prime = "U_prime",
  V_prime = "V_prime"
)
}
\arguments{
\item{.sutmats}{a data frame of supply-use table matrices with matrices arranged in columns.}

\item{k_prime}{a new column vector for the \code{K} matrix representing new
inputs to an industry or name of a column in \code{.sutmats} containing same.
Default is "\code{k_prime}".
The name of the single \code{k_prime} column must match the name of one of the columns of matrix \code{K}.}

\item{U}{use (\code{U}) matrix or name of the column in \code{.sutmats} that contains same. Default is "\code{U}".}

\item{V}{make (\code{V}) matrix or name of the column in \code{.sutmats}that contains same. Default is "\code{V}".}

\item{Y}{final demand (\code{Y}) matrix or name of the column in \code{.sutmats} that contains same. Default is "\code{Y}".}

\item{K}{a \code{K} matrix or name of the column in \code{.sutmats} that contains same. Default is "\code{K}".
\code{K} consists of columns that sum to 1.
Elements of \code{K} indicate the fraction of total input to industries (in columns)
provided by products (in rows).
\code{K} can be calculated by \code{\link{calc_io_mats}}.}

\item{L_ixp}{an (\code{L_ixp}) matrix or name of the column in \code{.sutmats} that contains same. Default is "\code{L_ixp}".}

\item{L_pxp}{an (\code{L_pxp}) matrix or name of the column in \code{.sutmats} that contains same. Default is "\code{L_pxp}".}

\item{Z}{a \code{Z} matrix or name of the column in \code{.sutmats} that contains same. Default is "\code{Z}".}

\item{D}{a \code{D} matrix or name of the column in \code{.sutmats} that contains same. Default is "\code{D}".}

\item{f}{an \code{f} vector or name of the column in \code{.sutmats} that contains same. Default is "\code{f}".}

\item{U_prime}{name for the \code{U_prime} matrix on output. Default is "\code{U_prime}".}

\item{V_prime}{name for the \code{V_prime} matrix on output. Default is "\code{V_prime}".}
}
\value{
a list or data frame containing \code{U_prime} and \code{V_prime} matrices
}
\description{
This function calculates the effect of changing perfectly-substitutable (ps) inputs
to an intermediate industry.
New versions of \code{U} and \code{V} matrices are returned
as \code{U_prime} and \code{V_prime}.
Changes are made upstream of the changed industry inputs.
The final demand matrix (\code{Y}) is unchanged.
}
\details{
Note that inputs \code{K}, \code{L_ixp}, \code{L_pxp},
\code{Z}, \code{D}, and \code{f} can be
conveniently calculated by the function \code{\link{calc_io_mats}}.

Internally, this function uses \code{\link[matsindf]{matsindf_apply}},
and documentation assumes that
\code{.sutmats} is not \code{NULL} and is a data frame.
If \code{.sutmats} is present, output is a data frame with columns named by string values of output arguments, and
input arguments should be character strings that name columns in \code{.sutmats}.
If \code{.sutmats} is \code{NULL} (the default), output is a list with items named by output strings,
and input arguments should be single matrices or vectors.
}
\examples{
library(dplyr)
library(matsbyname)
library(tidyr)
# To demonstrate calculating changes to an energy conversion chain due to changes
# in perfectly-substitutable inputs to an intermediate industry,
# we use the PerfectSubmats data frame.
# But we need to calculate several important input-output matrices first.
io_mats <- PerfectSubmats \%>\%
  spread(key = "matrix.name", value = "matrix") \%>\%
  calc_io_mats()
# Next, find the K matrix that contains the fraction of each type of energy
# that enters each industry
K <- io_mats$K[[1]]
# Develop a new column vector for inputs to the Electric transport sector.
# As provided, the Electric transport sector is dominated by Renewable elec.
# What if the electricity input to the Electric transport sector
# were split 50/50 between Renewable elect and FF elec?
k_prime_vec <- K[, "Electric transport", drop = FALSE]
k_prime_vec["FF elec", "Electric transport"] <- 0.5
k_prime_vec["Ren elec", "Electric transport"] <- 0.5
# Add k_prime_vec to the io_mats data frame.
io_mats <- io_mats \%>\%
  mutate(
    # Set up a new k_prime vector for Electric transport.
    # That vector will be used for the infininte substitution calculation.
    k_prime = select_cols_byname(K, retain_pattern = make_pattern("Electric transport",
                                                                  pattern_type = "exact")),
    k_prime = make_list(k_prime_vec, n = 1)
  )
# Now do the calculation of U_prime and V_prime matrices.
new_UV <- new_k_ps(io_mats)
# There is much more FF extraction now than before.
io_mats$U[[1]]["FF", "FF extraction"]
new_UV$U_prime[[1]]["FF", "FF extraction"]
}
