% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruct.R
\name{new_Y}
\alias{new_Y}
\title{Reconstruct an economy given a new final demand matrix}
\usage{
new_Y(
  .sutmats = NULL,
  Y_prime = "Y_prime",
  L_ixp = "L_ixp",
  L_pxp = "L_pxp",
  Z = "Z",
  Z_feed = "Z_feed",
  D = "D",
  O = "O",
  R_prime = "R_prime",
  U_prime = "U_prime",
  U_feed_prime = "U_feed_prime",
  U_eiou_prime = "U_EIOU_prime",
  r_eiou_prime = "r_EIOU_prime",
  V_prime = "V_prime"
)
}
\arguments{
\item{.sutmats}{A data frame of supply-use table matrices with matrices arranged in columns.}

\item{Y_prime}{A new final demand matrix or name of a column in \code{.sutmats} containing same.
Default is "Y_prime".}

\item{L_ixp, L_pxp, Z, Z_feed, D, O}{Input matrices that describe the structure of the energy conversion chain.
Values can be string names (the default) for columns in a data frame \code{.sutmats}
or names of items in a list \code{.sutmats}}

\item{R_prime, U_prime, U_feed_prime, U_eiou_prime, r_eiou_prime, V_prime}{The new names for new matrices.
Defaults are each argument name as a string.}
}
\value{
A list or data frame containing \strong{R_prime}, \strong{U_prime}, \strong{U_feed}, \strong{U_eiou}, \strong{r_eiou}, and \strong{V_prime} matrices.
}
\description{
When the final demand matrix changes from \strong{Y} to \strong{Y_prime},
this function calculates new
resource (\strong{R_prime}),
use (\strong{U_prime}),
feed (\strong{U_feed}),
energy industry own use (\strong{U_eiou}),
ratio (\strong{r_eiou}),
and make (\strong{V_prime}) matrices
that would be required to meet the new final demand (\strong{Y_prime}).
}
\details{
Note that inputs \strong{L_ixp}, \strong{L_pxp},
\strong{Z}, and \strong{D} can be
conveniently calculated by the function \code{calc_io_mats()}.

Internally, this function uses \code{matsindf::matsindf_apply()},
and documentation assumes that
\code{.sutmats} is not \code{NULL} and is a data frame.
But \code{.sutmats} can also be a named list of matrices.
Or matrices can be supplied individually to the
\code{Y_prime}, \code{L_ixp}, \code{L_pxp}, \code{Z}, \code{Z_feed}, \code{D}, and \code{O} arguments.
If \code{.sutmats} is present, output is a data frame with columns named by string values of output arguments, and
input arguments should be character strings that name columns in \code{.sutmats}.
If \code{.sutmats} is \code{NULL} (the default), output is a list with items named by output strings,
and input arguments should be single matrices or vectors.
}
\examples{
library(dplyr)
library(matsbyname)
library(tidyr)
UKEnergy2000mats \%>\%
  spread(key = matrix.name, value = matrix) \%>\%
  select(Country, Year, EnergyType, LastStage, R, U, U_feed, V, Y, r_EIOU, S_units) \%>\%
  calc_io_mats() \%>\%
  mutate(
    # Give new Y matrices that are double the existing Y matrices
    Y_prime = matsbyname::hadamardproduct_byname(2, Y)
  ) \%>\%
  # Should give U_prime and V_prime matrices that are double the existing U and V matrices
  new_Y()
}
