% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{data}
\name{nonenergy_products}
\alias{nonenergy_products}
\title{Non-energy products}
\format{
An object of class \code{list} of length 7.
}
\usage{
nonenergy_products
}
\description{
A string vector containing names of products classified as "Non-energy" by
the package development team. This list include hydrocarbon-derived
materials such as "Lubricants".
This list also includes products classified as "Crude, NGL, refinery
feedstocks" in the IEA World Extended Energy Balances 2021 documentation such
as "Additives/blending components".
}
\examples{
Recca::nonenergy_products
}
\keyword{datasets}
