% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{outputs_unit_homogeneous}
\alias{outputs_unit_homogeneous}
\title{Tell whether industry outputs are unit-homogeneous}
\usage{
outputs_unit_homogeneous(.sutdata = NULL, V_colname = "V",
  S_units_colname = "S_units",
  outs_unit_homogeneous = "outputs_unit_homogeneous")
}
\arguments{
\item{.sutdata}{a data frame of supply-use table matrices with matrices arranged in columns.}

\item{V_colname}{the name of the column in \code{.sutdata} that contains
\code{V} matrices. Default is "\code{V}".}

\item{S_units_colname}{the name of the column in \code{.sutdata} that contains
\code{S_units} matrices. Default is "\code{S_units}".}

\item{outs_unit_homogeneous_colname}{the name of the output column
that tells whether each industry's outputs are unit-homogeneous.
Default is "\code{outputs_unit_homogeneous}".}
}
\value{
\code{.sutdata} with additional column "\code{outputs_unit_homogeneous}"
        containing \code{TRUE} if each industry's outputs are unit-homogeneous, \code{FALSE} otherwise.
}
\description{
Returns \code{TRUE} if each industry's output are unit-homogeneous.
}
\details{
The \code{V_bar} matrix is queried for the number of non-zero entries in each row.
If the number of non-zero entries in each row is exactly 1,
industry outputs are unit-homogeneous.
Note that \code{V_bar = \link[matsbyname]{matrixproduct_byname}(V, S_units)}.
}
\examples{
library(magrittr)
UKEnergy2000mats \%>\%
  spread(key = "matrix.name", value = "matrix") \%>\%
  outputs_unit_homogeneous()
}
