% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency.R
\name{pivot_clean_complete_eta_pfd}
\alias{pivot_clean_complete_eta_pfd}
\title{Pivot, clean, and complete an efficiency data frame}
\usage{
pivot_clean_complete_eta_pfd(
  .eta_df,
  efficiency_name_suffix = Recca::efficiency_cols$efficiency_name_suffix,
  last_stage = Recca::psut_cols$last_stage,
  eta_pfd_gross = Recca::efficiency_cols$eta_pfd_gross,
  eta_pfd_net = Recca::efficiency_cols$eta_pfd_net,
  eta_pfd_gross_colname = paste0(eta_pfd_gross, efficiency_name_suffix),
  eta_pfd_net_colname = paste0(eta_pfd_net, efficiency_name_suffix),
  gross_net = Recca::efficiency_cols$gross_net,
  gross = Recca::efficiency_cols$gross,
  net = Recca::efficiency_cols$net,
  eta_pf = Recca::efficiency_cols$eta_pf,
  eta_fu = Recca::efficiency_cols$eta_fu,
  eta_pu = Recca::efficiency_cols$eta_pu,
  eta_ps = Recca::efficiency_cols$eta_ps,
  eta_us = Recca::efficiency_cols$eta_us,
  eta_pw = Recca::efficiency_cols$eta_pw,
  eta_sw = Recca::efficiency_cols$eta_sw,
  aggregate_primary = Recca::aggregate_cols$aggregate_primary,
  gross_aggregate_demand = Recca::aggregate_cols$gross_aggregate_demand,
  net_aggregate_demand = Recca::aggregate_cols$net_aggregate_demand,
  .eta = ".eta",
  .eta_type = ".eta_type",
  .eta_name = ".eta_name"
)
}
\arguments{
\item{.eta_df}{A data frame of efficiencies, likely created by \code{calc_eta_pfd()}.}

\item{efficiency_name_suffix}{A string identifying the suffix for names of efficiency variables.
Default is \code{Recca::efficiency_cols$efficiency_name_suffix}.}

\item{last_stage}{The name of the last stage column.
Default is \code{Recca::psut_cols$last_stage}.}

\item{eta_pfd_gross, eta_pfd_net}{Names of columns containing gross and net primary-to-final-demand efficiencies.
Default values from \code{Recca::efficiency_cols}.}

\item{eta_pfd_gross_colname, eta_pfd_net_colname}{Names of columns containing names of gross and net primary-to-final-demand efficiencies.
Defaults values are \code{paste0(eta_pfd_gross, efficiency_name_suffix)} and
\code{paste0(eta_pfd_net, efficiency_name_suffix)}.}

\item{gross_net, gross, net}{The name (\code{gross_net}) and values for \code{gross} and \code{net} efficiencies.
Default values from \code{Recca::efficiency_cols}.}

\item{eta_pf, eta_fu, eta_pu, eta_ps, eta_us, eta_pw, eta_sw}{Names of columns for various efficiencies between energy conversion chain stages..
"p" means primary stage,
"f" means final stage,
"u" means useful stage,
"s" means services stage, and
"w" means well-being stage.}

\item{aggregate_primary, gross_aggregate_demand, net_aggregate_demand}{Names of columns for net and gross aggregate energy consumption.
See \code{Recca::aggregate_cols} for default values.}

\item{.eta, .eta_type, .eta_name}{Names of internally-generated columns used for intermediate calculations.}
}
\value{
A cleaned version of \code{.eta_df}.
}
\description{
Efficiency data frames (created by \code{calc_eta_pfd()})
include (by default) columns of both
primary->final demand efficiencies and efficiency names.
The efficiency names are preferred to be columns to themselves.
The efficiency data frames can contain residual PSUT matrices,
which are cumbersome to carry around.
Furthermore, the efficiency data frames
contain only primary-to-final demand efficiencies,
not any intermediate efficiencies.
E.g., the data frame would contain
primary-to-final and primary-to-useful efficiencies
but not final-to-useful efficiencies.
This function cleans up these problems by
deleting columns of PSUT matrices (thereby cleaning the data frame),
pivoting to put all efficiencies in columns (thereby pivoting the data frame), and
calculating missing efficiencies (thereby completing the data frame).
}
\details{
This function knows about the following stages
in energy conversion chains: Primary, Final, Useful, Services, and Well-being.

The cleaning step eliminates all columns that contain exclusively matrices.
\code{matsindf::matrix_cols()} identifies the matrix columns.
}
\examples{
psut_mats <- UKEnergy2000mats \%>\%
  tidyr::pivot_wider(names_from = matrix.name, values_from = matrix)
p_industries <- c("Resources - Crude", "Resources - NG")
fd_sectors <- c("Residential", "Transport", "Oil fields")
chop_Y_aggs <- psut_mats \%>\%
  Recca::chop_Y(p_industries = p_industries, fd_sectors = fd_sectors, unnest = TRUE)
etas <- chop_Y_aggs \%>\%
  calc_eta_pfd()
etas \%>\%
  pivot_clean_complete_eta_pfd()
}
