% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregates.R
\name{primary_aggregates}
\alias{primary_aggregates}
\title{Primary aggregate energy}
\usage{
primary_aggregates(
  .sutdata,
  p_industries,
  R = Recca::psut_cols$R,
  V = Recca::psut_cols$V,
  Y = Recca::psut_cols$Y,
  by = c("Total", "Product", "Flow"),
  aggregate_primary = Recca::aggregate_cols$aggregate_primary
)
}
\arguments{
\item{.sutdata}{a data frame with columns of matrices from a supply-use analysis.}

\item{p_industries}{a vector of names of industries to be aggregated as "primary".
These industries will appear in rows of the resource (\code{R}) and make (\code{V}) matrices and
columns of the final demand matrix (\code{Y}).
Entries in \code{Y_p} will be subtracted from entries in \code{R_p + V_p} to obtain
the total primary energy aggregate.}

\item{R, V, Y}{See \code{Recca::psut_cols}.}

\item{V}{make (\code{V}) matrix or the name of the column in \code{.sutdata} containing same}

\item{Y}{final demand (\code{Y}) matrix or the name of the column in \code{.sutdata} containing same}

\item{by}{one of \code{Total}, \code{Product}, or \code{Flow} to indicate the desired aggregation:
\itemize{
\item \code{Total}: aggregation over both Product and Flow (the default)
\item \code{Product}: aggregation by energy carrier (Crude oil, Primary solid biofuels, etc.)
\item \code{Flow}: aggregation by type of flow (Production, Imports, Exports, etc.)
}}

\item{aggregate_primary}{the name for aggregates of primary energy on output}
}
\value{
a list or data frame containing aggregate primary energy
}
\description{
Calculates aggregate primary energy from a data frame of Supply-Use matrices.
}
