% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregates.R
\name{primary_aggregates}
\alias{primary_aggregates}
\title{Primary aggregate energy}
\usage{
primary_aggregates(
  .sutdata,
  p_industries,
  pattern_type = c("exact", "leading", "trailing", "anywhere"),
  R = Recca::psut_cols$R,
  V = Recca::psut_cols$V,
  Y = Recca::psut_cols$Y,
  by = c("Total", "Product", "Flow"),
  aggregate_primary = Recca::aggregate_cols$aggregate_primary
)
}
\arguments{
\item{.sutdata}{A data frame with columns of matrices from a supply-use analysis.}

\item{p_industries}{a vector of names of industries to be aggregated as "primary."
If \code{.sutdata} is a data frame, \code{p_industries} should be the name of a column in the data frame.
If \code{.sutdata} is \code{NULL}, \code{p_industries} can be a single vector of industry names.
These industries in \code{p_industries} will appear in rows of the resource (\code{R}) and make (\code{V}) matrices and
columns of the final demand matrix (\code{Y}).
Entries in \code{Y_p} will be subtracted from entries in \code{R_p + V_p} to obtain
the total primary energy aggregate,
where \verb{*_p} is the primary part of those matrices.
The function \code{find_p_industry_names()} might be helpful to find
primary industry names if they can be identified by prefixes.}

\item{pattern_type}{One of "exact", "leading", "trailing", or "anywhere" which specifies
how matches are made for \code{fd_sectors}.
If "exact", exact matches specify the sectors to be aggregated.
If "leading", sectors are aggregated if any entry in \code{fd_sectors} matches the leading part of a final demand sector's name.
If "trailing", sectors are aggregated if any entry in \code{fd_sectors} matches the trailing part of a final demand sector's name.
If "anywhere", sectors are aggregated if any entry in \code{fd_sectors} matches any part of a final demand sector's name.
Default is "leading" to enable primary industries such as "Resources [of Crude oil]".}

\item{R, V, Y}{See \code{Recca::psut_cols}.}

\item{by}{One of "Total", "Product", or "Flow" to indicate the desired aggregation:
\itemize{
\item "Total": aggregation over both Product and Flow (the default),
\item "Product": aggregation by energy carrier (Crude oil, Primary solid biofuels, etc.), or
\item "Flow": aggregation by type of flow (Production, Imports, Exports, etc.).
}}

\item{aggregate_primary}{The name for aggregates of primary energy on output.}
}
\value{
A list or data frame containing aggregate primary energy.
}
\description{
Calculates aggregate primary energy from a data frame of Supply-Use matrices.
}
\examples{
library(matsbyname)
p_industries <- c("Resources - Crude", "Resources - NG")
# Calculate primary total aggregates
res <- UKEnergy2000mats \%>\%
  tidyr::pivot_wider(names_from = "matrix.name", values_from = "matrix") \%>\%
  dplyr::mutate(
    p_industries = rep(list(p_industries), times = nrow(.))
  ) \%>\%
  primary_aggregates(p_industries = "p_industries", by = "Total")
res[[Recca::aggregate_cols$aggregate_primary]]
}
