% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregates.R
\name{primary_aggregates}
\alias{primary_aggregates}
\title{Primary aggregate energy}
\usage{
primary_aggregates(.sutdata, p_industries, V_colname = "V", Y_colname = "Y",
  by = c("Total", "Product", "Flow"),
  aggregate_primary_colname = "EX_p.ktoe")
}
\arguments{
\item{.sutdata}{a data frame with columns of matrices from a supply-use analysis.}

\item{p_industries}{a vector of names of industries to be aggregated as "primary".
These industries will appear in rows of the make matrix (\code{V}) and
columns of the final demand matrix (\code{Y}).
Entries in \code{Y_p} will be subtracted from entries in \code{V_p} to obtain
the total primary energy aggregate.}

\item{V_colname}{the name of the column in \code{.sutdata} containing make (\code{V}) matrices.}

\item{Y_colname}{the name of the column in \code{.sutdata} containing final demand (\code{Y}) matrices.}

\item{by}{one of \code{Total}, \code{Product}, or \code{Flow} to indicate the desired aggregation:
\itemize{
  \item{\code{Total}: aggregation over both Product and Flow (the default)}
  \item{\code{Product}: aggregation by energy carrier (Crude oil, Primary solid biofuels, etc.)}
  \item{\code{Flow}: aggregation by type of flow (Production, Imports, Exports, etc.)}
}}

\item{aggregate_primary_colname}{the name of the output column containing aggregates of primary energy.}
}
\value{
a data frame containing columns of
the aggregate primary energy for each row of \code{.sutdata}.
}
\description{
Calculates aggregate primary energy from a data frame of Supply-Use matrices.
}
