% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregates.R
\name{primary_aggregates_IEA}
\alias{primary_aggregates_IEA}
\title{Primary aggregate energy from IEA tables}
\usage{
primary_aggregates_IEA(.ieadata, flow = "Flow", energy = "EX.ktoe",
  p_industries = "p_industries", aggregate_primary_colname = "EX_IEA.ktoe")
}
\arguments{
\item{.ieadata}{the data frame containing an IEA table}

\item{aggregate_primary_colname}{the name of the aggregate primary energy
column to be created in the output data frame.}

\item{UVY_colname}{the name of the column in \code{.ieadata}
containing names of submatrices of U, V, and G to which this
row of data belongs.}

\item{energy_colname}{the name of the column in \code{.ieadata}
that contains energy data.}

\item{energy_type_colname}{the name of the column in \code{.ieadata}
that contains the energy type for this row in the table}

\item{V_p_varname}{the name of the primary submatrix (\code{_p})
of the Make (\code{V}) matrix.}

\item{Y_p_varname}{the name of the primary submatrix (\code{_p})
of the final demand (\code{G}) matrix.}
}
\value{
a data frame containing the grouping columns of \code{.ieadata}
as well as a column named \code{aggregate_primary_colname}.
}
\description{
Calculates total aggregate primary energy from a data frame of IEA data.
This function is named with "_IEA", because it is meant to operate on
tidy, IEA-style data frames.
The function \code{primary_aggregates} does the same thing,
but it is meant to operate on tidy SUT-style data frames.
}
\details{
This function works similar to \code{\link[dplyr]{summarise}}:
it distills \code{.ieadata} to many fewer rows
according to the grouping variables.
Thus, \code{.ieadata} should be grouped prior to sending into this function.
Grouping columns are preserved on output.
}
