% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregates.R
\name{primary_aggregates_IEA}
\alias{primary_aggregates_IEA}
\title{Total primary aggregate energy from IEA tables}
\usage{
primary_aggregates_IEA(
  .ieadata,
  flow = "Flow",
  flow_aggregation_point = "Flow.aggregation.point",
  energy = "E.dot",
  p_industries = c("Production", "Coal mines", "Oil and gas extraction", "Imports",
    "Exports", "International aviation bunkers", "International marine bunkers",
    "Stock changes"),
  eiou = "Energy industry own use",
  aggregate_primary = Recca::aggregate_cols$aggregate_primary_iea
)
}
\arguments{
\item{.ieadata}{the data frame containing an IEA table}

\item{flow}{the name of the column that contains flow information. Default is "\code{Flow}".}

\item{flow_aggregation_point}{the name of the column that identifies flow aggregation points.
Default is "\code{Flow.aggregation.point}".}

\item{energy}{the name of the column that contains energy information. Default is "\code{E.dot}".}

\item{p_industries}{a vector of names of primary industries. Default is
"\code{c("Coal mines", "Oil and gas extraction",
       "Production", "Imports", "Exports",
       "International aviation bunkers", "International marine bunkers",
       "Stock changes")}"}

\item{eiou}{the string that identifies energy industry own use in the \code{Flow.aggregation.point} column.
Default is "\code{Energy industry own use}".}

\item{aggregate_primary}{the name of the aggregate primary energy
column to be created in the output data frame.
Default is "\code{EX_p_IEA.ktoe}".}
}
\value{
a data frame containing the grouping columns of \code{.ieadata}
as well as a column named with the value of \code{aggregate_primary}.
}
\description{
Calculates total aggregate primary energy from a data frame of IEA data.
This function is named with "_IEA", because it is meant to operate on
tidy, IEA-style data frames.
The function \code{primary_aggregates} does the same thing,
but it is meant to operate on tidy SUT-style data frames.
}
\details{
This function works similar to \code{\link[dplyr]{summarise}}:
it distills \code{.ieadata} to many fewer rows
according to the grouping variables.
Thus, \code{.ieadata} should be grouped prior to sending into this function.
Grouping columns are preserved on output.
}
\examples{
library(dplyr)
library(tidyr)
r_ind <- resource_industries(UKEnergy2000mats \%>\%
  spread(key = matrix.name, value = matrix))[["r_industries"]][[1]]
UKEnergy2000tidy \%>\%
  group_by(Country, Year, Energy.type, Last.stage) \%>\%
  primary_aggregates_IEA(p_industries = r_ind)
}
