% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregates.R
\name{primary_aggregates_IEA}
\alias{primary_aggregates_IEA}
\title{Primary aggregate energy from IEA tables}
\usage{
primary_aggregates_IEA(.ieadata, UVY_colname = "UVY",
  energy_colname = "EX.ktoe", V_p_varname = "V_p", Y_p_varname = "Y_p",
  aggregate_primary_colname)
}
\arguments{
\item{.ieadata}{the data frame containing an IEA table}

\item{UVY_colname}{the name of the column in \code{.ieadata}
containing names of submatrices of U, V, and G to which this
row of data belongs.}

\item{energy_colname}{the name of the column in \code{.ieadata}
that contains energy data.}

\item{V_p_varname}{the name of the primary submatrix (\code{_p})
of the Make (\code{V}) matrix.}

\item{Y_p_varname}{the name of the primary submatrix (\code{_p})
of the final demand (\code{G}) matrix.}

\item{aggregate_primary_colname}{the name of the aggregate primary energy
column to be created in the output data frame.}

\item{energy_type_colname}{the name of the column in \code{.ieadata}
that contains the energy type for this row in the table}
}
\value{
a data frame containing the grouping columns of \code{.ieadata}
as well as a column named \code{aggregate_primary_colname}.
}
\description{
Calculates aggregate primary energy from a data frame of IEA data tables.
This function is named with "_IEA", because it is meant to operate on
tidy, IEA-style data frames.
The function \code{primary_aggregates} does the same thing,
but it is meant to operate on tidy SUT-style data frames.
}
\details{
Note that \code{.ieadata} should be grouped prior to sending into this function.
Grouping columns are preserved on output.
}
