% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{read_ecc_from_excel}
\alias{read_ecc_from_excel}
\title{Read an energy conversion chain from a Excel file}
\usage{
read_ecc_from_excel(
  path,
  worksheets = NULL,
  R = Recca::psut_cols$R,
  U = Recca::psut_cols$U,
  V = Recca::psut_cols$V,
  Y = Recca::psut_cols$Y,
  r_eiou = Recca::psut_cols$r_eiou,
  U_eiou = Recca::psut_cols$U_eiou,
  U_feed = Recca::psut_cols$U_feed,
  S_units = Recca::psut_cols$S_units,
  sep = "_"
)
}
\arguments{
\item{path}{The path to the Excel file.}

\item{worksheets}{Names of worksheets from which matrices
are to be read.
Default is \code{NULL}, meaning that all
worksheets are read.}

\item{R, U, V, Y, r_eiou, U_eiou, U_feed, S_units}{String names for matrices
in the file at \code{path}
containing matrices.
See \code{Recca::psut_cols}
for defaults.}

\item{sep}{The separator between matrix name and worksheet name
for named regions.
Default is "_".}
}
\value{
A data frame in \code{matsindf} format containing
the matrices from named regions in \code{path}.
}
\description{
Reads matrices from named regions in an Excel file
into \code{matsindf} format.
The named regions are assumed to be global
to the workbook.
The format for the region names is assumed to be
\verb{<<matrix symbol>><<sep>><<worksheet name>>},
as written by \code{\link[=write_ecc_to_excel]{write_ecc_to_excel()}}.
For example, "R_4" for the \strong{R} matrix on the sheet named "4".
}
\details{
Named regions are assumed to include
the rectangle of numerical values,
row names (to the left of the rectangle of numbers), and
column names (above the rectangle of numbers),
the same format
as written by \code{\link[=write_ecc_to_excel]{write_ecc_to_excel()}}.

This function is an inverse of \code{\link[=write_ecc_to_excel]{write_ecc_to_excel()}}.
}
\seealso{
\code{\link[=write_ecc_to_excel]{write_ecc_to_excel()}}
}
