% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recca-package.R
\docType{package}
\name{Recca-package}
\alias{Recca}
\alias{Recca-package}
\title{Recca: R Energy Conversion Chain Analysis}
\description{
Provides functions to analyze energy conversion chains using a physical supply use table (PSUT) approach.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MatthewHeun/Recca}
  \item \url{https://matthewheun.github.io/Recca/}
  \item Report bugs at \url{https://github.com/MatthewHeun/Recca/issues}
}

}
\author{
\strong{Maintainer}: Matthew Heun \email{matthew.heun@me.com} (\href{https://orcid.org/0000-0002-7438-214X}{ORCID})

Authors:
\itemize{
  \item Emmanuel Aramendia \email{e.aramendia@leeds.ac.uk} (\href{https://orcid.org/0000-0002-5964-6776}{ORCID})
}

}
\keyword{internal}
