% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruct.R
\name{reconstruct_UV}
\alias{reconstruct_UV}
\title{Reconstruct an economy given a new final demand matrix}
\usage{
reconstruct_UV(.sutdata, Y_prime_colname = "Y_prime",
  L_ixp_colname = "L_ixp", L_pxp_colname = "L_pxp", Z_colname = "Z",
  D_colname = "D", keep_cols = NULL, U_prime_colname = "U_prime",
  V_prime_colname = "V_prime", y_prime_colname = ".yprime",
  g_prime_colname = ".gprime", q_prime_colname = ".qprime")
}
\arguments{
\item{.sutdata}{a data frame of supply-use table matrices with matrices arranged in columns.}

\item{Y_prime_colname}{the name of a column containing new final demand matrices
that will be used to reconstruct the economy}

\item{L_ixp_colname}{the name of a column containing industry-by-product L matrices.}

\item{L_pxp_colname}{the name of a column containing product-by-product L matrices.}

\item{Z_colname}{the name of a column containing \code{Z} matrices.}

\item{D_colname}{the name of a column containing \code{D} matrices.}

\item{keep_cols}{a vector of names of columns of \code{.sutdata} to return with the output}

\item{U_prime_colname}{the name of the output column that contains new Use (\code{U}) matrices.}

\item{V_prime_colname}{the name of the output column that contains new Make (\code{V}) matrices.}

\item{y_prime_colname}{the name of the column containing y_prime vectors}

\item{g_prime_colname}{the name of the column containing g_prime vectors}

\item{q_prime_colname}{the name of the column containing q_prime vectors}
}
\value{
a data frame containing columns specified in \code{keep_cols},
\code{U_colname} and \code{V_colname}.
}
\description{
Reconstruct an economy given a new final demand matrix
}
\examples{

}
