% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{resource_industries}
\alias{resource_industries}
\title{Resource industries}
\usage{
resource_industries(.sutdata = NULL, U = "U", V = "V",
  r_industries = "r_industries")
}
\arguments{
\item{.sutdata}{a list or data frame containing use matrix(ces) and make matrix(ces)}

\item{U}{identifier for the use matrix (a string). Default is "\code{U}".}

\item{V}{identifier for the make matrix (a string). Default is "\code{V}".}

\item{r_industries}{name for the resource industries vector (a string). Default is "\code{r_industries}".}
}
\value{
\code{.sutdata} with an additional column (named with the value of the \code{p_industries} argument)
        containing the resource industries for each row
}
\description{
Identifies resource industries.
}
\details{
Resource industries are industries that make a product without using any products.
Resource industries are identified by interrogating
the use (\code{U}) and make (\code{V}) matrices.
Resource industries have all zeroes in their column of the use matrix (\code{U})
and at least one non-zero value in their row of the make (\code{V}) matrix.

Argument and value descriptions are written assuming that \code{.sutdata} is a data frame.
Alternatively, \code{.sutdata} can be unspecified, and \code{U} and \code{V} can be matrices.
In that case, the return value is a list with a single item: \code{r_industries}
which contains a vector of names of resource industries for the \code{U} and \code{V} matrices.
}
\examples{
library(magrittr)
library(tidyr)
resource_industries(UKEnergy2000mats \%>\% spread(key = matrix.name, value = matrix))
}
