% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{separate_RV}
\alias{separate_RV}
\title{Separate resource (\code{R}) and make (\code{V}) matrices from make plus resource (\code{R_plus_V}) matrices}
\usage{
separate_RV(.sutmats = NULL, U = "U", R_plus_V = "R_plus_V",
  R = "R", V = "V")
}
\arguments{
\item{.sutmats}{a list or data frame containing use matrix(ces) and make matrix(ces)}

\item{U}{a use (\code{U}) matrix or name of the column in \code{.sutmats} that contains same. Default is "\code{U}".}

\item{R_plus_V}{an \code{R_plus_V} matrix or name of the column in \code{.sutmats} that contains same. Default is "\code{R_plus_V}".}

\item{R}{name for resource (\code{R}) matrix on output. Default is "\code{R}".}

\item{V}{name for make (\code{V}) matrix on output. Default is "\code{V}".}
}
\value{
a list or data frame containing \code{R} and \code{V} matrices
}
\description{
Resource industries are industries that make a product without using any products.
Resource industries are identified by interrogating
the use (\code{U}) and make (\code{R_plus_V}) matrices.
Resource industries have all zeroes in their column of the use matrix (\code{U})
and at least one non-zero value in their row of the make (\code{R_plus_V}) matrix.
}
\details{
A resource matrix (\code{R}) has industries in rows and products in columns.
The elements of of \code{R} indicate extraction of resources from the biosphere.
The industries of \code{R} are the reserves of the extracted products.

This function uses the \code{\link{resource_industries}} function to
identify the resource industries in the \code{R_plus_V} matrix.
Thereafter, the function extracts the resource industries from the \code{R_plus_V} matrix
to form the \code{R} matrix.
Finally, the \code{R} matrix is subtracted from the \code{R_plus_V} matrix
and saved as the \code{V} matrix.
If there are no resource industries in the \code{R_plus_V} matrix,
a warning is emitted,
no \code{R} matrix is created, and
no changes are made to the \code{R_plus_V} matrix.

\code{\link{separate_RV}} is the inverse of \code{\link{combine_RV}}.
}
\examples{
library(dplyr)
library(tidyr)
UKEnergy2000mats \%>\%
  spread(key = "matrix.name", value = "matrix") \%>\%
  # Rename the V matrix, because it includes the R matrix.
  rename(
    R_plus_V = V
  ) \%>\%
  separate_RV()
}
