% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_edge_list.R
\name{simplify_edge_list}
\alias{simplify_edge_list}
\title{Simplify an edge list}
\usage{
simplify_edge_list(edge_list, from = "From", to = "To",
  value = "Value", product = "Product")
}
\arguments{
\item{edge_list}{the edge list to be simplified}

\item{from}{the name of the from column in the edge list. (Default is "\code{From}".)}

\item{to}{the name of the to column in the edge list. (Default is "\code{To}".)}

\item{value}{the name of the value column in the edge list. (Default is "\code{Value}".)}

\item{product}{the name of the product column in the edge list. (Default is "\code{Product}".)}
}
\value{
a simplified edge list
}
\description{
A PSUT energy conversion chain edge can be simplified if
a product has only one supplier (i.e., one "from").
Then, every "to" node for that product can have the product's "from" be its "from".
See examples.
}
\examples{
el <- data.frame(From = c("A", "Oil"), To = c("Oil", "C"),
                 Value = c(42, 42), Product = c("Oil", "Oil"),
                 stringsAsFactors = FALSE)
# Oil flows from A to C through its product node (Oil)
el
# Simplify to have Oil flow from A to C without the product node
simplify_edge_list(el)
}
