% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UKEnergy2000tidy}
\alias{UKEnergy2000tidy}
\title{Energy consumption in the UK in 2000}
\format{
A data frame with 186 rows and 9 variables:
\describe{
\item{Country}{country, GB (Great Britain, only one country)}
\item{Year}{year, 2000 (only one year)}
\item{LedgerSide}{Supply or Consumption}
\item{FlowAggregationPoint}{tells where each row should be aggregated}
\item{EnergyType}{E.ktoe (for energy) or X.ktoe (for exergy)}
\item{LastStage}{tells the final stage of the energy conversion chain: final, useful, or services}
\item{Flow}{the Industry or Sector involved in this flow}
\item{Product}{the energy product involved in this flow}
\item{Edot}{value of the energy, exergy, or service flow in ktoe}
\item{Unit}{unit in which quantity is expressed}
}
}
\source{
\doi{10.1016/j.apenergy.2018.05.109}
}
\usage{
UKEnergy2000tidy
}
\description{
A dataset containing approximations to
some of the energy flows in the UK in the year 2000.
These data first appeared as the example in
Figures 3, 4, 6, 7, 10, 11, B.1, and B.2 of
M.K. Heun, A. Owen, and P. E. Brockway.
A physical supply-use table framework for energy analysis on the energy conversion chain.
Applied Energy, 226:1134–1162, Sep 2018.
}
\details{
\code{UKEnergy2000tidy} gives each non-zero entry in \code{UKEnergy2000mats}
as a single column in a data frame.
These data are in tidy format.
}
\keyword{datasets}
