% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy_balance.R
\name{verify_IEATable_energy_balance}
\alias{verify_IEATable_energy_balance}
\title{Confirm that an IEA-style data frame conserves energy.}
\usage{
verify_IEATable_energy_balance(.ieatidydata, ledger.side = "Ledger.side",
  energy = "E.ktoe", supply = "Supply", consumption = "Consumption",
  tol = 1e-06)
}
\arguments{
\item{.ieatidydata}{an IEA-style data frame containing the following columns:
\code{Country}, \code{Year}, \code{Ledger.side}, \code{Product}, \code{E.ktoe}.}

\item{tol}{the maximum amount by which Supply and Consumption can be out of balance}
}
\value{
Nothing is returned.
This function should be called
for its side-effect of testing whether energy is balanced in \code{.ieatidydata}.
}
\description{
Energy balances are confirmed (within \code{tol}) for every combination of
grouping variables in \code{.ieatidydata}.
If energy is in balance for every group of \code{.ieatidydata},
execution returns to the caller.
If energy balance is not observed for any group of \code{.ieatidydata},
a message is printed which shows the first few non-balancing Products, and
execution halts.
}
\details{
Be sure to group \code{.ieatidydata} prior to calling this function,
as shown in the example.
}
\examples{
verify_IEATable_energy_balance(IEAData_orig \%>\% group_by(Country, Year, Product))
}
