% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy_balance.R
\name{verify_IEATable_energy_balance}
\alias{verify_IEATable_energy_balance}
\title{Confirm that an IEA-style data frame conserves energy.}
\usage{
verify_IEATable_energy_balance(.ieatidydata, ledger.side = "Ledger.side",
  energy = "E.ktoe", supply = "Supply", consumption = "Consumption",
  err = ".err", tol = 1e-06)
}
\arguments{
\item{.ieatidydata}{an IEA-style data frame containing grouping columns
(typically \code{Country}, \code{Year}, \code{Product}, and others),
a \code{Ledger.side} column, and
an energy column (\code{E.ktoe}).
\code{.ieatidydata} should be grouped prior to sending to this function.}

\item{ledger.side}{the name of the column in \code{.ieatidydata}
that contains ledger side information (a string). Default is "\code{Ledger.side}".}

\item{energy}{the name of the column in \code{.ieatidydata}
that contains energy data (a string). Default is "\code{E.ktoe}".}

\item{supply}{the identifier for supply data in the \code{ledger.side} column (a string).
Default is "\code{Supply}".}

\item{consumption}{the identifier for consumption data in the \code{ledger.side} column (a string).
Default is "\code{Consumption}".}

\item{err}{the name of the error column in the output. Default is "\code{.err}".}

\item{tol}{the maximum amount by which Supply and Consumption can be out of balance}
}
\value{
a data frame containing with grouping variables and
        an additional column whose name is the value of \code{err}.
        The \code{err} column should be 0.
}
\description{
Energy balances are confirmed (within \code{tol}) for every combination of
grouping variables in \code{.ieatidydata}.
}
\details{
Be sure to group \code{.ieatidydata} prior to calling this function,
as shown in the example.

If energy is in balance for every group, a data frame with additional column \code{err}
is returned.
If energy balance is not observed for one or more of the groups,
a warning is emitted.
}
\examples{
UKEnergy2000tidy \%>\%
  filter(Last.stage \%in\% c("final", "useful")) \%>\%
  group_by(Country, Year, Energy.type, Last.stage) \%>\%
  verify_IEATable_energy_balance(energy = "EX.ktoe")
}
