% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy_balance.R
\name{verify_inter_industry_balance}
\alias{verify_inter_industry_balance}
\title{Confirm that an SUT-style data frame conserves products between industries}
\usage{
verify_inter_industry_balance(
  .sutmats_with_balances = NULL,
  balances = Recca::balance_cols$inter_industry_balance_colname,
  tol = 1e-06,
  balanced_colname = Recca::balance_cols$inter_industry_balanced_colname
)
}
\arguments{
\item{.sutmats_with_balances}{An SUT-style data frame with a column
named \code{balances}.}

\item{balances}{The name of a column that contains balance vectors.
Default is \link{balance_cols}\verb{$inter_industry_balance_colname}.}

\item{tol}{The maximum amount by which products can be out of balance.
Default is \code{1e-6}.}

\item{balanced_colname}{The name for booleans telling if balance is present.
Default is \link{balance_cols}\verb{$inter_industry_balanced_colname}.}
}
\value{
A list or data frame saying whether \code{.sutmats_with_balances} are in balance.
}
\description{
Inter-industry energy balances are verified by product
(within \code{tol}) for every row in \code{.sutmats}.
}
\details{
In a PSUT description of an economy, all of every product
leaving one industry must arrive at another industry.

If every row is balanced,
\code{.sutmats} is returned with an additional column, and
execution returns to the caller.
If balance is not observed for one or more of the rows,
a warning is emitted, and
the additional column (\code{balanced_colname})
indicates where the problem occurred, with \code{FALSE}
showing where products are not balanced
to within \code{tol}.

Typically, one would call \code{\link[=calc_inter_industry_balance]{calc_inter_industry_balance()}}
before calling \code{\link[=verify_inter_industry_balance]{verify_inter_industry_balance()}}.
See the examples.
}
\examples{
library(dplyr)
library(tidyr)
result <- UKEnergy2000mats |>
  dplyr::filter(LastStage \%in\% c("Final", "Useful")) |>
  tidyr::pivot_wider(names_from = matrix.name,
                     values_from = matrix) |>
  calc_inter_industry_balance() |>
  verify_inter_industry_balance(tol = 1e-4)
result
result[[Recca::balance_cols$inter_industry_balanced_colname]]
}
\seealso{
\code{\link[=calc_inter_industry_balance]{calc_inter_industry_balance()}}
}
