% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy_balance.R
\name{verify_SUT_energy_balance}
\alias{verify_SUT_energy_balance}
\title{Confirm that an SUT-style data frame conserves energy.}
\usage{
verify_SUT_energy_balance(
  .sutmats = NULL,
  R = "R",
  U = "U",
  V = "V",
  Y = "Y",
  tol = 1e-06,
  SUT_energy_balance = ".SUT_energy_balance"
)
}
\arguments{
\item{.sutmats}{an SUT-style data frame with columns of matrices,
including \code{U}, \code{V}, and \code{Y} columns.}

\item{R}{resources (\code{R}) matrix or name of the column in \code{.sutmats} that contains same. Default is "R".}

\item{U}{use (\code{U}) matrix or name of the column in \code{.sutmats} that contains same. Default is "U".}

\item{V}{make (\code{V}) matrix or name of the column in \code{.sutmats}that contains same. Default is "V".}

\item{Y}{final demand (\code{Y}) matrix or name of the column in \code{.sutmats} that contains same. Default is "Y".}

\item{tol}{the maximum amount by which Supply and Consumption can be out of balance.
Default is \code{1e-6}.}

\item{SUT_energy_balance}{the name for booleans telling if energy is in balance. Default is ".SUT_energy_balance".}
}
\value{
a list or data frame saying whether \code{.sutmats} are in balance.
}
\description{
Energy balances are confirmed by Product (within \code{tol}) for every row in \code{.sutmats}.
}
\details{
If energy is in balance for every row, \code{.sutmats} is returned with an additional column, and
execution returns to the caller.
If energy balance is not observed for one or more of the rows,
a warning is emitted, and
the additional column (\code{SUT_energy_blance})
indicates where the problem occurred, with \code{FALSE}
showing where energy is not in balance.
}
\examples{
library(dplyr)
library(tidyr)
verify_SUT_energy_balance(UKEnergy2000mats \%>\%
                            dplyr::filter(Last.stage \%in\% c("final", "useful")) \%>\%
                            tidyr::spread(key = matrix.name, value = matrix),
                          tol = 1e-4)
}
