% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{verify_SUT_energy_balance_with_units}
\alias{verify_SUT_energy_balance_with_units}
\title{Confirm that an SUT-style data frame conserves energy.}
\usage{
verify_SUT_energy_balance_with_units(
  .sutmats = NULL,
  R = "R",
  U = "U",
  U_feed = "U_feed",
  U_eiou = "U_EIOU",
  r_eiou = "r_EIOU",
  V = "V",
  Y = "Y",
  S_units = "S_units",
  tol = 1e-06,
  matnames = "matnames",
  matvals = "matvals",
  rowtypes = "rowtypes",
  coltypes = "coltypes",
  rownames = "rownames",
  colnames = "colnames",
  prod_diff = ".prod_diff",
  ind_diff = ".ind_diff",
  SUT_prod_energy_balanced = ".SUT_prod_energy_balanced",
  SUT_ind_energy_balanced = ".SUT_ind_energy_balanced",
  ebal_error = "ebal_error",
  product = "Product"
)
}
\arguments{
\item{.sutmats}{An SUT-style data frame containing columns
\code{R} (optionally), \code{U}, \code{V}, \code{Y}, and \code{S_units}.}

\item{R}{Resource (\code{R}) matrix or name of the column in \code{.sutmats} that contains same. Default is "R".}

\item{U}{Use (\code{U}) matrix or name of the column in \code{.sutmats} that contains same. Default is "U".}

\item{U_feed, U_eiou, r_eiou}{Optional matrices or columns in \code{.sutmats}.}

\item{V}{Make (\code{V}) matrix or name of the column in \code{.sutmats}that contains same. Default is "V".}

\item{Y}{Final demand (\code{Y}) matrix or name of the column in \code{.sutmats} that contains same. Default is "Y".}

\item{S_units}{\code{S_units} A matrix or name of the column in \code{.sutmats} that contains same. Default is "S_units".}

\item{tol}{The maximum amount by which energy can be out of balance. Default is \code{1e-6}.}

\item{matnames, matvals, rowtypes, coltypes, rownames, colnames}{Column names used internally.}

\item{prod_diff, ind_diff, ebal_error, product}{Column names for product and industry energy balance errors.}

\item{SUT_prod_energy_balanced}{The name for booleans telling if product energy is in balance. Default is ".SUT_prod_energy_balance".}

\item{SUT_ind_energy_balanced}{The name for booleans telling if product energy is in balance. Default is ".SUT_inds_energy_balance".}
}
\value{
\code{.sutmats} with additional columns.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\details{
If energy is in balance for every row,
\code{.sutmats} is returned with two additional columns, and
execution returns to the caller.
If energy balance is not observed for one or more rows,
a warning is emitted, and
columns named \code{SUT_prod_energy_blance} and \code{SUT_ind_energy_blance} are added to \code{.sutmats}.
\code{FALSE} indicates energy is not in balance.

This function should be called
for its side-effect of testing whether energy is in balance in \code{.sutmats}.

Both product and industry energy balance are verified.
Units (as supplied by the \code{S_units} matrix) are respected.
}
\examples{
library(tidyr)
verify_SUT_energy_balance_with_units(UKEnergy2000mats \%>\%
                                       tidyr::spread(key = matrix.name, value = matrix),
                                       tol = 1e-3)
}
\keyword{internal}
