% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy_balance.R
\name{verify_SUT_energy_balance_with_units}
\alias{verify_SUT_energy_balance_with_units}
\title{Confirm that an SUT-style data frame conserves energy.}
\usage{
verify_SUT_energy_balance_with_units(.sutdata = NULL, U = "U", V = "V",
  Y = "Y", S_units = "S_units", tol = 1e-06,
  SUT_prod_energy_balance = ".SUT_prod_energy_balance",
  SUT_ind_energy_balance = ".SUT_ind_energy_balance")
}
\arguments{
\item{.sutdata}{an SUT-style data frame containing columns
\code{U}, \code{V}, \code{Y}, and \code{S_units}.}

\item{U}{the name of the column that contains \code{U} matrices}

\item{V}{the name of the column that contains \code{V} matrices}

\item{Y}{the name of the column that contains \code{Y} matrices}

\item{S_units}{the name of the column that contains \code{S_units} matrices}

\item{tol}{the maximum amount by which energy can be out of balance}
}
\value{
\code{.sutdata} with additional columns.
}
\description{
If energy is in balance for every row, \code{.sutdata} is returned with two additional columns, and
execution returns to the caller.
If energy balance is not observed for one or more rows,
a warning is emitted, and
columns named \code{SUT_prod_energy_blance} and \code{SUT_ind_energy_blance} are added to \code{.sutdata}.
\code{FALSE} indicates energy is not in balance.
}
\details{
This function should be called
for its side-effect of testing whether energy is in balance in \code{.sutdata}.

Both product and industry energy balance are verified.
Units (as supplied by the \code{S_units} matrix) are respected.
}
\examples{
verify_SUT_energy_balance_with_units(UKEnergy2000mats)
}
