% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy_balance.R
\name{verify_SUT_industry_production}
\alias{verify_SUT_industry_production}
\title{Confirm that all Industries in an SUT-style data frame produce energy.}
\usage{
verify_SUT_industry_production(
  .sutmats = NULL,
  R = "R",
  U = "U",
  V = "V",
  industry_production_OK = ".industry_production_OK",
  problem_industries = ".problem_industries"
)
}
\arguments{
\item{.sutmats}{an SUT-style data frame containing metadata columns
(typically \code{Country}, \code{Year}, \code{Ledger.side}, \code{Product}, etc.)
and columns of SUT matrices, including \code{U} and \code{V}.}

\item{R}{resources (\strong{R}) matrix or name of the column in \code{.sutmats} that contains same. Default is "R".}

\item{U}{use (\strong{U}) matrix or name of the column in \code{.sutmats} that contains same. Default is "U".}

\item{V}{make (\strong{V}) matrix or name of the column in \code{.sutmats}that contains same. Default is "V".}

\item{industry_production_OK}{the name of the column in the output that
tells whether all industries produce something. Default is ".industry_production_OK".}

\item{problem_industries}{the name of the column in the output that
tells which transformation processes consume energy but do not produce anything.}
}
\value{
\code{.sutmats} with added column named with the value of \code{industry_production_OK}.
}
\description{
If a transformation process industry consumes energy (in the \code{U} matrix)
but does not make energy (in the \code{V} matrix),
it is most certainly an error.
(In contrast, there can be Industries that make energy but do not consume it,
such as Industries involved in Production.
And final demand sectors consume energy but do not produce any.)
This function emits a warning if an Industry in the \code{U} matrix
is found to consume energy but not make energy.
Look at the \code{industry_production_OK} column of the output to see which rows of
\code{.sutmats} exhibit the problem.
Look at the \code{problem_industries} column of the output to see which industries
exhibit this problem.
}
\examples{
library(tidyr)
verify_SUT_industry_production(UKEnergy2000mats \%>\%
                                 spread(key = matrix.name, value = matrix))
}
