% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy_balance.R
\name{verify_SUT_industry_production}
\alias{verify_SUT_industry_production}
\title{Confirm that all Industries in an SUT-style data frame produce energy.}
\usage{
verify_SUT_industry_production(.sutdata, U_colname = "U", V_colname = "V",
  industry_production_OK_colname = "industry production OK",
  check_colname = ".check", problems_data_frame_name = "Problems")
}
\arguments{
\item{.sutdata}{an SUT-style data frame containing metadata columns
(typically \code{Country}, \code{Year}, \code{Ledger.side}, \code{Product}, etc.)
and columns of SUT matrices, including \code{U} and \code{V}.}

\item{U_colname}{the name of the column of use matrices}

\item{V_colname}{the name of the column of make matrices}

\item{check_colname}{the name of a column of intermediate results
that contains \bold{V}*i (rowsums of \bold{V}) completed against \bold{U}^T on rows.}

\item{problems_data_frame_name}{the name of the data frame created
in the event that any problems are discovered}
}
\value{
Nothing is returned.
This function should be called
for its side-effect of testing whether energy is in balance in \code{.sutdata}.
}
\description{
If an Industry consumes energy (in the \strong{U} matrix)
but does not make energy (in the \strong{V} matrix),
it is most certainly an error.
(In contrast, there can be Industries that make energy but do not consume it,
such as Industries involved in Production.)
This function errors if an Industry is found that consumes energy but does not make energy.
But it will also place an object named \code{problems_data_frame_name}
into the global environment.
\code{problems_data_frame_name} can assist debugging this problem
First, look at the \code{industry_production_OK_colname} column of \code{problems_data_frame_name}.
It shows which rows of \code{.sutdata} had a problem (FALSE in that column).
If you find a row with a problem, look at the \code{check_colname} column.
Any \code{0} entries indicate industries that consume energy but do not make energy.
}
\examples{
verify_SUT_industry_production(SUTMatsWne)
}
