% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_edge_list.R
\name{waste_edges}
\alias{waste_edges}
\title{Create waste energy edges for an edge map}
\usage{
waste_edges(Umat, Vmat, from = "From", to = "To", value = "Value",
  product = "Product", waste = "Waste")
}
\arguments{
\item{Umat}{a use matrix.}

\item{Vmat}{a make matrix.}

\item{from}{the name of the edge list column containing source nodes. (Default is "\code{From}".)}

\item{to}{the name of the edge list column containing destination nodes. (Default is "\code{To}".)}

\item{value}{the name of the edge list column containing magnitudes of the flows. (Default is "\code{Value}".)}

\item{product}{the name of the edge list column containing the product of the edge list flow. (Default is "\code{Product}".)}

\item{waste}{the name of the waste product and the destination node for wastes. (Default is "\code{Waste}".)}
}
\value{
waste energy edges computed from the \code{Umat} and \code{Vmat} matrices
}
\description{
Waste edges are created from the \code{W} matrix.
}
\details{
The \code{waste} argument supplies both the name of the waste flow (default is "\code{Waste}")
and the name of the destination of the waste flows.
}
\examples{
sutmats <- UKEnergy2000mats \%>\% spread(key = matrix.name, value = matrix)
edge_list(sutmats)$`Edge list`[[1]] \%>\% filter(Product == "Waste")
}
