% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{write_ecc_to_excel}
\alias{write_ecc_to_excel}
\title{Write energy conversion chain matrices in an Excel file}
\usage{
write_ecc_to_excel(
  .psut_data = NULL,
  path,
  overwrite_file = FALSE,
  pad = 2,
  include_named_regions = TRUE,
  R = Recca::psut_cols$R,
  U = Recca::psut_cols$U,
  V = Recca::psut_cols$V,
  Y = Recca::psut_cols$Y,
  r_eiou = Recca::psut_cols$r_eiou,
  U_eiou = Recca::psut_cols$U_eiou,
  U_feed = Recca::psut_cols$U_feed,
  S_units = Recca::psut_cols$S_units,
  worksheet_names = NULL,
  .wrote_mats_colname = "Wrote mats",
  UV_bg_color = "#FDF2D0",
  RY_bg_color = "#D3712D",
  calculated_bg_color = "#D9D9D9",
  col_widths = 7
)
}
\arguments{
\item{.psut_data}{A list or data frame of energy conversion chains.
Default is \code{NULL}, in which case
single matrices can be supplied in the
\code{R}, \code{U}, \code{V}, \code{Y}, \code{r_eiou}, \code{U_eiou}, \code{U_feed}, and \code{S_units}
arguments.}

\item{path}{The path of the Excel file to be created.}

\item{overwrite_file}{A boolean that tells whether you want to overwrite
the file at \code{path}, if it already exists.
Default is \code{FALSE}.}

\item{pad}{The number of rows and columns between adjacent matrices in the Excel sheet.
Default is \code{2}.}

\item{include_named_regions}{A boolean that tells whether to name regions of
the Excel tabs according to matrices.
Default is \code{TRUE}.}

\item{R, U, U_feed, U_eiou, r_eiou, V, Y, S_units}{Names of ECC matrices or actual matrices.
See \code{Recca::psut_cols}.}

\item{worksheet_names}{A string or string vector identifying
the names for the worksheets in the workbook.
Alternatively, when \code{.psut_data} is a data frame,
the string name of a column in the data frame
containing the names of the worksheets.
When \code{NULL}, the default, tabs are
numbered sequentially.}

\item{.wrote_mats_colname}{The name of the outgoing column
that tells whether a worksheet was written successfully.
Default is "Wrote mats".}

\item{UV_bg_color}{The color of cells containing U and V matrices.
Default is a creamy yellow.}

\item{RY_bg_color}{The color of cells containing R and Y matrices.
Default is a rust color.}

\item{calculated_bg_color}{The color of cells containing calculated matrices.
Default is gray.}

\item{col_widths}{The widths of columns of matrices.
Default is \code{7} to save space.}
}
\value{
An unmodified version of \code{.psut_data} (if not \code{NULL}) or a list of
the incoming matrices.
}
\description{
It is often helpful to see energy conversion chain (ECC) matrices in Excel format,
arranged spatially.
This function takes ECC matrices and writes them to an Excel file.
}
\details{
If \code{.psut_data} is a PSUT data frame,
each row is written to a different tab in the output file at \code{path}.

When \code{include_named_regions} is \code{TRUE} (the default),
named regions for matrices are added to Excel sheets.
The format for the names is \verb{<<matrix>>_<<sheet name>>}.
For example, "R_4" for the \strong{R} matrix on the sheet named "4".
The names help to identify matrices in high-level views of the Excel file.
The region names apply to the numbers only.
Row names are one column left of the named region.
Column names are one row above the named region.

When \code{worksheet_names} is not \code{NULL} (the default),
be sure that worksheet names are unique.
Also, be aware that worksheet names
must have 31 characters or fewer.
Furthermore, the worksheet names
may not contain any of the following characters:
\\  /  ?  *  [  ]
}
\examples{
\dontrun{
ecc <- UKEnergy2000mats \%>\%
  tidyr::pivot_wider(names_from = "matrix.name",
                     values_from = "matrix") |>
dplyr::mutate(
  tab_names = paste0(EnergyType, "-", LastStage)
)
ecc_temp_path <- tempfile(pattern = "write_excel_ecc_test_file", fileext = ".xlsx")
write_ecc_to_excel(ecc,
                   path = ecc_temp_path,
                   tab_names = "tab_names",
                   overwrite = TRUE)
}
}
