% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{write_ecc_to_excel}
\alias{write_ecc_to_excel}
\title{Write energy conversion chain matrices in an Excel format}
\usage{
write_ecc_to_excel(
  .psut_data = NULL,
  path,
  overwrite_file = FALSE,
  pad = 2,
  R = Recca::psut_cols$R,
  U = Recca::psut_cols$U,
  V = Recca::psut_cols$V,
  Y = Recca::psut_cols$Y,
  r_eiou = Recca::psut_cols$r_eiou,
  U_eiou = Recca::psut_cols$U_eiou,
  U_feed = Recca::psut_cols$U_feed,
  S_units = Recca::psut_cols$S_units,
  .wrote_mats_colname = "Wrote mats",
  matrix_bg_color = "#FDF2D0"
)
}
\arguments{
\item{.psut_data}{A list or data frame of energy conversion chains.}

\item{path}{The path of the Excel file to be created.}

\item{overwrite_file}{A boolean that tells whether you want to overwrite
the file at \code{path}, if it already exists.}

\item{pad}{The number of rows and columns between adjacent matrices in the Excel sheet.
Default is \code{2}.}

\item{R, U, V, Y, r_eiou, U_eiou, U_feed}{Names of ECC matrices or actual matrices.
See \code{Recca::psut_cols}.}

\item{.wrote_mats_colname}{The name of the outgoing column
that tells whether a worksheet was written successfully.
Default is "Wrote mats".}

\item{matrix_bg_color}{The color of cells containing matrix numbers.
Default is a creamy yellow.}
}
\value{
An unmodified version of \code{.psut_data} (if not \code{NULL}) or a list of
the incoming matrices.
}
\description{
It is often helpful to see energy conversion chain (ECC) matrices in Excel format,
arranged spatially.
This function takes ECC matrices and writes them to an Excel file.
}
\details{
If \code{.psut_data} is a PSUT data frame,
each row is written to a different tab in the output file at \code{path}.
}
\examples{
\dontrun{
ecc <- UKEnergy2000mats \%>\%
  tidyr::spread(key = "matrix.name", value = "matrix")
ecc_temp_path <- tempfile(pattern = "write_excel_ecc_test_file", fileext = ".xlsx")
write_ecc_to_excel(ecc, path = ecc_temp_path, overwrite = TRUE)
}
}
